% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anovatable.R
\name{anovatable}
\alias{anovatable}
\title{Analysis of Variance}
\usage{
anovatable(X, y)
}
\arguments{
\item{X}{\code{n} by \code{k} numeric matrix.
The data matrix \eqn{\mathbf{X}}
(also known as design matrix, model matrix or regressor matrix)
is an \eqn{n \times k} matrix of \eqn{n} observations of \eqn{k} regressors,
which includes a regressor whose value is 1 for each observation on the first column.}

\item{y}{Numeric vector of length \code{n} or \code{n} by \code{1} matrix.
The vector \eqn{\mathbf{y}} is an \eqn{n \times 1} vector of observations
on the regressand variable.}
}
\value{
Returns the ANOVA table.
}
\description{
Analysis of Variance
}
\examples{
# Simple regression------------------------------------------------
X <- jeksterslabRdatarepo::wages.matrix[["X"]]
X <- X[, c(1, ncol(X))]
y <- jeksterslabRdatarepo::wages.matrix[["y"]]
anovatable(X = X, y = y)

# Multiple regression----------------------------------------------
X <- jeksterslabRdatarepo::wages.matrix[["X"]]
# age is removed
X <- X[, -ncol(X)]
anovatable(X = X, y = y)
}
\seealso{
Other hypothesis testing functions: 
\code{\link{.anovatable}()},
\code{\link{ci}()},
\code{\link{nhst}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{hypothesis testing functions}
\keyword{inference}
