% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_hat.R
\name{beta_hat_inv}
\alias{beta_hat_inv}
\title{Beta-hat - Inverse of X}
\usage{
beta_hat_inv(X, y)
}
\arguments{
\item{X}{Matrix.
The data matrix
\eqn{\mathbf{X}}
is an \eqn{n \times k} matrix
of \eqn{n} observations
of \eqn{k} regressors,
which includes a regressor
whose value is 1 for each observation.}

\item{y}{Vector or \code{n} by \code{1} matrix.
The vector
\eqn{\mathbf{y}}
is an \eqn{n \times 1} vector
of observations on the regressand variable.}
}
\value{
Returns \eqn{\boldsymbol{\hat{\beta}}}
that is,
a \eqn{k \times 1} vector of estimates
of \eqn{k} unknown regression coefficients
estimated using ordinary least squares.
}
\description{
Estimates coefficients of a linear regression model
using
\deqn{
    \boldsymbol{\hat{\beta}}
    =
    \left(
      \mathbf{X}^{\prime}
      \mathbf{X}
    \right)^{-1}
    \left(
      \mathbf{X}^{\prime}
      \mathbf{y}
    \right).
  }
}
\references{
\href{https://en.wikipedia.org/wiki/Linear_regression}{Wikipedia: Linear Regression}

\href{https://en.wikipedia.org/wiki/Ordinary_least_squares}{Wikipedia: Ordinary Least Squares}

\href{https://en.wikipedia.org/wiki/Numerical_methods_for_linear_least_squares#Inverting_the_matrix_of_the_normal_equations}{Wikipedia: Inverting the matrix of the normal equations}
}
\seealso{
Other beta-hat functions: 
\code{\link{beta_hat_qr}()},
\code{\link{beta_hat_svd}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{beta-hat functions}
