% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betahat.R
\name{betahat}
\alias{betahat}
\title{Beta-hat (\eqn{\boldsymbol{\hat{\beta}}})}
\usage{
betahat(X, y, FUN = betahat_inv)
}
\arguments{
\item{X}{Matrix.
The data matrix
\eqn{\mathbf{X}}
is an \eqn{n \times k} matrix
of \eqn{n} observations
of \eqn{k} regressors,
which includes a regressor
whose value is 1 for each observation.}

\item{y}{Vector or \code{n} by \code{1} matrix.
The vector
\eqn{\mathbf{y}}
is an \eqn{n \times 1} vector
of observations on the regressand variable.}

\item{FUN}{Function.
Beta-hat function to use.
\code{\link[=betahat_inv]{betahat_inv()}} is used by default.}
}
\description{
See \code{\link[=betahat_inv]{betahat_inv()}} for inverse of X,
\code{\link[=betahat_qr]{betahat_qr()}} for QR decomposition, and
\code{\link[=betahat_svd]{betahat_svd()}} for singular value decomposition.
}
\references{
\href{https://en.wikipedia.org/wiki/Linear_regression}{Wikipedia: Linear Regression}

\href{https://en.wikipedia.org/wiki/Ordinary_least_squares}{Wikipedia: Ordinary Least Squares}

\href{https://en.wikipedia.org/wiki/Numerical_methods_for_linear_least_squares#Inverting_the_matrix_of_the_normal_equations}{Wikipedia: Inverting the matrix of the normal equations}

\href{https://en.wikipedia.org/wiki/QR_decomposition}{Wikipedia: QR Decomposition}

\href{https://en.wikipedia.org/wiki/Singular_value_decomposition}{Wikipedia: Singular Value Decomposition}

\href{https://en.wikipedia.org/wiki/Numerical_methods_for_linear_least_squares#Orthogonal_decomposition_methods}{Wikipedia: Orthogonal decomposition methods}
}
\seealso{
Other beta-hat functions: 
\code{\link{betahat_inv}()},
\code{\link{betahat_qr}()},
\code{\link{betahat_svd}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{beta-hat functions}
