% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betahat.R
\name{betahat_cor}
\alias{betahat_cor}
\title{Beta-hat (\eqn{\boldsymbol{\hat{\beta}}}) - Correlations}
\usage{
betahat_cor(R_X, r_yX)
}
\arguments{
\item{R_X}{Matrix.
Correlations between the regressors.}

\item{r_yX}{Vector
or
\eqn{k - 1 \times 1}
matrix
of correlations between the regressand
and the regressors.}
}
\value{
Returns the slopes of a linear regression model
derived from correlations.
}
\description{
Estimates slopes of a linear regression model
(\eqn{\boldsymbol{\hat{\beta}}} minus the intercept)
as a function of correlations
given by
\deqn{
    \boldsymbol{\hat{\beta}}
    =
    \mathrm{R}_{\mathbf{X}}
    \mathrm{r}_{\mathbf{yX}}
  }
where
\eqn{\mathbf{R}_{\mathbf{X}}}
is the correlation matrix
of the regressor variables
and
\eqn{ \mathbf{r}_{\mathbf{yX}}}
is the column vector
of the correlations between
the regressand
and the regressors.
}
\author{
Ivan Jacob Agaloos Pesigan
}
