% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betahat.R
\name{betahat_cov}
\alias{betahat_cov}
\title{Beta-hat (\eqn{\boldsymbol{\hat{\beta}}}) - Covariances}
\usage{
betahat_cov(V_X, v_yX)
}
\arguments{
\item{V_X}{Matrix.
Covariances between the regressors.}

\item{v_yX}{Vector
or
\eqn{k - 1 \times 1}
matrix
of covariances between the regressand
and the regressors.}
}
\value{
Returns the slopes of a linear regression model
derived from covariances.
}
\description{
Estimates slopes of a linear regression model
(\eqn{\boldsymbol{\hat{\beta}}} minus the intercept)
as a function of covariances
given by
\deqn{
    \boldsymbol{\hat{\beta}}
    =
    \mathrm{V}_{\mathbf{X}}
    \mathrm{v}_{\mathbf{yX}}
  }
where
\eqn{\mathbf{V}_{\mathbf{X}}}
is the covariance matrix
of the regressor variables
and
\eqn{ \mathbf{v}_{\mathbf{yX}}}
is the column vector
of the covariances between
the regressand
and the regressors.
}
\author{
Ivan Jacob Agaloos Pesigan
}
