% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betahat_mat.R
\name{betahat_intercept}
\alias{betahat_intercept}
\title{Beta-hat Intercept
\eqn{
  \left(
    \boldsymbol{
      \hat{
        \beta
      }
    }_{
      \mathrm{
        intercept
      }
    }
  \right)
}}
\usage{
betahat_intercept(betahat_slopes, muhatX, muhaty)
}
\arguments{
\item{betahat_slopes}{Numeric vector.
Estimates of unknown regression slopes.}

\item{muhatX}{Numeric vector.
Sample means of regressors.}

\item{muhaty}{Numeric.
Sample mean of regressand.}
}
\description{
Calculates the intercept of a linear regression model
(
\eqn{
    \boldsymbol{
      \hat{
        \beta
      }
    }
  }
minus the slopes
)
as a function of sample means and
regression slopes
given by
\deqn{
  \boldsymbol{
    \hat{
      \beta
    }
  }_{
    \mathrm{
      intercept
    }
  }
  =
  \boldsymbol{
    \hat{
      \beta
    }
  }_{
    \mathrm{
      slopes
    }
  }^{
    \prime
  }
  \boldsymbol{
    \hat{
      \mu
    }
  }
}
where
\eqn{
  \boldsymbol{
    \hat{
      \beta
    }
  }_{
    \mathrm{
      slopes
    }
  }
}
is a vector of estimated regression slopes
with the last element as 0,
and
\eqn{
  \boldsymbol{
    \hat{\mu}
  }
}
is a vector of sample means
of the regressor variables
and the regressand variable.
}
\seealso{
Other beta-hat from summary matrices functions: 
\code{\link{.betahat_mat}()},
\code{\link{betahat_mat}()},
\code{\link{betahat_slopes_cov}()},
\code{\link{betahat_std_slopes_cor}()}
}
\concept{beta-hat from summary matrices functions}
