% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betahat_mat.R
\name{betahat_slopes_cov}
\alias{betahat_slopes_cov}
\title{Beta-hat Slopes
\eqn{
  \left(
    \boldsymbol{
      \hat{
        \beta
      }
    }_{
      \mathrm{
        slopes
      }
    }
  \right)
} - Covariances}
\usage{
betahat_slopes_cov(VX, vyX)
}
\arguments{
\item{VX}{Matrix.
Covariances between the regressors.}

\item{vyX}{Vector
or
\eqn{
  \left(
    k
    -
    1
  \right)
  \times
  1
}
matrix
of covariances between the regressand
and the regressors.}
}
\value{
Returns the slopes
\eqn{
  \left(
    \boldsymbol{
      \hat{
        \beta
      }
    }_{
      \mathrm{
        slopes
      }
    }
  \right)
}
of a linear regression model
derived from the variance-covariance matrix.
}
\description{
Calculates the slopes of a linear regression model
(
\eqn{
    \boldsymbol{
      \hat{
        \beta
      }
    }
  }
minus the intercept
)
as a function of covariances
given by
\deqn{
  \boldsymbol{
    \hat{
      \beta
    }
  }_{
    \mathrm{
      slopes
    }
  }
  =
  \mathrm{
    V
  }_{
    \mathbf{
      X
    }
  }^{
    \prime
  }
  \mathrm{
    v
  }_{
    \mathbf{
      yX
    }
  }
}
where
\eqn{
  \mathbf{
    V
  }_{
    \mathbf{
      X
    }
  }
}
is the covariance matrix
of the regressor variables
and
\eqn{
  \mathbf{
    v
  }_{
    \mathbf{
      yX
    }
  }
}
is the column vector
of the covariances between
the regressand
and the regressors.
}
\seealso{
Other beta-hat from summary matrices functions: 
\code{\link{.betahat_mat}()},
\code{\link{betahat_intercept}()},
\code{\link{betahat_mat}()},
\code{\link{betahat_std_slopes_cor}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{beta-hat from summary matrices functions}
