% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betahat_mat.R
\name{betahat_std_mat}
\alias{betahat_std_mat}
\title{Standardized Beta-hat Slopes
\eqn{
  \left(
    \boldsymbol{
      \hat{
        \beta
      }
    }_{
      \mathrm{
        std
      }
    }
  \right)
} - Correlations}
\usage{
betahat_std_mat(X, y)
}
\arguments{
\item{X}{Matrix.
The data matrix
\eqn{\mathbf{X}}
(also known as
design matrix,
model matrix or
regressor matrix)
is an \eqn{n \times k} matrix
of \eqn{n} observations
of \eqn{k} regressors,
which includes a regressor
whose value is 1 for each observation.}

\item{y}{Vector or \code{n} by \code{1} matrix.
The vector
\eqn{\mathbf{y}}
is an \eqn{n \times 1} vector
of observations on the regressand variable.}
}
\value{
Returns the standardized slopes
\eqn{
  \left(
    \boldsymbol{
      \hat{
        \beta
      }
    }_{
      \mathrm{
        std
      }
    }
  \right)
}
of a linear regression model
derived from the correlation matrix.
}
\description{
Calculates
\eqn{
  \boldsymbol{
    \hat{
      \beta
    }
  }_{
    \mathrm{
      std
    }
  }
}
from the correlation matrix.
See \code{\link[=betahat_std_slopes_cor]{betahat_std_slopes_cor()}}.
}
\author{
Ivan Jacob Agaloos Pesigan
}
