% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betahat_mat.R
\name{betahat_std_slopes_cor}
\alias{betahat_std_slopes_cor}
\title{Standardized Beta-hat Slopes (\eqn{\boldsymbol{\hat{\beta}}_{\mathrm{std}}}) - Correlations}
\usage{
betahat_std_slopes_cor(RX, ryX)
}
\arguments{
\item{RX}{Matrix.
Correlations between the regressors.}

\item{ryX}{Vector
or
\eqn{\left( k - 1 \right) \times 1}
matrix
of correlations between the regressand
and the regressors.}
}
\value{
Returns the standardized slopes
(\eqn{\boldsymbol{\hat{\beta}}_\mathrm{std}})
of a linear regression model
derived from the correlation matrix.
}
\description{
Calculates the standardized slopes of a linear regression model
(\eqn{\boldsymbol{\hat{\beta}_{\mathrn{std}}}})
as a function of correlations
given by
\deqn{
    \boldsymbol{\hat{\beta}}_{\mathrm{std}}
    =
    \mathrm{R}_{\mathbf{X}}^{\prime}
    \mathrm{r}_{\mathbf{yX}}
  }
where
\eqn{\mathbf{R}_{\mathbf{X}}}
is the correlation matrix
of the regressor variables
and
\eqn{ \mathbf{r}_{\mathbf{yX}}}
is the column vector
of the correlations between
the regressand
and the regressors.
Note that the \eqn{y}-intercept
is zero when the variables are scaled.
}
\seealso{
Other beta-hat from summary matrices functions: 
\code{\link{.betahat_mat}()},
\code{\link{betahat_intercept}()},
\code{\link{betahat_mat}()},
\code{\link{betahat_slopes_cov}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{beta-hat from summary matrices functions}
