% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcov.R
\name{cov_beta_hat}
\alias{cov_beta_hat}
\title{Variance-Covariance Matrix of Estimates Regression Coefficients}
\usage{
cov_beta_hat(beta_hat = NULL, X, y, s_sqr_est = "ols")
}
\arguments{
\item{beta_hat}{Vector or \code{k} by \code{1} matrix.
The vector
\eqn{\boldsymbol{\hat{\beta}}}
is a \eqn{k \times 1} vector of estimates
of \eqn{k} unknown regression coefficients.}

\item{X}{Matrix.
The data matrix
\eqn{\mathbf{X}}
is an \eqn{n \times k} matrix
of \eqn{n} observations
of \eqn{k} regressors,
which includes a regressor
whose value is 1 for each observation.}

\item{y}{Vector or \code{n} by \code{1} matrix.
The vector
\eqn{\mathbf{y}}
is an \eqn{n \times 1} vector
of observations on the regressand variable.}

\item{s_sqr_est}{String.
Residual variance estimator.
If \code{"both"},
returns both OLS and ML estimates as a vector.
If \code{"ols"},
returns OLS estimate.
If \code{"ml"},
returns ML estimate.}
}
\value{
Returns variance-covariance matrix of estimates of regression coefficients.
}
\description{
Calculates the variance-covariance matrix of estimates of regression coefficients
\deqn{
    \sigma^2 \left(\mathbf{X}^{\prime} \mathbf{X} \right)^{-1}.
  }
}
\author{
Ivan Jacob Agaloos Pesigan
}
