% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptives.R
\name{descriptives}
\alias{descriptives}
\title{Descriptive Statistics}
\usage{
descriptives(X, y, muhat = TRUE)
}
\arguments{
\item{X}{Matrix.
The data matrix
\eqn{\mathbf{X}}
(also known as
design matrix,
model matrix or
regressor matrix)
is an \eqn{n \times k} matrix
of \eqn{n} observations
of \eqn{k} regressors,
which includes a regressor
whose value is 1 for each observation.}

\item{y}{Vector or \code{n} by \code{1} matrix.
The vector
\eqn{\mathbf{y}}
is an \eqn{n \times 1} vector
of observations on the regressand variable.}

\item{muhat}{Logical.
If \code{TRUE}, compute means.}
}
\value{
Returns a list with the following items
\describe{
\item{Sigmahat}{Variance-covariance matrix (\eqn{\boldsymbol{\hat{\Sigma}}}).}
\item{R}{Correlation matrix (\eqn{\mathrm{R}}).}
\item{VX}{Variance-covariance matrix (\eqn{\mathbf{V}_{\mathbf{X}}}).}
\item{vyX}{Vector of covariances between \eqn{\mathbf{y}} and \eqn{\mathbf{X}} (\eqn{\mathbf{v}_{\mathbf{yX}}}).}
\item{RX}{Correlation matrix (\eqn{\mathbf{R}_{\mathbf{X}}}).}
\item{ryX}{Vector of correlations between \eqn{\mathbf{y}} and \eqn{\mathbf{X}} (\eqn{\mathbf{r}_{\mathbf{yX}}}).}
\item{muhatX}{Means of regressor variables \eqn{\boldsymbol{\hat{\mu}}_\mathbf{X}}.}
\item{muhaty}{Mean of regressand variable \eqn{\hat{\mu}_\mathbf{y}}.}
}
Note that \code{muhatX} and \code{muhaty} are \code{NA}
when \code{muhat = FALSE}.
}
\description{
Calculate means, covariances and correlations.
}
\author{
Ivan Jacob Agaloos Pesigan
}
