% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anovatable.R
\name{.anovatable}
\alias{.anovatable}
\title{Analysis of Variance (from \eqn{RSS} and \eqn{ESS})}
\usage{
.anovatable(RSS = NULL, ESS = NULL, n, k, X, y)
}
\arguments{
\item{RSS}{Numeric.
Residual sum of squares.}

\item{ESS}{Numeric.
Explained sum of squares.}

\item{n}{Integer.
Sample size.}

\item{k}{Integer.
Number of regressors including a regressor whose value is 1 for each observation.}

\item{X}{\code{n} by \code{k} numeric matrix.
The data matrix \eqn{\mathbf{X}}
(also known as design matrix, model matrix or regressor matrix)
is an \eqn{n \times k} matrix of \eqn{n} observations of \eqn{k} regressors,
which includes a regressor whose value is 1 for each observation on the first column.}

\item{y}{Numeric vector of length \code{n} or \code{n} by \code{1} matrix.
The vector \eqn{\mathbf{y}} is an \eqn{n \times 1} vector of observations
on the regressand variable.}
}
\description{
Analysis of Variance (from \eqn{RSS} and \eqn{ESS})
}
\seealso{
Other hypothesis testing functions: 
\code{\link{anovatable}()},
\code{\link{ci}()},
\code{\link{nhst}()}
}
\author{
Ivan Jacob Agaloos Pesigan

Ivan Jacob Agaloos Pesigan
}
\concept{hypothesis testing functions}
\keyword{inference}
