% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aov.R
\name{.aov}
\alias{.aov}
\title{Analysis of Variance}
\usage{
.aov(RSS = NULL, ESS = NULL, n, k, X = X, y = y, betahat = NULL)
}
\arguments{
\item{RSS}{Numeric.
Residual sum of squares.}

\item{ESS}{Numeric.
Explained sum of squares.}

\item{n}{Integer.
Sample size.}

\item{k}{Integer.
Number of regressors
including a regressor
whose value is 1 for each observation.}

\item{X}{Matrix.
The data matrix
\eqn{\mathbf{X}}
(also known as
design matrix,
model matrix or
regressor matrix)
is an \eqn{n \times k} matrix
of \eqn{n} observations
of \eqn{k} regressors,
which includes a regressor
whose value is 1 for each observation.}

\item{y}{Vector or \code{n} by \code{1} matrix.
The vector
\eqn{\mathbf{y}}
is an \eqn{n \times 1} vector
of observations on the regressand variable.}

\item{betahat}{Vector or \code{k} by \code{1} matrix.
The vector
\eqn{\boldsymbol{\hat{\beta}}}
is a \eqn{k \times 1} vector of estimates
of \eqn{k} unknown regression coefficients.}
}
\description{
Analysis of Variance
}
\author{
Ivan Jacob Agaloos Pesigan
}
