% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betahat_inf.R
\name{.betahat_inf}
\alias{.betahat_inf}
\title{Beta-hat Inference}
\usage{
.betahat_inf(
  betahat = NULL,
  vcov = NULL,
  n,
  X,
  y,
  FUN = betahat_inv,
  type = "unbiased"
)
}
\arguments{
\item{betahat}{Vector or \code{k} by \code{1} matrix.
The vector
\eqn{\boldsymbol{\hat{\beta}}}
is a \eqn{k \times 1} vector of estimates
of \eqn{k} unknown regression coefficients.}

\item{vcov}{Matrix.
Variance-covariance matrix
of estimates
of regression coefficients.}

\item{n}{Integer.
Sample size.}

\item{X}{Matrix.
The data matrix
\eqn{\mathbf{X}}
(also known as
design matrix,
model matrix or
regressor matrix)
is an \eqn{n \times k} matrix
of \eqn{n} observations
of \eqn{k} regressors,
which includes a regressor
whose value is 1 for each observation.}

\item{y}{Vector or \code{n} by \code{1} matrix.
The vector
\eqn{\mathbf{y}}
is an \eqn{n \times 1} vector
of observations on the regressand variable.}

\item{FUN}{Function.
Beta-hat function to use.
\code{\link[=betahat_inv]{betahat_inv()}} is used by default.}

\item{type}{String.
Residual variance estimator.
If \code{type = "unbiased"},
returns unbiased estimate.
If \code{type = "biased"},
returns biased estimate.
If \code{type = "both"},
returns a vector of
unbiased and biased estimates.}
}
\description{
Beta-hat Inference
}
\author{
Ivan Jacob Agaloos Pesigan
}
