% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betahat_mat.R
\name{.betahat_mat}
\alias{.betahat_mat}
\title{Beta-hat (\eqn{\boldsymbol{\hat{\beta}}}) - From Means and Covariance Matrix}
\usage{
.betahat_mat(
  VX = NULL,
  vyX = NULL,
  muhatX = NULL,
  muhaty = NULL,
  X = NULL,
  y = NULL
)
}
\arguments{
\item{VX}{Matrix.
Covariances between the regressors.}

\item{vyX}{Vector
or
\eqn{\left( k - 1 \right) \times 1}
matrix
of covariances between the regressand
and the regressors.}

\item{muhatX}{Vector.
Sample means of regressors.}

\item{muhaty}{Numeric.
Sample mean of regressand.}

\item{X}{Matrix.
The data matrix
\eqn{\mathbf{X}}
(also known as
design matrix,
model matrix or
regressor matrix)
is an \eqn{n \times k} matrix
of \eqn{n} observations
of \eqn{k} regressors,
which includes a regressor
whose value is 1 for each observation.}

\item{y}{Vector or \code{n} by \code{1} matrix.
The vector
\eqn{\mathbf{y}}
is an \eqn{n \times 1} vector
of observations on the regressand variable.}
}
\description{
Calculates \eqn{\boldsymbol{\hat{\beta}}}
from means and covariance matrix.
See
\code{\link[=.betahat_mat]{.betahat_mat()}},
\code{\link[=betahat_intercept]{betahat_intercept()}},
\code{\link[=betahat_slopes_cov]{betahat_slopes_cov()}}.
}
\details{
If
\code{VX},
\code{vyX},
\code{muhatX},
or
\code{muhaty}
is \code{NULL}
these values are calculated
from
\code{X}
and
\code{y}.
\code{X} and \code{y} are ignored when
\code{VX},
\code{vyX},
\code{muhatX},
and
\code{muhaty}
are provided.
}
\seealso{
Other beta-hat from summary matrices functions: 
\code{\link{betahat_intercept}()},
\code{\link{betahat_mat}()},
\code{\link{betahat_slopes_cov}()},
\code{\link{betahat_std_slopes_cor}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{beta-hat from summary matrices functions}
