% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SS.R
\name{.ESS}
\alias{.ESS}
\title{Explained Sum of Squares (from \eqn{\mathbf{\hat{y}}} and \eqn{\mathbf{\bar{Y}}})}
\usage{
.ESS(yhat = NULL, ybar = NULL, X, y, betahat = NULL)
}
\arguments{
\item{yhat}{Vector or \code{n} by \code{1} matrix.
\eqn{n \times 1} vector
of predicted values of \eqn{\mathbf{y}}
(\eqn{\mathbf{\hat{y}}}).}

\item{ybar}{Numeric.
Mean of y.}

\item{X}{Matrix.
The data matrix
\eqn{\mathbf{X}}
(also known as
design matrix,
model matrix or
regressor matrix)
is an \eqn{n \times k} matrix
of \eqn{n} observations
of \eqn{k} regressors,
which includes a regressor
whose value is 1 for each observation.}

\item{y}{Vector or \code{n} by \code{1} matrix.
The vector
\eqn{\mathbf{y}}
is an \eqn{n \times 1} vector
of observations on the regressand variable.}

\item{betahat}{Vector or \code{k} by \code{1} matrix.
The vector
\eqn{\boldsymbol{\hat{\beta}}}
is a \eqn{k \times 1} vector of estimates
of \eqn{k} unknown regression coefficients.}
}
\value{
Returns the explained sum of squares (\eqn{ESS}).
}
\description{
Calculates the explained sum of squares (ESS)
using
\deqn{
    ESS
    =
    \sum_{i = 1}^{n}
    \left(
      \hat{Y}_{i}
      -
      \bar{Y}
    \right)^2 \\
    =
    \sum_{i = 1}^{n}
    \left(
      \hat{\beta}_{1}
      +
      \hat{\beta}_{2}
      X_{2i}
      +
      \hat{\beta}_{3}
      X_{3i}
      +
      \dots
      +
      \hat{\beta}_{k}
      X_{ki}
      -
      \bar{Y}
    \right)^2 \\
  }
In matrix form
\deqn{
    ESS
    =
    \sum_{i = 1}^{n}
    \left(
      \mathbf{\hat{y}}
      -
      \mathbf{\bar{Y}}
    \right)^2 \\
    =
    \sum_{i = 1}^{n}
    \left(
      \mathbf{X}
      \boldsymbol{\hat{\beta}}
      -
      \mathbf{\bar{Y}}
    \right)^2
    }
where
\eqn{\mathbf{\hat{y}}}
(\eqn{\mathbf{X} \boldsymbol{\hat{\beta}}})
is an \eqn{n \times 1} matrix
of predicted values of
\eqn{\mathbf{y}},
and
\eqn{\mathbf{\bar{Y}}}
is the mean of
\eqn{\mathbf{y}}.
Equivalent computational matrix formula
\deqn{
    ESS
    =
    \boldsymbol{\hat{\beta}}^{\prime}
    \mathbf{X}^{\prime}
    \mathbf{X}
    \boldsymbol{\hat{\beta}}
    -
    n
    \mathbf{\bar{Y}}^{2}.
  }
Note that
\deqn{
    TSS
    =
    ESS
    +
    RSS.
  }
}
\details{
If \code{yhat = NULL},
it is computed using \code{\link[=yhat]{yhat()}}
with \code{X} and \code{y} as required arguments
and \code{betahat} as an optional argument.
}
\references{
\href{https://en.wikipedia.org/wiki/Residual_sum_of_squares}{Wikipedia: Residual Sum of Squares}

\href{https://en.wikipedia.org/wiki/Explained_sum_of_squares}{Wikipedia: Explained Sum of Squares}

\href{https://en.wikipedia.org/wiki/Total_sum_of_squares}{Wikipedia: Total Sum of Squares}

\href{https://en.wikipedia.org/wiki/Coefficient_of_determination}{Wikipedia: Coefficient of Determination}
}
\seealso{
Other sum of squares functions: 
\code{\link{.ESS_raw}()},
\code{\link{.RSS_raw}()},
\code{\link{.RSS}()},
\code{\link{ESS}()},
\code{\link{RSS}()},
\code{\link{TSS}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{sum of squares functions}
