% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aov.R
\name{.F_R2}
\alias{.F_R2}
\title{F-Test (From \eqn{R^2})}
\usage{
.F_R2(R2 = NULL, n, k, betahat = NULL, X = X, y = y)
}
\arguments{
\item{R2}{Numeric.
Coefficient of determinims \eqn{R^2}.}

\item{n}{Integer.
Sample size.}

\item{k}{Integer.
Number of regressors
including a regressor
whose value is 1 for each observation.}

\item{betahat}{Vector or \code{k} by \code{1} matrix.
The vector
\eqn{\boldsymbol{\hat{\beta}}}
is a \eqn{k \times 1} vector of estimates
of \eqn{k} unknown regression coefficients.}

\item{X}{Matrix.
The data matrix
\eqn{\mathbf{X}}
(also known as
design matrix,
model matrix or
regressor matrix)
is an \eqn{n \times k} matrix
of \eqn{n} observations
of \eqn{k} regressors,
which includes a regressor
whose value is 1 for each observation.}

\item{y}{Vector or \code{n} by \code{1} matrix.
The vector
\eqn{\mathbf{y}}
is an \eqn{n \times 1} vector
of observations on the regressand variable.}
}
\value{
Returns a vector with the following elements
\describe{
\item{R2}{Coefficient of determination (\eqn{R^2}).}
\item{F}{F Statistic (\eqn{F}).}
\item{p}{p value (\eqn{p}).}
\item{df1}{Degrees of Freedom - Model (\eqn{df_1 = k - 1}).}
\item{df2}{Degrees of Freedom - Error (\eqn{df_2 = n - k}).}
}
}
\description{
Calculates the \eqn{F} statistic from \eqn{R^2}
using
\deqn{
    F
    =
    \frac{
      \frac{
        R^2
      }
      {
        k - 1
      }
    }
    {
      \frac{
        1
        -
        R^2
      }
      {
        n - k
      }
    }
  }
}
\author{
Ivan Jacob Agaloos Pesigan
}
