% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leverage.R
\name{.h}
\alias{.h}
\title{Leverage}
\usage{
.h(P = NULL, X = NULL)
}
\arguments{
\item{P}{Numeric matrix
The projection matrix.}

\item{X}{\code{n} by \code{k} numeric matrix.
The data matrix \eqn{\mathbf{X}}
(also known as design matrix, model matrix or regressor matrix)
is an \eqn{n \times k} matrix of \eqn{n} observations of \eqn{k} regressors,
which includes a regressor whose value is 1 for each observation on the first column.}
}
\value{
Returns leverage.
}
\description{
Calculates leverage, that is, how far away the regressor values
of an observation are from those of the other observations using
\deqn{
    h_{ii} = x_{i}^{T} \left( \mathbf{X}^{T} \mathbf{X} \right)^{-1} x_{i}
  }
where \eqn{x_{i}^{T}}
is the \eqn{i}th row of the \eqn{\mathbf{X}} matrix.
Note that
\eqn{
    \mathbf{X} \left( \mathbf{X}^{T} \mathbf{X} \right)^{-1} \mathbf{X}^{T}
  }
is the projection matrix (or hat matrix) \eqn{\mathbf{P}}
and \eqn{h_{ii}} is the diagonal of \eqn{\mathbf{P}}.
}
\details{
If \code{P = NULL}, \code{P} is computed with \code{X} as a required argument.
\code{X} is ignored if \code{P} is provided.
}
\references{
\href{https://en.wikipedia.org/wiki/Leverage_(statistics)}{Wikipedia: Leverage}
}
\seealso{
Other projection matrix functions: 
\code{\link{.M}()},
\code{\link{M}()},
\code{\link{P}()},
\code{\link{h}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{projection matrix functions}
\keyword{projection}
