% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proj.R
\name{.M}
\alias{.M}
\title{M Matrix}
\usage{
.M(X, P = NULL)
}
\arguments{
\item{X}{\code{n} by \code{k} numeric matrix.
The data matrix \eqn{\mathbf{X}}
(also known as design matrix, model matrix or regressor matrix)
is an \eqn{n \times k} matrix of \eqn{n} observations of \eqn{k} regressors,
which includes a regressor whose value is 1 for each observation on the first column.}

\item{P}{\code{n} by \code{n} numeric matrix.
The \eqn{n \times n} projection matrix \eqn{\left( \mathbf{P} \right)}.}
}
\value{
Returns the residual maker matrix \eqn{\left( \mathbf{M} \right)} .
}
\description{
Calculates the residual maker matrix
\eqn{\left( \mathbf{M} \right)} using
\deqn{
    \mathbf{M} = \mathbf{I} - \mathbf{P}
  }
where
\deqn{
    \mathbf{P} = \mathbf{X} \left( \mathbf{X}^{T} \mathbf{X} \right)^{-1} \mathbf{X}^{T} .
  }
}
\details{
If \code{P = NULL}, the \code{P} matrix is computed using \code{\link[=P]{P()}}.
The residual maker matrix \eqn{\left( \mathbf{M} \right)}
transforms the \eqn{\mathbf{y}} vector
to the vector of residuals \eqn{\left( \mathbf{e} = \mathbf{My} \right)} .
}
\references{
\href{https://en.wikipedia.org/wiki/Projection_matrix}{Wikipedia: Projection Matrix}
}
\seealso{
Other projection matrix functions: 
\code{\link{M}()},
\code{\link{P}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{projection matrix functions}
\keyword{projection}
