% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{.model}
\alias{.model}
\title{Model Assessment}
\usage{
.model(RSS = NULL, TSS = NULL, n, k, X, y)
}
\arguments{
\item{RSS}{Numeric.
Residual sum of squares.}

\item{TSS}{Numeric.
Total sum of squares.}

\item{n}{Integer.
Sample size.}

\item{k}{Integer.
Number of regressors including a regressor whose value is 1 for each observation.}

\item{X}{\code{n} by \code{k} numeric matrix.
The data matrix \eqn{\mathbf{X}}
(also known as design matrix, model matrix or regressor matrix)
is an \eqn{n \times k} matrix of \eqn{n} observations of \eqn{k} regressors,
which includes a regressor whose value is 1 for each observation on the first column.}

\item{y}{Numeric vector of length \code{n} or \code{n} by \code{1} matrix.
The vector \eqn{\mathbf{y}} is an \eqn{n \times 1} vector of observations
on the regressand variable.}
}
\value{
Returns a vector with the following elements
\describe{
\item{RSS}{Residual sum of squares.}
\item{MSE}{Mean squared error.}
\item{RMSE}{Root mean squared error.}
\item{R2}{R-squared \eqn{\left( R^2 \right)}.}
\item{Rbar2}{Adjusted R-squared \eqn{\left( \bar{R}^2 \right)} .}
}
}
\description{
Model Assessment
}
\details{
If \code{RSS = NULL}, \code{RSS} is computed using \code{\link[=RSS]{RSS()}}
with \code{X} and \code{y} as required arguments.
If \code{RSS} is provided, \code{X}, and \code{y} are not needed.
If \code{TSS = NULL}, \code{TSS} is computed using \code{\link[=TSS]{TSS()}}
with \code{y} as r equired argument.
If \code{TSS} is provided, \code{y} is not needed.
}
\references{
\href{https://en.wikipedia.org/wiki/Residual_sum_of_squares}{Wikipedia: Residual Sum of Squares}

\href{https://en.wikipedia.org/wiki/Explained_sum_of_squares}{Wikipedia: Explained Sum of Squares}

\href{https://en.wikipedia.org/wiki/Total_sum_of_squares}{Wikipedia: Total Sum of Squares}

\href{https://en.wikipedia.org/wiki/Coefficient_of_determination}{Wikipedia: Coefficient of Determination}
}
\seealso{
Other assessment of model quality functions: 
\code{\link{.MSE}()},
\code{\link{.R2fromESS}()},
\code{\link{.R2fromRSS}()},
\code{\link{.RMSE}()},
\code{\link{.Rbar2}()},
\code{\link{MSE}()},
\code{\link{R2}()},
\code{\link{RMSE}()},
\code{\link{Rbar2}()},
\code{\link{model}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{assessment of model quality functions}
\keyword{model-assessment}
