% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yhat.R
\name{.Py}
\alias{.Py}
\title{y-hat \eqn{\left( \mathbf{\hat{y}} = \mathbf{P} \mathbf{y} \right)}}
\usage{
.Py(y, P = NULL, X = NULL)
}
\arguments{
\item{y}{Numeric vector of length \code{n} or \code{n} by \code{1} matrix.
The vector \eqn{\mathbf{y}} is an \eqn{n \times 1} vector of observations
on the regressand variable.}

\item{P}{\code{n} by \code{n} numeric matrix.
The \eqn{n \times n} projection matrix \eqn{\left( \mathbf{P} \right)}.}

\item{X}{\code{n} by \code{k} numeric matrix.
The data matrix \eqn{\mathbf{X}}
(also known as design matrix, model matrix or regressor matrix)
is an \eqn{n \times k} matrix of \eqn{n} observations of \eqn{k} regressors,
which includes a regressor whose value is 1 for each observation on the first column.}
}
\value{
Returns y-hat \eqn{\left( \mathbf{\hat{y}} \right)}.
}
\description{
Calculates y-hat \eqn{\left( \mathbf{\hat{y}} \right)},
that is, the predicted value of \eqn{\mathbf{y}} given \eqn{\mathbf{X}}
using
\deqn{
    \mathbf{\hat{y}} = \mathbf{P} \mathbf{y}
  }
where
\deqn{
    \mathbf{P} = \mathbf{X} \left( \mathbf{X}^{T} \mathbf{X} \right)^{-1}
    \mathbf{X}^{T} .
  }
}
\details{
If \code{P = NULL}, the \code{P} matrix is computed using \code{\link[=P]{P()}}
with \code{X} as its argument. If \code{P} is provided, \code{X} is not needed.
}
\references{
\href{https://en.wikipedia.org/wiki/Linear_regression}{Wikipedia: Linear Regression}

\href{https://en.wikipedia.org/wiki/Ordinary_least_squares}{Wikipedia: Ordinary Least Squares}
}
\seealso{
Other y-hat functions: 
\code{\link{.Xbetahat}()},
\code{\link{Py}()},
\code{\link{Xbetahat}()},
\code{\link{yhat}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{y-hat functions}
\keyword{predicted}
