% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R2.R
\name{.R2_ESS}
\alias{.R2_ESS}
\title{R-square
(from
\eqn{
  ESS
}
)}
\usage{
.R2_ESS(ESS = NULL, TSS = NULL, X = NULL, y = NULL, betahat = NULL)
}
\arguments{
\item{ESS}{Numeric.
Explained sum of squares.}

\item{TSS}{Numeric.
Total sum of squares.}

\item{X}{Matrix.
The data matrix
\eqn{\mathbf{X}}
(also known as
design matrix,
model matrix or
regressor matrix)
is an \eqn{n \times k} matrix
of \eqn{n} observations
of \eqn{k} regressors,
which includes a regressor
whose value is 1 for each observation.}

\item{y}{Vector or \code{n} by \code{1} matrix.
The vector
\eqn{\mathbf{y}}
is an \eqn{n \times 1} vector
of observations on the regressand variable.}

\item{betahat}{Vector or \code{k} by \code{1} matrix.
The vector
\eqn{\boldsymbol{\hat{\beta}}}
is a \eqn{k \times 1} vector of estimates
of \eqn{k} unknown regression coefficients.}
}
\value{
Returns the coefficient of determination
\eqn{
  R^2
}.
}
\description{
Calculates the coefficient of determination using
\deqn{
  R^2
  =
  \frac{
    \textrm{
      Explained sum of squares
    }
  }
  {
    \textrm{
      Total sum of squares
    }
  } .
}
}
\details{
If \code{ESS = NULL},
\code{ESS} is computed
using \code{\link[=ESS]{ESS()}}
with \code{X} and \code{y} as required arguments
and \code{betahat} as an optional argument.
If \code{TSS = NULL},
\code{TSS} is computed
using \code{\link[=TSS]{TSS()}}
with \code{y} as a required argument.
If \code{ESS} and \code{TSS} are provided,
\code{betahat}, \code{X}, and \code{y}
are not needed.
}
\references{
\href{https://en.wikipedia.org/wiki/Residual_sum_of_squares}{Wikipedia: Residual Sum of Squares}

\href{https://en.wikipedia.org/wiki/Explained_sum_of_squares}{Wikipedia: Explained Sum of Squares}

\href{https://en.wikipedia.org/wiki/Total_sum_of_squares}{Wikipedia: Total Sum of Squares}

\href{https://en.wikipedia.org/wiki/Coefficient_of_determination}{Wikipedia: Coefficient of Determination}
}
\seealso{
Other assessment of model quality functions: 
\code{\link{.MSE}()},
\code{\link{.R2_RSS}()},
\code{\link{.RMSE}()},
\code{\link{.Rbar2}()},
\code{\link{MSE}()},
\code{\link{R2}()},
\code{\link{RMSE}()},
\code{\link{Rbar2}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{assessment of model quality functions}
\keyword{coefficient}
\keyword{determination}
\keyword{of}
