% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R2.R
\name{.R2fromRSS}
\alias{.R2fromRSS}
\title{Coefficient of Determination \eqn{R^2} (from \eqn{RSS})}
\usage{
.R2fromRSS(RSS = NULL, TSS = NULL, epsilonhat = NULL, X, y, betahat = NULL)
}
\arguments{
\item{RSS}{Numeric.
Residual sum of squares.}

\item{TSS}{Numeric.
Total sum of squares.}

\item{epsilonhat}{Numeric vector of length \code{n} or \code{n} by 1 matrix.
\eqn{n \time 1} vector of residuals.}

\item{X}{\code{n} by \code{k} numeric matrix.
The data matrix \eqn{\mathbf{X}}
(also known as design matrix, model matrix or regressor matrix)
is an \eqn{n \times k} matrix of \eqn{n} observations of \eqn{k} regressors,
which includes a regressor whose value is 1 for each observation on the first column.}

\item{y}{Numeric vector of length \code{n} or \code{n} by \code{1} matrix.
The vector \eqn{\mathbf{y}} is an \eqn{n \times 1} vector of observations
on the regressand variable.}

\item{betahat}{Vector of length \code{k} or \code{k} by \code{1} matrix.
The vector \eqn{\boldsymbol{\hat{\beta}}} is a \eqn{k \times 1} vector of estimates
of \eqn{k} unknown regression coefficients.}
}
\value{
Returns the coefficient of determination \eqn{R^2} .
}
\description{
Calculates the coefficient of determination using
\deqn{
    R^2 = 1 - \frac{\textrm{Residual sum of squares}}
    {\textrm{Total sum of squares}} .
  }
}
\details{
If \code{RSS = NULL}, \code{RSS} is computed using \code{\link[=.RSS]{.RSS()}}.
If \code{TSS = NULL}, \code{TSS} is computed using \code{\link[=TSS]{TSS()}} with \code{y} as a required argument.
If \code{RSS} and \code{TSS} are provided, \code{epsilonhat}, \code{betahat}, \code{X}, and \code{y} are not needed.
}
\references{
\href{https://en.wikipedia.org/wiki/Residual_sum_of_squares}{Wikipedia: Residual Sum of Squares}

\href{https://en.wikipedia.org/wiki/Explained_sum_of_squares}{Wikipedia: Explained Sum of Squares}

\href{https://en.wikipedia.org/wiki/Total_sum_of_squares}{Wikipedia: Total Sum of Squares}

\href{https://en.wikipedia.org/wiki/Coefficient_of_determination}{Wikipedia: Coefficient of Determination}
}
\seealso{
Other coefficient of determination functions: 
\code{\link{.R2fromESS}()},
\code{\link{.Rbar2}()},
\code{\link{R2}()},
\code{\link{Rbar2}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{coefficient of determination functions}
\keyword{R-squared}
