% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R2.R
\name{.Rbar2}
\alias{.Rbar2}
\title{Adjusted R-squared \eqn{\bar{R}^{2}} (from \eqn{R^2})}
\usage{
.Rbar2(R2 = NULL, n, k, X, y, fromRSS = TRUE)
}
\arguments{
\item{R2}{Numeric.
Coefficient of determination \eqn{R^2} .}

\item{n}{Integer.
Sample size.}

\item{k}{Integer.
Number of regressors including a regressor whose value is 1 for each observation.}

\item{X}{\code{n} by \code{k} numeric matrix.
The data matrix \eqn{\mathbf{X}}
(also known as design matrix, model matrix or regressor matrix)
is an \eqn{n \times k} matrix of \eqn{n} observations of \eqn{k} regressors,
which includes a regressor whose value is 1 for each observation on the first column.}

\item{y}{Numeric vector of length \code{n} or \code{n} by \code{1} matrix.
The vector \eqn{\mathbf{y}} is an \eqn{n \times 1} vector of observations
on the regressand variable.}

\item{fromRSS}{Logical.
If \code{TRUE}, calculates the coefficient of determination from \code{RSS}.
If \code{FALSE}, calculates the coefficient of determination from \code{ESS}.}
}
\value{
Returns the adjusted coefficient of determination \eqn{\bar{R}^{2}} .
}
\description{
Calculates the adjusted coefficient of determination
\deqn{
    \bar{R}^{2} = 1 - \left(\frac{RSS / \left( n - k \right)}
    {TSS / \left( n - 1 \right)} \right)
    = 1 - \left( 1 - R^2 \right) \frac{n - 1}{n - k} .
  }
}
\details{
If \code{R2 = NULL}, \code{R2} is computed using \code{\link[=R2]{R2()}}
with \code{X} and \code{y} as required arguments.
If \code{R2} is provided, \code{X}, and \code{y} are not needed.
}
\references{
\href{https://en.wikipedia.org/wiki/Residual_sum_of_squares}{Wikipedia: Residual Sum of Squares}

\href{https://en.wikipedia.org/wiki/Explained_sum_of_squares}{Wikipedia: Explained Sum of Squares}

\href{https://en.wikipedia.org/wiki/Total_sum_of_squares}{Wikipedia: Total Sum of Squares}

\href{https://en.wikipedia.org/wiki/Coefficient_of_determination}{Wikipedia: Coefficient of Determination}
}
\seealso{
Other assessment of model quality functions: 
\code{\link{.MSE}()},
\code{\link{.R2fromESS}()},
\code{\link{.R2fromRSS}()},
\code{\link{.RMSE}()},
\code{\link{.model}()},
\code{\link{MSE}()},
\code{\link{R2}()},
\code{\link{RMSE}()},
\code{\link{Rbar2}()},
\code{\link{model}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{assessment of model quality functions}
\keyword{model-assessment}
