% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sehat.R
\name{.sehatbetahatbiased}
\alias{.sehatbetahatbiased}
\title{Standard Errors of Estimates of Regression Coefficients (from \eqn{\hat{\sigma}_{\varepsilon \ \textrm{biased}}^{2}})}
\usage{
.sehatbetahatbiased(vcovhatbetahatbiased = NULL, X, y)
}
\arguments{
\item{vcovhatbetahatbiased}{\code{k} by \code{k} matrix.
\eqn{k \times k} biased variance-covariance matrix of estimates of regression coefficients.}

\item{X}{\code{n} by \code{k} numeric matrix.
The data matrix \eqn{\mathbf{X}}
(also known as design matrix, model matrix or regressor matrix)
is an \eqn{n \times k} matrix of \eqn{n} observations of \eqn{k} regressors,
which includes a regressor whose value is 1 for each observation on the first column.}

\item{y}{Numeric vector of length \code{n} or \code{n} by \code{1} matrix.
The vector \eqn{\mathbf{y}} is an \eqn{n \times 1} vector of observations
on the regressand variable.}
}
\description{
Standard Errors of Estimates of Regression Coefficients (from \eqn{\hat{\sigma}_{\varepsilon \ \textrm{biased}}^{2}})
}
\seealso{
Other standard errors of estimates of regression coefficients functions: 
\code{\link{.sehatbetahat}()},
\code{\link{.sehatslopesprimedelta}()},
\code{\link{.sehatslopesprimetb}()},
\code{\link{sehatbetahatbiased}()},
\code{\link{sehatbetahat}()},
\code{\link{sehatslopesprimedelta}()},
\code{\link{sehatslopesprimetb}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{standard errors of estimates of regression coefficients functions}
\keyword{inference}
