% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sehat.R
\name{.sehatbetahatprimebiased}
\alias{.sehatbetahatprimebiased}
\title{Standard Errors of Standardized Estimates of Regression Coefficients (Biased)}
\usage{
.sehatbetahatprimebiased(
  betahat = NULL,
  sehatbetahat = NULL,
  betahatprime = NULL,
  X,
  y
)
}
\arguments{
\item{betahat}{Numeric vector of length \code{k} or \code{k} by \code{1} matrix.
The vector \eqn{\boldsymbol{\hat{\beta}}} is a \eqn{k \times 1} vector of estimates
of \eqn{k} unknown regression coefficients.}

\item{sehatbetahat}{Numeric vector of length \code{k} or \code{k} by \code{1} matrix.
Standard errors of standardized estimates of regression coefficients.}

\item{betahatprime}{Numeric vector of length \code{k} or \code{k} by \code{1} matrix.
The vector \eqn{\boldsymbol{\hat{\beta}^{\prime}}} is a \eqn{k \times 1} vector of standardized estimates
of \eqn{k} unknown regression coefficients.
The first item (intercept) is equal to zero.}

\item{X}{\code{n} by \code{k} numeric matrix.
The data matrix \eqn{\mathbf{X}}
(also known as design matrix, model matrix or regressor matrix)
is an \eqn{n \times k} matrix of \eqn{n} observations of \eqn{k} regressors,
which includes a regressor whose value is 1 for each observation on the first column.}

\item{y}{Numeric vector of length \code{n} or \code{n} by \code{1} matrix.
The vector \eqn{\mathbf{y}} is an \eqn{n \times 1} vector of observations
on the regressand variable.}
}
\description{
\deqn{
    \mathrm{\widehat{se}}_{\hat{\beta}^{\prime}} =
    \mathrm{\widehat{se}}_{\hat{\beta}} \frac{\hat{\beta}^{\prime}}{\hat{\beta}}
  }
According to Yuan and Chan (2011), this standard error is biased.
}
\seealso{
Other standard errors of estimates of regression coefficients functions: 
\code{\link{.sehatbetahatbiased}()},
\code{\link{.sehatbetahat}()},
\code{\link{sehatbetahatbiased}()},
\code{\link{sehatbetahatprimebiased}()},
\code{\link{sehatbetahat}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{standard errors of estimates of regression coefficients functions}
\keyword{inference}
