% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sehat.R
\name{.sehatslopeshatprimedelta}
\alias{.sehatslopeshatprimedelta}
\title{Standard Errors of Standardized Estimates of Regression Coefficients (Yuan and Chan (2011))}
\usage{
.sehatslopeshatprimedelta(
  slopeshat,
  sigma2hatepsilonhat,
  SigmaXhat,
  sigmayXhat,
  sigma2yhat,
  adjust = FALSE,
  n,
  X,
  y
)
}
\arguments{
\item{slopeshat}{Numeric vector of length \code{p} or \code{p} by \code{1} matrix.
\eqn{p \times 1} column vector of estimated regression slopes \eqn{\left( \boldsymbol{\hat{\beta}}_{2, 3, \cdots, k} = \left\{ \hat{\beta}_2, \hat{\beta}_3, \cdots, \hat{\beta}_k \right\}^{T} \right)} .}

\item{sigma2hatepsilonhat}{Numeric.
Estimate of error variance.}

\item{SigmaXhat}{\code{p} by \code{p} numeric matrix.
\eqn{p \times p} matrix of estimated variances and covariances between regressor variables
\eqn{X_2, X_3, \cdots, X_k}
\eqn{\left( \boldsymbol{\hat{\Sigma}}_{\mathbf{X}} \right)}.}

\item{sigmayXhat}{Numeric vector of length \code{p} or \code{p} by \code{1} matrix.
\eqn{p \times 1} vector of estimated covariances between the regressand \eqn{y} variable
and regressor variables \eqn{X_2, X_3, \cdots, X_k}
\eqn{\left( \boldsymbol{\hat{\sigma}}_{\mathbf{y}, \mathbf{X}}
    = \left\{ \hat{\sigma}_{y, X_2}, \hat{\sigma}_{y, X_3}, \cdots, \hat{\sigma}_{y, X_k} \right\}^{T} \right)}.}

\item{sigma2yhat}{Numeric.
Estimated variance of the regressand \eqn{\left( \hat{\sigma}_{y}^{2} \right)}}

\item{adjust}{Logical.
Use \eqn{n - 3} adjustment for small samples.}

\item{n}{Integer.
Sample size.}

\item{X}{\code{n} by \code{k} numeric matrix.
The data matrix \eqn{\mathbf{X}}
(also known as design matrix, model matrix or regressor matrix)
is an \eqn{n \times k} matrix of \eqn{n} observations of \eqn{k} regressors,
which includes a regressor whose value is 1 for each observation on the first column.}

\item{y}{Numeric vector of length \code{n} or \code{n} by \code{1} matrix.
The vector \eqn{\mathbf{y}} is an \eqn{n \times 1} vector of observations
on the regressand variable.}
}
\description{
Standard Errors of Standardized Estimates of Regression Coefficients (Yuan and Chan (2011))
}
\details{
The \eqn{p}th estimated standard error is calculated using
\deqn{
    \mathbf{\widehat{se}}_{\boldsymbol{\hat{\beta}}_{\text{p}}^{\prime}} =
    \sqrt{
      \frac{\hat{\sigma}_{X_{p}}^{2} \hat{c}_{p} \hat{\sigma}_{\hat{\varepsilon}}^{2}}{n \hat{\sigma}_{y}^{2}} +
    \frac{\hat{\beta}_{p}
      \left[
        \hat{\sigma}_{X_{p}}^{2}
        \left( \boldsymbol{\hat{\beta}}^{T} \boldsymbol{\hat{\Sigma}}_{X} \boldsymbol{\hat{\beta}} \right) -
        \hat{\sigma}_{X_{p}}^{2} \hat{\sigma}_{\hat{\varepsilon}}^{2} -
        \hat{\sigma}_{y, X_{p}}^{2}
      \right]}{n \hat{\sigma}_{y}^{4}}
    }
  }
where
\itemize{
\item \eqn{p = \left\{2, 3, \cdots, k \right\}}
\item \eqn{\hat{\sigma}_{\hat{\varepsilon}}^{2}} is the estimated residual variance
\item \eqn{\boldsymbol{\hat{\beta}}_{2, 3, \cdots, k} = \left\{ \hat{\beta}_{2}, \hat{\beta}_{3}, \cdots, \hat{\beta}_{k}\right\}^{T}}
is the \eqn{p \times 1} column vector of estimated regression slopes
\item \eqn{\hat{\sigma}_{y}^{2}} is the variance of the regressand variable \eqn{y}
\item \eqn{\boldsymbol{\hat{\Sigma}}_{\mathbf{X}}} is the \eqn{p \times p} estimated covariance matrix of the regressor variables \eqn{X_2, X_3, \cdots, X_k}
\item \eqn{\hat{\sigma}_{X_p}^{2}} is the variance of the corresponding \eqn{p}th regressor variable.
\item \eqn{\hat{\sigma}_{y, X_{p}}^{2}} is the covariance of the regressand variable \eqn{y} and the regressor variables \eqn{X_2, X_3, \cdots, X_k}
\item \eqn{c_p} is the diagonal element that corresponds to the regressor variable in \eqn{\boldsymbol{\Sigma}_{\mathbf{X}}^{-1}}
\item \eqn{n} is the sample size
}
}
\examples{
slopes <- c(-3.0748755, -1.5653133, 1.0959758, 1.3703010, 0.1666065)
SigmaXhat <- matrix(
  data = c(
    0.25018672, 0.00779108, -0.01626038, -0.04424864, -0.13217068,
    0.00779108, 0.12957466, 0.01061297, -0.08818286, -0.16427222,
    -0.016260378, 0.010612975, 0.133848763, 0.004083767, 0.658462191,
    -0.044248635, -0.088182856, 0.004083767, 7.917601877, -5.910469742,
    -0.1321707, -0.1642722, 0.6584622, -5.9104697, 136.0217584
  ),
  ncol = 5
)
sigma2hatepsilonhat <- 42.35584
sigma2yhat <- 62.35235
sigmayXhat <- c(-0.8819639, -0.3633559, 0.2953811, 10.1433433, 15.9481950)
n <- 1289
.sehatslopeshatprimedelta(
  slopes = slopes, sigma2hatepsilonhat = sigma2hatepsilonhat,
  SigmaXhat = SigmaXhat, sigma2yhat = sigma2yhat, sigmayXhat = sigmayXhat, n = n
)
}
\references{
Yuan, K., Chan, W. (2011).
Biases and Standard Errors of Standardized Regression Coefficients.
\emph{Psychometrika} 76, 670-690.
\href{https://doi.org/10.1007/s11336-011-9224-6}{doi:10.1007/s11336-011-9224-6}.
}
\seealso{
Other standard errors of estimates of regression coefficients functions: 
\code{\link{.sehatbetahatbiased}()},
\code{\link{.sehatbetahat}()},
\code{\link{.sehatslopeshatprimetb}()},
\code{\link{sehatbetahatbiased}()},
\code{\link{sehatbetahat}()},
\code{\link{sehatslopeshatprimedelta}()},
\code{\link{sehatslopeshatprimetb}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{standard errors of estimates of regression coefficients functions}
\keyword{se}
