% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sehat.R
\name{.sehatslopesprimedelta}
\alias{.sehatslopesprimedelta}
\title{Standard Errors of Standardized Estimates of Regression Coefficients (Yuan and Chan (2011))}
\usage{
.sehatslopesprimedelta(
  slopes,
  sigma2hatepsilonhat,
  SigmaXhat,
  sigmayXhat,
  sigma2yhat,
  adjust = FALSE,
  n,
  X,
  y
)
}
\arguments{
\item{slopes}{Numeric vector of length \code{p} or \code{p} by \code{1} matrix.
The vector \eqn{\boldsymbol{\hat{\beta}_{2, \cdots, k}}} is a \eqn{p \times 1} vector of estimates
of \eqn{p} unknown regression slopes.}

\item{sigma2hatepsilonhat}{Numeric.
Estimate of error variance.}

\item{SigmaXhat}{\code{p} by \code{p} numeric matrix.
Estimated \eqn{p \times p} variances and covariances of \eqn{{X}_{2}, {X}_{3}, \cdots, {X}_{k}}
\eqn{\left( \boldsymbol{\hat{\Sigma}}_{\mathbf{X}} \right)}.}

\item{sigmayXhat}{Numeric vector of length \code{p} or \code{p} by \code{1} matrix.
Estimated \eqn{p \times 1} covariances between the regressand and the regressors
\eqn{\left( \boldsymbol{\sigma}_{\mathbf{yX}} \right)}.}

\item{sigma2yhat}{Numeric.
Estimated variance of the regressand \eqn{\left( \hat{\sigma}_{y} \right)}}

\item{adjust}{Logical.
Use \eqn{n - 3} adjustment for small samples.}

\item{n}{Integer.
Sample size.}

\item{X}{\code{n} by \code{k} numeric matrix.
The data matrix \eqn{\mathbf{X}}
(also known as design matrix, model matrix or regressor matrix)
is an \eqn{n \times k} matrix of \eqn{n} observations of \eqn{k} regressors,
which includes a regressor whose value is 1 for each observation on the first column.}

\item{y}{Numeric vector of length \code{n} or \code{n} by \code{1} matrix.
The vector \eqn{\mathbf{y}} is an \eqn{n \times 1} vector of observations
on the regressand variable.}
}
\description{
Standard Errors of Standardized Estimates of Regression Coefficients (Yuan and Chan (2011))
}
\examples{
slopes <- c(-3.0748755, -1.5653133, 1.0959758, 1.3703010, 0.1666065)
SigmaXhat <- matrix(
  data = c(
    0.25018672, 0.00779108, -0.01626038, -0.04424864, -0.13217068,
    0.00779108, 0.12957466, 0.01061297, -0.08818286, -0.16427222,
    -0.016260378, 0.010612975, 0.133848763, 0.004083767, 0.658462191,
    -0.044248635, -0.088182856, 0.004083767, 7.917601877, -5.910469742,
    -0.1321707, -0.1642722, 0.6584622, -5.9104697, 136.0217584
  ),
  ncol = 5
)
sigma2hatepsilonhat <- 42.35584
sigma2yhat <- 62.35235
sigmayXhat <- c(-0.8819639, -0.3633559, 0.2953811, 10.1433433, 15.9481950)
n <- 1289
.sehatslopesprimedelta(
  slopes = slopes, sigma2hatepsilonhat = sigma2hatepsilonhat,
  SigmaXhat = SigmaXhat, sigma2yhat = sigma2yhat, sigmayXhat = sigmayXhat, n = n
)
}
\seealso{
Other standard errors of estimates of regression coefficients functions: 
\code{\link{.sehatbetahatbiased}()},
\code{\link{.sehatbetahat}()},
\code{\link{.sehatslopesprimetb}()},
\code{\link{sehatbetahatbiased}()},
\code{\link{sehatbetahat}()},
\code{\link{sehatslopesprimedelta}()},
\code{\link{sehatslopesprimetb}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{standard errors of estimates of regression coefficients functions}
\keyword{inference}
