% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sehat.R
\name{.sehatslopesprimetb}
\alias{.sehatslopesprimetb}
\title{Standard Errors of Standardized Estimates of Regression Coefficients (Textbook)}
\usage{
.sehatslopesprimetb(
  slopes = NULL,
  sehatslopes = NULL,
  slopesprime = NULL,
  X,
  y
)
}
\arguments{
\item{slopes}{Numeric vector of length \code{p} or \code{p} by \code{1} matrix.
The vector \eqn{\boldsymbol{\hat{\beta}_{2, \cdots, k}}} is a \eqn{p \times 1} vector of estimates
of \eqn{p} unknown regression slopes.}

\item{sehatslopes}{Numeric vector of length \code{p} or \code{p} by \code{1} matrix.
Standard errors of estimates of regression slopes.}

\item{slopesprime}{Numeric vector of length \code{p} or \code{p} by \code{1} matrix.
The vector \eqn{\boldsymbol{\hat{\beta}_{2, \cdots, k}^{\prime}}} is a \eqn{p \times 1} vector of standardized estimates
of \eqn{p} unknown regression slopes.}

\item{X}{\code{n} by \code{k} numeric matrix.
The data matrix \eqn{\mathbf{X}}
(also known as design matrix, model matrix or regressor matrix)
is an \eqn{n \times k} matrix of \eqn{n} observations of \eqn{k} regressors,
which includes a regressor whose value is 1 for each observation on the first column.}

\item{y}{Numeric vector of length \code{n} or \code{n} by \code{1} matrix.
The vector \eqn{\mathbf{y}} is an \eqn{n \times 1} vector of observations
on the regressand variable.}
}
\description{
\deqn{
    \mathbf{\widehat{se}}_{\boldsymbol{\hat{\beta}}_{2, \cdots, k}^{\prime}} =
    \mathbf{\widehat{se}}_{\boldsymbol{\hat{\beta}}_{2, \cdots, k}} \frac{\boldsymbol{\hat{\beta}}_{2, \cdots, k}^{\prime}}{\boldsymbol{\hat{\beta}}_{2, \cdots, k}}
  }
According to Yuan and Chan (2011), this standard error is biased.
}
\seealso{
Other standard errors of estimates of regression coefficients functions: 
\code{\link{.sehatbetahatbiased}()},
\code{\link{.sehatbetahat}()},
\code{\link{sehatbetahatbiased}()},
\code{\link{sehatbetahat}()},
\code{\link{sehatslopesprimetb}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{standard errors of estimates of regression coefficients functions}
\keyword{inference}
