% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigma2epsilonhat.R
\name{.sigma2epsilonhatbiased}
\alias{.sigma2epsilonhatbiased}
\title{Residual Variance \eqn{\hat{\sigma}_{\varepsilon \ \textrm{biased}}^{2}} (from \eqn{RSS})}
\usage{
.sigma2epsilonhatbiased(RSS = NULL, n, X, y)
}
\arguments{
\item{RSS}{Numeric.
Residual sum of squares.}

\item{n}{Integer.
Sample size.}

\item{X}{\code{n} by \code{k} numeric matrix.
The data matrix \eqn{\mathbf{X}}
(also known as design matrix, model matrix or regressor matrix)
is an \eqn{n \times k} matrix of \eqn{n} observations of \eqn{k} regressors,
which includes a regressor whose value is 1 for each observation on the first column.}

\item{y}{Numeric vector of length \code{n} or \code{n} by \code{1} matrix.
The vector \eqn{\mathbf{y}} is an \eqn{n \times 1} vector of observations
on the regressand variable.}
}
\value{
Returns the estimated residual variance \eqn{\hat{\sigma}_{\varepsilon \ \textrm{biased}}^{2}} .
}
\description{
Calculates an estimate of the error variance
\deqn{
    \mathbf{E} \left( \sigma^2 \right) = \hat{\sigma}_{\varepsilon}^{2}
  }
\deqn{
    \hat{\sigma}_{\varepsilon \ \textrm{biased}}^{2}
    = \frac{1}{n} \sum_{i = 1}^{n} \left( \mathbf{y} - \mathbf{X} \boldsymbol{\hat{\beta}} \right)^2 \\
    = \frac{\boldsymbol{\varepsilon}^{\prime} \boldsymbol{\varepsilon}}{n} \\
    = \frac{RSS}{n}
  }
where
\eqn{\boldsymbol{\varepsilon}} is the vector of residuals,
\eqn{RSS} is the residual sum of squares, and
\eqn{n} is the sample size.
}
\details{
If \code{RSS = NULL}, \code{RSS} is computed using \code{\link[=RSS]{RSS()}}.
If \code{RSS} is provided, \code{X}, and \code{y} are not needed.
}
\references{
\href{https://en.wikipedia.org/wiki/Linear_regression}{Wikipedia: Linear Regression}

\href{https://en.wikipedia.org/wiki/Ordinary_least_squares}{Wikipedia: Ordinary Least Squares}
}
\seealso{
Other residual variance functions: 
\code{\link{.sigma2epsilonhat}()},
\code{\link{sigma2epsilonhatbiased}()},
\code{\link{sigma2epsilonhat}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{residual variance functions}
\keyword{sigma2epsilonhat}
