% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcov.R
\name{.sigma2hat}
\alias{.sigma2hat}
\title{Residual Variance (from \eqn{RSS})}
\usage{
.sigma2hat(
  RSS = NULL,
  n,
  k,
  type = "unbiased",
  X = NULL,
  y = NULL,
  betahat = NULL
)
}
\arguments{
\item{RSS}{Numeric.
Residual sum of squares.}

\item{n}{Integer.
Sample size.}

\item{k}{Integer.
Number of regressors
including a regressor
whose value is 1 for each observation.}

\item{type}{String.
Residual variance estimator.
If \code{type = "unbiased"},
returns unbiased estimate.
If \code{type = "biased"},
returns biased estimate.
If \code{type = "both"},
returns a vector of
unbiased and biased estimates.}

\item{X}{Matrix.
The data matrix
\eqn{\mathbf{X}}
(also known as
design matrix,
model matrix or
regressor matrix)
is an \eqn{n \times k} matrix
of \eqn{n} observations
of \eqn{k} regressors,
which includes a regressor
whose value is 1 for each observation.}

\item{y}{Vector or \code{n} by \code{1} matrix.
The vector
\eqn{\mathbf{y}}
is an \eqn{n \times 1} vector
of observations on the regressand variable.}

\item{betahat}{Vector or \code{k} by \code{1} matrix.
The vector
\eqn{\boldsymbol{\hat{\beta}}}
is a \eqn{k \times 1} vector of estimates
of \eqn{k} unknown regression coefficients.}
}
\value{
Returns the estimated
residual variance (\eqn{\hat{\sigma}^2}).
}
\description{
Calculates estimates of the error variance
\deqn{
    \mathbf{E}
      \left(
        \sigma^2
      \right)
  =
  \hat{\sigma}^2
  }
\deqn{
    \hat{\sigma}^2_{\textrm{unbiased}}
    =
    \frac{1}{n - k}
    \sum_{i = 1}^{n}
    \left(
      \mathbf{y}
      -
      \mathbf{X} \boldsymbol{\hat{\beta}}
    \right)^2 \\
    =
    \frac{
    \mathbf{e}^{\prime}
    \mathbf{e}
    }
    {
      n - k
    } \\
    =
    \frac{
    RSS
    }
    {
      n - k
    }
  }
or
\deqn{
    \hat{\sigma}^2_{\textrm{biased}}
    =
    \frac{1}{n}
    \sum_{i = 1}^{n}
    \left(
      \mathbf{y}
      -
      \mathbf{X} \boldsymbol{\hat{\beta}}
    \right)^2 \\
    =
    \frac{
    \mathbf{e}^{\prime}
    \mathbf{e}
    }
    {
      n
    } \\
    =
    \frac{
    RSS
    }
    {
      n
    }
  }
where
\eqn{\mathbf{e}} is the vector of residuals,
\eqn{RSS} is the residual sum of squares,
\eqn{n} is the sample size,
and \eqn{k} is the number of regressors
including a regressor whose value is 1 for each observation.
}
\details{
If \code{RSS = NULL},
\code{RSS} is computed
using \code{\link[=RSS]{RSS()}}
with \code{X} and \code{y} as a required arguments
and \code{betahat} as an optional argument.
If \code{RSS} is provided,
\code{betahat}, \code{X}, and \code{y}
are not needed.
}
\references{
\href{https://en.wikipedia.org/wiki/Linear_regression}{Wikipedia: Linear Regression}

\href{https://en.wikipedia.org/wiki/Ordinary_least_squares}{Wikipedia: Ordinary Least Squares}
}
\seealso{
Other residual variance functions: 
\code{\link{sigma2hat}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{residual variance functions}
