% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigma2hatepsilonhat.R
\name{.sigma2hatepsilonhat}
\alias{.sigma2hatepsilonhat}
\title{Residual Variance \eqn{\hat{\sigma}_{\hat{\varepsilon}}^{2}} (from \eqn{\mathrm{RSS}})}
\usage{
.sigma2hatepsilonhat(RSS = NULL, n, k, X, y)
}
\arguments{
\item{RSS}{Numeric.
Residual sum of squares.}

\item{n}{Integer.
Sample size.}

\item{k}{Integer.
Number of regressors including a regressor
whose value is 1 for each observation on the first column.}

\item{X}{\code{n} by \code{k} numeric matrix.
The data matrix \eqn{\mathbf{X}}
(also known as design matrix, model matrix or regressor matrix)
is an \eqn{n \times k} matrix of \eqn{n} observations of \eqn{k} regressors,
which includes a regressor whose value is 1 for each observation on the first column.}

\item{y}{Numeric vector of length \code{n} or \code{n} by \code{1} matrix.
The vector \eqn{\mathbf{y}} is an \eqn{n \times 1} vector of observations
on the regressand variable.}
}
\value{
Returns the estimated residual variance \eqn{\hat{\sigma}_{\hat{\varepsilon}}^{2}} .
}
\description{
Calculates an estimate of the error variance
\deqn{
    \mathbf{E} \left( \sigma^2 \right) = \hat{\sigma}_{\hat{\varepsilon}}^{2}
  }
\deqn{
    \hat{\sigma}_{\hat{\varepsilon}}^{2}
    = \frac{1}{n - k} \sum_{i = 1}^{n} \left( \mathbf{y} - \mathbf{X} \boldsymbol{\hat{\beta}} \right)^2 \\
    = \frac{\boldsymbol{\hat{\varepsilon}}^{\prime} \boldsymbol{\hat{\varepsilon}}}{n - k} \\
    = \frac{\mathrm{RSS}}{n - k}}
where
\eqn{\boldsymbol{\hat{\varepsilon}}} is the vector of residuals,
\eqn{\mathrm{RSS}} is the residual sum of squares,
\eqn{n} is the sample size,
and \eqn{k} is the number of regressors
including a regressor whose value is 1 for each observation on the first column.
}
\details{
If \code{RSS = NULL}, \code{RSS} is computed using \code{\link[=RSS]{RSS()}}.
If \code{RSS} is provided, \code{X}, and \code{y} are not needed.
}
\references{
\href{https://en.wikipedia.org/wiki/Linear_regression}{Wikipedia: Linear Regression}

\href{https://en.wikipedia.org/wiki/Ordinary_least_squares}{Wikipedia: Ordinary Least Squares}
}
\seealso{
Other residual variance functions: 
\code{\link{.sigma2hatepsilonhatbiased}()},
\code{\link{sigma2hatepsilonhatbiased}()},
\code{\link{sigma2hatepsilonhat}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{residual variance functions}
\keyword{sigma2hatepsilonhat}
