% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betahat_matrix.R
\name{.slopeshatprime}
\alias{.slopeshatprime}
\title{Estimates of Regression Standardized Slopes \eqn{\boldsymbol{\hat{\beta}}_{2, \cdots, k}^{\prime}}}
\usage{
.slopeshatprime(RXhat = NULL, ryXhat = NULL, X, y)
}
\arguments{
\item{RXhat}{\code{p} by \code{p} numeric matrix.
\eqn{p \times p} matrix of estimates of correlations between the regressor variables \eqn{X_2, X_3, \cdots, X_k}
\eqn{\left( \mathbf{\hat{R}}_{\mathbf{X}} \right)}.}

\item{ryXhat}{Numeric vector of length \code{p} or \code{p} by \code{1} matrix.
\eqn{p \times 1} vector of estimates of correlations between the regressand variable \eqn{y} and the regressor variables \eqn{X_2, X_3, \cdots, X_k}
\eqn{\left( \mathbf{\hat{r}}_{\mathbf{y}, \mathbf{X}}
    = \left\{ \hat{r}_{y, X_2}, \hat{r}_{y, X_3}, \cdots, \hat{r}_{y, X_k} \right\}^{T} \right)}.}

\item{X}{\code{n} by \code{k} numeric matrix.
The data matrix \eqn{\mathbf{X}}
(also known as design matrix, model matrix or regressor matrix)
is an \eqn{n \times k} matrix of \eqn{n} observations of \eqn{k} regressors,
which includes a regressor whose value is 1 for each observation on the first column.}

\item{y}{Numeric vector of length \code{n} or \code{n} by \code{1} matrix.
The vector \eqn{\mathbf{y}} is an \eqn{n \times 1} vector of observations
on the regressand variable.}
}
\value{
Returns the estimated standardized slopes
\eqn{\boldsymbol{\hat{\beta}}_{2, \cdots, k}^{\prime}}
of a linear regression model derived from the estimated correlation matrix.
}
\description{
Estimates of Regression Standardized Slopes \eqn{\boldsymbol{\hat{\beta}}_{2, \cdots, k}^{\prime}}
}
\details{
Estimates of the linear regression standardized slopes are calculated using
\deqn{
    \boldsymbol{\hat{\beta}}_{2, \cdots, k}^{\prime} =
    \mathbf{\hat{R}}_{\mathbf{X}}^{T} \mathbf{\hat{r}}_{\mathbf{y}, \mathbf{X}}
  }

where
\itemize{
\item \eqn{\mathbf{\hat{R}}_{\mathbf{X}}}
is the \eqn{p \times p} estimated correlation matrix of the regressor variables \eqn{X_2, X_3, \cdots, X_k} and
\item \eqn{\mathbf{\hat{r}}_{\mathbf{y}, \mathbf{X}}}
is the \eqn{p \times 1} column vector
of the estimated correlations between the regressand \eqn{y} variable
and regressor variables \eqn{X_2, X_3, \cdots, X_k}
}
}
\seealso{
Other beta-hat functions: 
\code{\link{.betahatnorm}()},
\code{\link{.betahatqr}()},
\code{\link{.betahatsvd}()},
\code{\link{.intercepthat}()},
\code{\link{.slopeshat}()},
\code{\link{betahat}()},
\code{\link{intercepthat}()},
\code{\link{slopeshatprime}()},
\code{\link{slopeshat}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{beta-hat functions}
\keyword{beta-hat-ols}
