% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betahatinference.R
\name{.slopeshatprimeinference}
\alias{.slopeshatprimeinference}
\title{Standardized Regression Slopes Hypothesis Test and Confidence Intervals}
\usage{
.slopeshatprimeinference(
  slopeshatprime = NULL,
  sehatslopeshatprime = NULL,
  sehatslopeshatprimetype = "textbook",
  adjust = FALSE,
  n,
  X,
  y
)
}
\arguments{
\item{slopeshatprime}{Numeric vector of length \code{p} or \code{p} by \code{1} matrix.
\eqn{p \times 1} column vector of estimated standardized regression slopes
\eqn{\left( \boldsymbol{\hat{\beta}}_{2, 3, \cdots, k}  = \left\{ \hat{\beta}_2, \hat{\beta}_3, \cdots, \hat{\beta}_k \right\}^{T} \right)} .}

\item{sehatslopeshatprime}{Numeric vector of length \code{p} or \code{p} by \code{1} matrix.
Standard errors of estimates of standardized regression slopes.}

\item{sehatslopeshatprimetype}{Character string.
Standard errors for standardized regression slopes hypothesis test.
Options are \code{sehatslopeshatprimetype = "textbook"} and \code{sehatslopeshatprimetype = "delta"}.}

\item{adjust}{Logical.
If \code{sehatslopeshatprimetype = "delta"} and \code{adjust = TRUE},
uses \code{n - 3} to adjust \code{sehatslopeshatprime} for bias.
This adjustment is recommended for small sample sizes.}

\item{n}{Integer.
Sample size.}

\item{X}{\code{n} by \code{k} numeric matrix.
The data matrix \eqn{\mathbf{X}}
(also known as design matrix, model matrix or regressor matrix)
is an \eqn{n \times k} matrix of \eqn{n} observations of \eqn{k} regressors,
which includes a regressor whose value is 1 for each observation on the first column.}

\item{y}{Numeric vector of length \code{n} or \code{n} by \code{1} matrix.
The vector \eqn{\mathbf{y}} is an \eqn{n \times 1} vector of observations
on the regressand variable.}
}
\description{
Standardized Regression Slopes Hypothesis Test and Confidence Intervals
}
\seealso{
Other inference functions: 
\code{\link{.betahatinference}()},
\code{\link{betahatinference}()},
\code{\link{slopeshatprimeinference}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{inference functions}
\keyword{inference}
