% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta.R
\name{.slopesprime}
\alias{.slopesprime}
\title{Regression Standardized Slopes \eqn{\boldsymbol{\beta}_{2, \cdots, k}^{\prime}}}
\usage{
.slopesprime(RX = NULL, ryX = NULL, X, y)
}
\arguments{
\item{RX}{\code{p} by \code{p} numeric matrix.
\eqn{p \times p} matrix of correlations between the regressor variables \eqn{X_2, X_3, \cdots, X_k}
\eqn{\left( \mathbf{R}_{\mathbf{X}} \right)}.}

\item{ryX}{Numeric vector of length \code{p} or \code{p} by \code{1} matrix.
\eqn{p \times 1} vector of correlations between the regressand variable \eqn{y} and the regressor variables \eqn{X_2, X_3, \cdots, X_k}
\eqn{\left( \mathbf{r}_{\mathbf{y}, \mathbf{X}}
    = \left\{ r_{y, X_2}, r_{y, X_3}, \cdots, r_{y, X_k} \right\}^{T} \right)}.}

\item{X}{\code{n} by \code{k} numeric matrix.
The data matrix \eqn{\mathbf{X}}
(also known as design matrix, model matrix or regressor matrix)
is an \eqn{n \times k} matrix of \eqn{n} observations of \eqn{k} regressors,
which includes a regressor whose value is 1 for each observation on the first column.}

\item{y}{Numeric vector of length \code{n} or \code{n} by \code{1} matrix.
The vector \eqn{\mathbf{y}} is an \eqn{n \times 1} vector of observations
on the regressand variable.}
}
\value{
Returns the standardized slopes
\eqn{\boldsymbol{\beta}_{2, \cdots, k}^{\prime}}
of a linear regression model derived from the correlation matrix.
}
\description{
Derives the standardized slopes
\eqn{\boldsymbol{\beta}_{2, \cdots, k}^{\prime}} of a linear regression model
as a function of correlations.
}
\details{
The linear regression standardized slopes are calculated using
\deqn{
    \boldsymbol{\beta}_{2, \cdots, k}^{\prime} =
    \mathbf{R}_{\mathbf{X}}^{T} \mathbf{r}_{\mathbf{y}, \mathbf{X}}
  }

where
\itemize{
\item \eqn{\mathbf{R}_{\mathbf{X}}}
is the \eqn{p \times p} correlation matrix of the regressor variables \eqn{X_2, X_3, \cdots, X_k} and
\item \eqn{\mathbf{r}_{\mathbf{y}, \mathbf{X}}}
is the \eqn{p \times 1} column vector
of the correlations between the regressand \eqn{y} variable
and regressor variables \eqn{X_2, X_3, \cdots, X_k}
}
}
\seealso{
Other parameter functions: 
\code{\link{.intercept}()},
\code{\link{.slopes}()},
\code{\link{intercept}()},
\code{\link{sigma2epsilon}()},
\code{\link{slopesprime}()},
\code{\link{slopes}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{parameter functions}
\keyword{parameter}
