% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betahatinference.R
\name{.slopesprimeinference}
\alias{.slopesprimeinference}
\title{Standardized Regression Slopes Hypothesis Test and Confidence Intervals}
\usage{
.slopesprimeinference(
  slopesprime = NULL,
  sehatslopesprime = NULL,
  sehatslopesprimetype = "textbook",
  adjust = FALSE,
  n,
  X,
  y
)
}
\arguments{
\item{slopesprime}{Numeric vector of length \code{p} or \code{p} by \code{1} matrix.
The vector \eqn{\boldsymbol{\hat{\beta}_{2, \cdots, k}^{\prime}}} is a \eqn{p \times 1} vector of standardized estimates
of \eqn{p} unknown regression slopes.}

\item{sehatslopesprime}{Numeric vector of length \code{p} or \code{p} by \code{1} matrix.
Standard errors of estimates of standardized regression slopes.}

\item{sehatslopesprimetype}{Character string.
Standard errors for standardized regression slopes hypothesis test.
Options are \code{sehatslopesprimetype = "textbook"} and \code{sehatslopesprimetype = "delta"}.}

\item{adjust}{Logical.
If \code{sehatslopesprimetype = "delta"} and \code{adjust = TRUE},
uses \code{n - 3} to adjust \code{sehatslopesprime} for bias.
This adjustment is recommended for small sample sizes.}

\item{n}{Integer.
Sample size.}

\item{X}{\code{n} by \code{k} numeric matrix.
The data matrix \eqn{\mathbf{X}}
(also known as design matrix, model matrix or regressor matrix)
is an \eqn{n \times k} matrix of \eqn{n} observations of \eqn{k} regressors,
which includes a regressor whose value is 1 for each observation on the first column.}

\item{y}{Numeric vector of length \code{n} or \code{n} by \code{1} matrix.
The vector \eqn{\mathbf{y}} is an \eqn{n \times 1} vector of observations
on the regressand variable.}
}
\description{
Standardized Regression Slopes Hypothesis Test and Confidence Intervals
}
\seealso{
Other inference functions: 
\code{\link{.betahatinference}()},
\code{\link{betahatinference}()},
\code{\link{slopesprimeinference}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{inference functions}
\keyword{inference}
