% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcov.R
\name{.vcov_betahat}
\alias{.vcov_betahat}
\title{Variance-Covariance Matrix of Estimates of Regression Coefficients (from \eqn{\hat{\sigma}^2})}
\usage{
.vcov_betahat(
  sigma2hat = NULL,
  RSS = NULL,
  X,
  y,
  betahat = NULL,
  type = "unbiased"
)
}
\arguments{
\item{sigma2hat}{Numeric.
Estimate of error variance.}

\item{RSS}{Numeric.
Residual sum of squares.}

\item{X}{Matrix.
The data matrix
\eqn{\mathbf{X}}
(also known as
design matrix,
model matrix or
regressor matrix)
is an \eqn{n \times k} matrix
of \eqn{n} observations
of \eqn{k} regressors,
which includes a regressor
whose value is 1 for each observation.}

\item{y}{Vector or \code{n} by \code{1} matrix.
The vector
\eqn{\mathbf{y}}
is an \eqn{n \times 1} vector
of observations on the regressand variable.}

\item{betahat}{Vector or \code{k} by \code{1} matrix.
The vector
\eqn{\boldsymbol{\hat{\beta}}}
is a \eqn{k \times 1} vector of estimates
of \eqn{k} unknown regression coefficients.}

\item{type}{String.
Residual variance estimator.
If \code{type = "unbiased"},
returns unbiased estimate.
If \code{type = "biased"},
returns biased estimate.
If \code{type = "both"},
returns a vector of
unbiased and biased estimates.}
}
\value{
Returns the variance-covariance matrix
of estimates of regression coefficients.
}
\description{
Calculates the variance-covariance matrix
of estimates of regression coefficients using
\deqn{
    \hat{\sigma}^2
    \left(
      \mathbf{X}^{\prime}
      \mathbf{X}
    \right)^{-1}
  }
where \eqn{\hat{\sigma}^2}
is the estimate of the error variance \eqn{{\sigma}^2}
and \eqn{\mathbf{X}} is
the data matrix,
that is,
an \eqn{n \times k} matrix
of \eqn{n} observations
of \eqn{k} regressors,
which includes a regressor
whose value is 1 for each observation.
}
\details{
If \code{sigma2hat = NULL},
\code{sigma2hat} is computed
using \code{\link[=sigma2hat]{sigma2hat()}}
with \code{X} and \code{y} as a required arguments
and \code{betahat} and \code{RSS} as an optional argument.
If \code{sigma2hat} is provided,
\code{RSS}, \code{betahat}, \code{X}, and \code{y}
are not needed.
}
\references{
\href{https://en.wikipedia.org/wiki/Linear_regression}{Wikipedia: Linear Regression}

\href{https://en.wikipedia.org/wiki/Ordinary_least_squares}{Wikipedia: Ordinary Least Squares}
}
\seealso{
Other variance-covariance functions: 
\code{\link{vcov_betahat}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{variance-covariance functions}
