% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcovhat.R
\name{.vcovhatbetahatbiased}
\alias{.vcovhatbetahatbiased}
\title{Variance-Covariance Matrix of Estimates of Regression Coefficients (from \eqn{\hat{\sigma}_{\varepsilon \ \textrm{biased}}^{2}})}
\usage{
.vcovhatbetahatbiased(sigma2hatepsilonhatbiased = NULL, X, y)
}
\arguments{
\item{sigma2hatepsilonhatbiased}{Numeric.
Biased estimate of error variance.}

\item{X}{\code{n} by \code{k} numeric matrix.
The data matrix \eqn{\mathbf{X}}
(also known as design matrix, model matrix or regressor matrix)
is an \eqn{n \times k} matrix of \eqn{n} observations of \eqn{k} regressors,
which includes a regressor whose value is 1 for each observation on the first column.}

\item{y}{Numeric vector of length \code{n} or \code{n} by \code{1} matrix.
The vector \eqn{\mathbf{y}} is an \eqn{n \times 1} vector of observations
on the regressand variable.}
}
\description{
Variance-Covariance Matrix of Estimates of Regression Coefficients (from \eqn{\hat{\sigma}_{\varepsilon \ \textrm{biased}}^{2}})
}
\references{
\href{https://en.wikipedia.org/wiki/Linear_regression}{Wikipedia: Linear Regression}

\href{https://en.wikipedia.org/wiki/Ordinary_least_squares}{Wikipedia: Ordinary Least Squares}
}
\seealso{
Other variance-covariance of estimates of regression coefficients functions: 
\code{\link{.vcovhatbetahat}()},
\code{\link{vcovhatbetahatbiased}()},
\code{\link{vcovhatbetahat}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{variance-covariance of estimates of regression coefficients functions}
\keyword{inference}
