% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yhat.R
\name{.Xbetahat}
\alias{.Xbetahat}
\title{y-hat
\eqn{\left( \mathbf{\hat{y}} = \mathbf{X} \boldsymbol{\hat{\beta}} \right)}}
\usage{
.Xbetahat(X, betahat = NULL, y = NULL)
}
\arguments{
\item{X}{\code{n} by \code{k} numeric matrix.
The data matrix \eqn{\mathbf{X}}
(also known as design matrix, model matrix or regressor matrix)
is an \eqn{n \times k} matrix of \eqn{n} observations of \eqn{k} regressors,
which includes a regressor whose value is 1 for each observation on the first column.}

\item{betahat}{Numeric vector of length \code{k} or \code{k} by \code{1} matrix.
The vector \eqn{\boldsymbol{\hat{\beta}}} is a \eqn{k \times 1} vector of estimates
of \eqn{k} unknown regression coefficients.}

\item{y}{Numeric vector of length \code{n} or \code{n} by \code{1} matrix.
The vector \eqn{\mathbf{y}} is an \eqn{n \times 1} vector of observations
on the regressand variable.}
}
\value{
Returns y-hat \eqn{\left( \mathbf{\hat{y}} \right)}.
}
\description{
Calculates y-hat \eqn{\left( \mathbf{\hat{y}} \right)},
that is, the predicted value of \eqn{\mathbf{y}} given \eqn{\mathbf{X}} using
\deqn{
    \mathbf{\hat{y}} = \mathbf{X} \boldsymbol{\hat{\beta}}
  }
where
\deqn{
    \boldsymbol{\hat{\beta}} = \left( \mathbf{X}^{T} \mathbf{X} \right)^{-1}
    \left( \mathbf{X}^{T} \mathbf{y} \right) .
  }
}
\details{
If \code{betahat = NULL}, the \code{betahat} vector is computed
using \code{\link[=betahat]{betahat()}} with \code{X} and \code{y} as arguments.
If \code{betahat} is provided, \code{y} is not needed.
}
\references{
\href{https://en.wikipedia.org/wiki/Linear_regression}{Wikipedia: Linear Regression}

\href{https://en.wikipedia.org/wiki/Ordinary_least_squares}{Wikipedia: Ordinary Least Squares}
}
\seealso{
Other y-hat functions: 
\code{\link{.Py}()},
\code{\link{Py}()},
\code{\link{Xbetahat}()},
\code{\link{yhat}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{y-hat functions}
\keyword{predicted}
