% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e.R
\name{e}
\alias{e}
\title{Residuals (\eqn{\mathbf{e} = \mathbf{y} - \mathbf{\hat{y}}})}
\usage{
e(X, y, beta_hat = NULL)
}
\arguments{
\item{X}{Matrix.
The data matrix
\eqn{\mathbf{X}}
is an \eqn{n \times k} matrix
of \eqn{n} observations
of \eqn{k} regressors,
which includes a regressor
whose value is 1 for each observation.}

\item{y}{Vector or \code{n} by \code{1} matrix.
The vector
\eqn{\mathbf{y}}
is an \eqn{n \times 1} vector
of observations on the regressand variable.}

\item{beta_hat}{Vector or \code{k} by \code{1} matrix.
The vector
\eqn{\boldsymbol{\hat{\beta}}}
is a \eqn{k \times 1} vector of estimates
of \eqn{k} unknown regression coefficients.}
}
\value{
Returns residuals (\eqn{\mathbf{e}}),
that is,
the difference between
the observed and predicted value of \eqn{\mathbf{y}}
(\eqn{\mathbf{e} = \mathbf{y}} - \eqn{\mathbf{\hat{y}}}).
}
\description{
Calculates residuals using
\deqn{
    \mathbf{e}
    =
    \mathbf{y}
    -
    \mathbf{\hat{y}}
    =
    \mathbf{y}
    -
    \mathbf{X}
    \boldsymbol{\hat{\beta}}.
  }
}
\details{
If \code{beta_hat = NULL},
the \code{beta_hat} vector is computed
using \code{\link[=beta_hat_inv]{beta_hat_inv()}}.
}
\references{
\href{https://en.wikipedia.org/wiki/Errors_and_residuals}{Wikipedia: Errors and Residuals}
}
\seealso{
Other residuals functions: 
\code{\link{e_My}()},
\code{\link{e_y_minus_y_hat}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{residuals functions}
