% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epsilonhat.R
\name{epsilonhat}
\alias{epsilonhat}
\title{Residuals (\eqn{\boldsymbol{\hat{\varepsilon}} = \mathbf{y} - \mathbf{\hat{y}}})}
\usage{
epsilonhat(X, y, betahat = NULL)
}
\arguments{
\item{X}{Matrix.
The data matrix
\eqn{\mathbf{X}}
(also known as
design matrix,
model matrix or
regressor matrix)
is an \eqn{n \times k} matrix
of \eqn{n} observations
of \eqn{k} regressors,
which includes a regressor
whose value is 1 for each observation.}

\item{y}{Vector or \code{n} by \code{1} matrix.
The vector
\eqn{\mathbf{y}}
is an \eqn{n \times 1} vector
of observations on the regressand variable.}

\item{betahat}{Vector or \code{k} by \code{1} matrix.
The vector
\eqn{\boldsymbol{\hat{\beta}}}
is a \eqn{k \times 1} vector of estimates
of \eqn{k} unknown regression coefficients.}
}
\value{
Returns
an
\eqn{n \times 1}
matrix of
residuals
(\eqn{\boldsymbol{\hat{\varepsilon}}}),
that is,
the difference between
the observed
(\eqn{\mathbf{y}})
and predicted
(\eqn{\mathbf{\hat{y}}})
values of
the regressand variable
(\eqn{\boldsymbol{\hat{\varepsilon}} = \mathbf{y}} - \eqn{\mathbf{\hat{y}}}).
}
\description{
Calculates residuals using
\deqn{
    \hat{\varepsilon}_{i}
    =
    Y_{i}
    -
    \hat{Y}_{i} \\
    =
    Y_{i}
    -
    \left(
      \hat{\beta}_{1}
      +
      \hat{\beta}_{2}
      X_{2i}
      +
      \hat{\beta}_{3}
      X_{3i}
      +
      \dots
      +
      \hat{\beta}_{k}
      X_{ki}
    \right) \\
    =
    Y_{i}
    -
    \hat{\beta}_{1}
    -
    \hat{\beta}_{2}
    X_{2i}
    -
    \hat{\beta}_{3}
    X_{3i}
    -
    \dots
    -
    \hat{\beta}_{k}
    X_{ki} .
  }
In matrix form
\deqn{
    \boldsymbol{\hat{\varepsilon}}
    =
    \mathbf{y}
    -
    \mathbf{\hat{y}} \\
    =
    \mathbf{y}
    -
    \mathbf{X}
    \boldsymbol{\hat{\beta}}.
  }
}
\details{
If \code{betahat = NULL},
the \code{betahat} vector is computed
using \code{\link[=betahat_inv]{betahat_inv()}}.
}
\references{
\href{https://en.wikipedia.org/wiki/Errors_and_residuals}{Wikipedia: Errors and Residuals}
}
\seealso{
Other residuals functions: 
\code{\link{My}()},
\code{\link{y_minus_yhat}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{residuals functions}
