% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss.R
\name{ess}
\alias{ess}
\title{Explained Sum of Squares}
\usage{
ess(betahat = NULL, X, y)
}
\arguments{
\item{betahat}{Vector or \code{k} by \code{1} matrix.
The vector
\eqn{\boldsymbol{\hat{\beta}}}
is a \eqn{k \times 1} vector of estimates
of \eqn{k} unknown regression coefficients.}

\item{X}{Matrix.
The data matrix
\eqn{\mathbf{X}}
is an \eqn{n \times k} matrix
of \eqn{n} observations
of \eqn{k} regressors,
which includes a regressor
whose value is 1 for each observation.}

\item{y}{Vector or \code{n} by \code{1} matrix.
The vector
\eqn{\mathbf{y}}
is an \eqn{n \times 1} vector
of observations on the regressand variable.}
}
\value{
Returns the explained sum of squares.
}
\description{
Calculates the explained sum of squares (ESS)
\deqn{
    ESS
    =
    \boldsymbol{\hat{\beta}}^{\prime}
    \mathbf{X}^{\prime}
    \mathbf{X}
    \boldsymbol{\hat{\beta}}
    -
    n
    \mathbf{\bar{Y}}^{2}.
  }
Note that
\deqn{
    TSS
    =
    ESS
    +
    RSS.
  }
}
\references{
\href{https://en.wikipedia.org/wiki/Residual_sum_of_squares}{Wikipedia: Residual Sum of Squares}

\href{https://en.wikipedia.org/wiki/Explained_sum_of_squares}{Wikipedia: Explained Sum of Squares}

\href{https://en.wikipedia.org/wiki/Total_sum_of_squares}{Wikipedia: Total Sum of Squares}

\href{https://en.wikipedia.org/wiki/Coefficient_of_determination}{Wikipedia: Coefficient of Determination}
}
\seealso{
Other sum of squares functions: 
\code{\link{.rss}()},
\code{\link{rss}()},
\code{\link{tss}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{sum of squares functions}
