% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aov.R
\name{F_inc}
\alias{F_inc}
\title{F-Test
\eqn{
  F_{
    m,
    \left(
      n
      -
      k
    \right)
  }
}}
\usage{
F_inc(X, y, X_r)
}
\arguments{
\item{X}{Matrix.
The data matrix
\eqn{\mathbf{X}}
(also known as
design matrix,
model matrix or
regressor matrix)
is an \eqn{n \times k} matrix
of \eqn{n} observations
of \eqn{k} regressors,
which includes a regressor
whose value is 1 for each observation.}

\item{y}{Vector or \code{n} by \code{1} matrix.
The vector
\eqn{\mathbf{y}}
is an \eqn{n \times 1} vector
of observations on the regressand variable.}

\item{X_r}{Vector.
Vector of column indices used to subset X
for the restricted regression.}
}
\description{
Tests the incremental contribution of regressors
using
\deqn{
  F
  =
  \frac{
    \frac{
      RSS_R
      RSS_{
        UR
      }
    }
    {
      m
    }
  }
  {
    \frac{
      RSS_{
        UR
      }
    }
    {
      n
      -
      k
    }
  }
  \sim
  F_{
    m,
    \left(
      n
      -
      k
    \right)
  } .
}
}
\author{
Ivan Jacob Agaloos Pesigan
}
