% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta.R
\name{intercept}
\alias{intercept}
\title{Regression Intercept \eqn{\beta_{1}}}
\usage{
intercept(slopes, muy, muX)
}
\arguments{
\item{slopes}{Vector or \code{p} by \code{1} matrix.
Regression slopes \eqn{\left( \boldsymbol{\beta}_{2 \cdots k} \right)} .}

\item{muy}{Numeric.
Mean of the regressand variable \eqn{\left( \mu_{\mathbf{y}} \right)} .}

\item{muX}{Vector or \code{p} by \code{1} matrix. \eqn{p \times 1} vector
of means of the regressors \eqn{{X}_{2}, {X}_{3}, \cdots, {X}_{k}}
\eqn{\left( \boldsymbol{\mu}_{\mathbf{X}} \right)} .}
}
\value{
Returns the intercept \eqn{\beta_1}
of a linear regression model derived from the means
and the slopes \eqn{\left( \boldsymbol{\beta}_{2 \cdots k} \right)} .
}
\description{
Derives the intercept \eqn{\beta_1} of a linear regression model
from the \eqn{p \times 1} regression slopes \eqn{\left( \boldsymbol{\beta}_{2 \cdots k} \right)},
the mean of the regressand \eqn{\left( \mu_y \right)},
and the \eqn{p \times 1} means of regressors \eqn{{X}_{2}, {X}_{3}, \dots, {X}_{k}}
\eqn{\left( \boldsymbol{\mu}_{\mathbf{X}} \right)} .
}
\details{
The intercept is given by
\deqn{
    \beta_1
    =
    \mu_y
    -
    \boldsymbol{\mu}_{\mathbf{X}}
    \boldsymbol{\beta}_{2 \cdots k}^{T}
  }
}
\seealso{
Other parameter functions: 
\code{\link{sigma2epsilon}()},
\code{\link{slopesprime}()},
\code{\link{slopes}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{parameter functions}
\keyword{parameter}
