% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betahat_matrix.R
\name{intercepthat}
\alias{intercepthat}
\title{Estimated Regression Intercept \eqn{\hat{\beta}_{1}}}
\usage{
intercepthat(X, y)
}
\arguments{
\item{X}{\code{n} by \code{k} numeric matrix.
The data matrix \eqn{\mathbf{X}}
(also known as design matrix, model matrix or regressor matrix)
is an \eqn{n \times k} matrix of \eqn{n} observations of \eqn{k} regressors,
which includes a regressor whose value is 1 for each observation on the first column.}

\item{y}{Numeric vector of length \code{n} or \code{n} by \code{1} matrix.
The vector \eqn{\mathbf{y}} is an \eqn{n \times 1} vector of observations
on the regressand variable.}
}
\value{
Returns the estimated intercept \eqn{\hat{\beta}_1}
of a linear regression model derived from the estimated means
and the slopes \eqn{\left( \boldsymbol{\hat{\beta}}_{2, \cdots, k} \right)} .
}
\description{
Estimated Regression Intercept \eqn{\hat{\beta}_{1}}
}
\details{
The intercept \eqn{\beta_1} is given by
\deqn{
    \hat{\beta}_1 = \hat{\mu}_y - \boldsymbol{\hat{\mu}}_{\mathbf{X}}
    \boldsymbol{\hat{\beta}}_{2, \cdots, k}^{T} .
  }
}
\examples{
# Simple regression------------------------------------------------
X <- jeksterslabRdatarepo::wages.matrix[["X"]]
X <- X[, c(1, ncol(X))]
y <- jeksterslabRdatarepo::wages.matrix[["y"]]
intercepthat(X = X, y = y)

# Multiple regression----------------------------------------------
X <- jeksterslabRdatarepo::wages.matrix[["X"]]
# age is removed
X <- X[, -ncol(X)]
intercepthat(X = X, y = y)
}
\seealso{
Other beta-hat functions: 
\code{\link{.betahatnorm}()},
\code{\link{.betahatqr}()},
\code{\link{.betahatsvd}()},
\code{\link{.intercepthat}()},
\code{\link{.slopeshatprime}()},
\code{\link{.slopeshat}()},
\code{\link{betahat}()},
\code{\link{slopeshatprime}()},
\code{\link{slopeshat}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{beta-hat functions}
\keyword{beta-hat-ols}
