% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{linreg}
\alias{linreg}
\title{Linear Regression}
\usage{
linreg(X, y, beta_hat = "inv")
}
\arguments{
\item{X}{Matrix.
The data matrix
\eqn{\mathbf{X}}
is an \eqn{n \times k} matrix
of \eqn{n} observations
of \eqn{k} regressors,
which includes a regressor
whose value is 1 for each observation.}

\item{y}{Vector or \code{n} by \code{1} matrix.
The vector
\eqn{\mathbf{y}}
is an \eqn{n \times 1} vector
of observations on the regressand variable.}

\item{beta_hat}{String.
How the regression coefficients are estimated.
\code{beta_hat = "inv"} for \code{\link[=beta_hat_inv]{beta_hat_inv()}},
\code{beta_hat = "qr"} for \code{\link[=beta_hat_qr]{beta_hat_qr()}},
\code{beta_hat = "svd"} for \code{\link[=beta_hat_svd]{beta_hat_svd()}}.}
}
\description{
Linear Regression
}
\author{
Ivan Jacob Agaloos Pesigan
}
