% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proj.R
\name{M}
\alias{M}
\title{M Matrix}
\usage{
M(X, P = NULL)
}
\arguments{
\item{X}{Matrix.
The data matrix
\eqn{\mathbf{X}}
is an \eqn{n \times k} matrix
of \eqn{n} observations
of \eqn{k} regressors,
which includes a regressor
whose value is 1 for each observation.}

\item{P}{Matrix.
The projection matrix
(\eqn{\mathbf{P}}).}
}
\value{
Returns the residual maker matrix (\eqn{\mathbf{M}}).
}
\description{
Calculates the residual maker matrix (\eqn{\mathbf{M}})
using
\deqn{
    \mathbf{M}
    =
    \mathbf{I} - \mathbf{X}
    \left(
      \mathbf{X}^{\prime}
      \mathbf{X}
    \right)^{-1}
    \mathbf{X}^{\prime}
    =
    \mathbf{I} - \mathbf{P}.
  }
}
\details{
If \code{P = NULL},
the \code{P} matrix is computed
using \code{\link[=P]{P()}}.
The residual maker matrix (\eqn{\mathbf{M}})
transforms the \eqn{\mathbf{y}} vector
to the vector of residuals
(\eqn{\mathbf{e} = \mathbf{My}}).
}
\references{
\href{https://en.wikipedia.org/wiki/Projection_matrix}{Wikipedia: Projection Matrix}
}
\seealso{
Other projection matrix functions: 
\code{\link{P}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{projection matrix functions}
