% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_implied.R
\name{mean_str}
\alias{mean_str}
\title{Mean Structure
(\eqn{\mathbf{M}})}
\usage{
mean_str(beta_slopes, muy, muX)
}
\arguments{
\item{beta_slopes}{Vector or \code{p} by \code{1} matrix.
\eqn{\left( p \right) \times 1}
vector of regression slopes.}

\item{muy}{Numeric.
Mean of the regressand variable
\eqn{\mu_{\mathbf{y}}}.}

\item{muX}{Vector or \code{p} by \code{1} matrix.
\eqn{\left( p \right) \times 1} vector
of means of
the regressors
\eqn{
      {X}_{2},
      {X}_{3},
      \dots,
      {X}_{k}
    }
excluding the mean for the vector of constants
(\eqn{\boldsymbol{\mu}_{\mathbf{X}}}).}
}
\description{
Mean structure column vector
(\eqn{\mathbf{M}})
from regression slopes
and means of observed variables
of a \eqn{k}-variable Linear Regression Model.
}
\details{
The parameters of a linear regression model
for the mean structure
are
the slopes
(\eqn{\boldsymbol{\beta}_{\mathrm{slopes}}}),
the mean of
the regressand
\eqn{\mu_\mathbf{y}}
and
the means of
the regressors
\eqn{
    {X}_{2},
    {X}_{3},
    \dots,
    {X}_{k}
  }
excluding the mean for the vector of constants
(\eqn{\boldsymbol{\mu}_{\mathbf{X}}}).
}
\author{
Ivan Jacob Agaloos Pesigan
}
