% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{model}
\alias{model}
\title{Model Assessment}
\usage{
model(X, y)
}
\arguments{
\item{X}{\code{n} by \code{k} numeric matrix.
The data matrix \eqn{\mathbf{X}}
(also known as design matrix, model matrix or regressor matrix)
is an \eqn{n \times k} matrix of \eqn{n} observations of \eqn{k} regressors,
which includes a regressor whose value is 1 for each observation on the first column.}

\item{y}{Numeric vector of length \code{n} or \code{n} by \code{1} matrix.
The vector \eqn{\mathbf{y}} is an \eqn{n \times 1} vector of observations
on the regressand variable.}
}
\value{
Returns a vector with the following elements
\describe{
\item{RSS}{Residual sum of squares.}
\item{MSE}{Mean square error.}
\item{RMSE}{Root mean square error.}
\item{R2}{R-squared \eqn{\left( R^2 \right)}.}
\item{Rbar2}{Adjusted R-squared \eqn{\left( \bar{R}^2 \right)} .}
}
}
\description{
Model Assessment
}
\examples{
X <- jeksterslabRdatarepo::wages.matrix[["X"]]
# age is removed
X <- X[, -ncol(X)]
y <- jeksterslabRdatarepo::wages.matrix[["y"]]
model(
  X = X,
  y = y
)
}
\seealso{
Other assessment of model quality functions: 
\code{\link{.MSE}()},
\code{\link{.R2fromESS}()},
\code{\link{.R2fromRSS}()},
\code{\link{.RMSE}()},
\code{\link{.Rbar2}()},
\code{\link{.model}()},
\code{\link{MSE}()},
\code{\link{R2}()},
\code{\link{RMSE}()},
\code{\link{Rbar2}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{assessment of model quality functions}
\keyword{model-assessment}
