% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSE.R
\name{MSE}
\alias{MSE}
\title{Mean Squared Error}
\usage{
MSE(X, y, betahat = NULL)
}
\arguments{
\item{X}{Matrix.
The data matrix
\eqn{\mathbf{X}}
(also known as
design matrix,
model matrix or
regressor matrix)
is an \eqn{n \times k} matrix
of \eqn{n} observations
of \eqn{k} regressors,
which includes a regressor
whose value is 1 for each observation.}

\item{y}{Vector or \code{n} by \code{1} matrix.
The vector
\eqn{\mathbf{y}}
is an \eqn{n \times 1} vector
of observations on the regressand variable.}

\item{betahat}{Vector or \code{k} by \code{1} matrix.
The vector
\eqn{\boldsymbol{\hat{\beta}}}
is a \eqn{k \times 1} vector of estimates
of \eqn{k} unknown regression coefficients.}
}
\value{
Returns the mean squared error.
}
\description{
Calculates the mean squared error (\eqn{MSE}) using
\deqn{
    MSE
    =
    \frac{1}{n}
    \sum_{i = 1}^{n}
    \left(
      \mathbf{y} - \mathbf{X} \boldsymbol{\hat{\beta}}
    \right)^{2} \\
    =
    \frac{1}{n}
    \sum_{i = 1}^{n}
    \left(
      \mathbf{y} - \mathbf{\hat{y}}
    \right)^{2} \\
    =
    \frac{RSS}{n}.
  }
}
\references{
\href{https://en.wikipedia.org/wiki/Mean_squared_error}{Wikipedia: Mean Squared Error}
}
\seealso{
Other assessment of model quality functions: 
\code{\link{.MSE}()},
\code{\link{.R2_ESS}()},
\code{\link{.R2_RSS}()},
\code{\link{.RMSE}()},
\code{\link{.Rbar2}()},
\code{\link{R2}()},
\code{\link{RMSE}()},
\code{\link{Rbar2}()}
}
\concept{assessment of model quality functions}
