% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ram.R
\name{mutheta}
\alias{mutheta}
\title{Model-Implied Mean Vector
\eqn{\boldsymbol{\mu} \left( \boldsymbol{\theta} \right)}}
\usage{
mutheta(beta, muX)
}
\arguments{
\item{beta}{Vector or \code{k} by \code{1} matrix.
\eqn{k \times 1} regression coefficients \eqn{\left( \boldsymbol{\beta} \right)}.}

\item{muX}{Vector or \code{p} by \code{1} matrix. \eqn{p \times 1} vector
of means of the regressors \eqn{ {X}_{2}, {X}_{3}, \dots, {X}_{k}}
\eqn{\left( \boldsymbol{\mu}_{\mathbf{X}} \right)} .}
}
\description{
Model-implied mean vector
\eqn{\boldsymbol{\mu} \left( \boldsymbol{\theta} \right)}
from parameters of a \eqn{k}-variable linear regression model.
}
\details{
The following are the parameters of a linear regression model
for the mean structure
\itemize{
\item \eqn{\boldsymbol{\beta}} - \eqn{k \times k} regression coefficients and
\item \eqn{\boldsymbol{\mu}_{\mathbf{X}}} - \eqn{p \times 1} vector of means of the regressors
\eqn{ {X}_{2}, {X}_{3}, \dots, {X}_{k}} .
}

The model-implied mean vector is given by
\deqn{
    \boldsymbol{\mu} \left( \boldsymbol{\theta} \right)
    =
    \left( \mathrm{I} - \mathrm{A} \right)^{-1}
    \mathrm{M}
  }
where
\deqn{
    \mathbf{A}
    =
    \begin{bmatrix}
      0 & \beta_2 & \beta_3 & \cdots & \beta_k \\
      0 & 0       & 0       & \cdots & 0       \\
      0 & 0       & 0       & \cdots & 0       \\
      0 & 0       & 0       & \cdots & 0       \\
      0 & 0       & 0       & \cdots & 0
    \end{bmatrix}
  }

\deqn{
    \mathbf{M}
    =
    \begin{bmatrix}
      \beta_1     \\
      \mu_{X_{2}} \\
      \mu_{X_{3}} \\
      \vdots      \\
      \mu_{X_{k}}
    \end{bmatrix}
  }
}
\seealso{
Other model-implied functions: 
\code{\link{Sigmatheta}()},
\code{\link{intercept}()},
\code{\link{slopes}()},
\code{\link{stdslopes}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{model-implied functions}
\keyword{model-implied}
