% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proj.R
\name{P}
\alias{P}
\title{P Matrix}
\usage{
P(X)
}
\arguments{
\item{X}{Matrix.
The data matrix
\eqn{\mathbf{X}}
(also known as
design matrix,
model matrix or
regressor matrix)
is an \eqn{n \times k} matrix
of \eqn{n} observations
of \eqn{k} regressors,
which includes a regressor
whose value is 1 for each observation.}
}
\value{
Returns the projection matrix (\eqn{\mathbf{P}}).
}
\description{
Calculates the projection matrix
\eqn{\left( \mathbf{P} \right)}
using
\deqn{
  \mathbf{P}
  =
  \mathbf{X}
  \left(
    \mathbf{X}^{\prime}
    \mathbf{X}
  \right)^{-1}
  \mathbf{X}^{\prime}.
}
}
\details{
The projection matrix
\eqn{\left( \mathbf{P} \right)},
also known as the hat matrix,
transforms the \eqn{\mathbf{y}} vector
to the vector of predicted values
\eqn{\left( \mathbf{\hat{y}} = \mathbf{Py} \right)}.
}
\references{
\href{https://en.wikipedia.org/wiki/Projection_matrix}{Wikipedia: Projection Matrix}
}
\seealso{
Other projection matrix functions: 
\code{\link{.h}()},
\code{\link{M}()},
\code{\link{h}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{projection matrix functions}
