% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yhat.R
\name{Py}
\alias{Py}
\title{y-hat (\eqn{\mathbf{\hat{y}} = \mathbf{P} \mathbf{y}})}
\usage{
Py(y, P = NULL, X = NULL)
}
\arguments{
\item{y}{Vector or \code{n} by \code{1} matrix.
The vector
\eqn{\mathbf{y}}
is an \eqn{n \times 1} vector
of observations on the regressand variable.}

\item{P}{Matrix.
The projection matrix
(\eqn{\mathbf{P}}).}

\item{X}{Matrix.
The data matrix
\eqn{\mathbf{X}}
is an \eqn{n \times k} matrix
of \eqn{n} observations
of \eqn{k} regressors,
which includes a regressor
whose value is 1 for each observation.}
}
\value{
Returns y-hat (\eqn{\mathbf{\hat{y}}}).
}
\description{
Calculates y-hat (\eqn{\mathbf{\hat{y}}}), that is,
the predicted value of \eqn{\mathbf{y}}
given \eqn{\mathbf{X}} using
\deqn{
    \mathbf{\hat{y}}
    =
    \mathbf{P}
    \mathbf{y}.
  }
}
\details{
If \code{P = NULL},
the \code{P} matrix is computed
using \code{\link[=P]{P()}}
with \code{X} as its argument.
If \code{P} is provided,
\code{X} is not needed.
}
\references{
\href{https://en.wikipedia.org/wiki/Linear_regression}{Wikipedia: Linear Regression}

\href{https://en.wikipedia.org/wiki/Ordinary_least_squares}{Wikipedia: Ordinary Least Squares}
}
\seealso{
Other y-hat functions: 
\code{\link{Xbetahat}()},
\code{\link{yhat_test}()},
\code{\link{yhat}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{y-hat functions}
