% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2.R
\name{r2_ess}
\alias{r2_ess}
\title{R-square (from \eqn{ESS})}
\usage{
r2_ess(ess = NULL, tss = NULL, beta_hat = NULL, X = NULL, y = NULL)
}
\arguments{
\item{ess}{Numeric.
Explained sum of squares.}

\item{tss}{Numeric.
Total sum of squares.}

\item{beta_hat}{Vector or \code{k} by \code{1} matrix.
The vector
\eqn{\boldsymbol{\hat{\beta}}}
is a \eqn{k \times 1} vector of estimates
of \eqn{k} unknown regression coefficients.}

\item{X}{Matrix.
The data matrix
\eqn{\mathbf{X}}
is an \eqn{n \times k} matrix
of \eqn{n} observations
of \eqn{k} regressors,
which includes a regressor
whose value is 1 for each observation.}

\item{y}{Vector or \code{n} by \code{1} matrix.
The vector
\eqn{\mathbf{y}}
is an \eqn{n \times 1} vector
of observations on the regressand variable.}
}
\value{
Returns the coefficient of determination \eqn{R^2}.
}
\description{
Calculates the coefficient of determination using
\deqn{
    R^2
    =
    \frac{\textrm{Explained sum of squares}}{\textrm{Total sum of squares}}.
  }
}
\details{
If \code{ess = NULL},
\code{ess} is computed
using \code{\link[=ss_e]{ss_e()}}
with \code{X} and \code{y} as required arguments
and \code{beta_hat} as an optional argument.
If \code{tss = NULL},
\code{tss} is computed
using \code{\link[=ss_t]{ss_t()}}
with \code{y} as a required argument.
If \code{ess} and \code{tss} are provided,
\code{beta_hat}, \code{X}, and \code{y}
are not needed.
}
\references{
\href{https://en.wikipedia.org/wiki/Residual_sum_of_squares}{Wikipedia: Residual Sum of Squares}

\href{https://en.wikipedia.org/wiki/Explained_sum_of_squares}{Wikipedia: Explained Sum of Squares}

\href{https://en.wikipedia.org/wiki/Total_sum_of_squares}{Wikipedia: Total Sum of Squares}

\href{https://en.wikipedia.org/wiki/Coefficient_of_determination}{Wikipedia: Coefficient of Determination}
}
\seealso{
Other assessment of model quality functions: 
\code{\link{mse_rss}()},
\code{\link{mse}()},
\code{\link{r2_rss}()},
\code{\link{r2}()},
\code{\link{rbar2_r2}()},
\code{\link{rbar2}()},
\code{\link{rmse_rss}()},
\code{\link{rmse}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{assessment of model quality functions}
