% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2.R
\name{r2_rss}
\alias{r2_rss}
\title{R-square (from \eqn{RSS})}
\usage{
r2_rss(rss = NULL, tss = NULL, beta_hat = NULL, X = NULL, y = NULL)
}
\arguments{
\item{rss}{Numeric.
Residual sum of squares.}

\item{tss}{Numeric.
Total sum of squares.}

\item{beta_hat}{Vector or \code{k} by \code{1} matrix.
The vector
\eqn{\boldsymbol{\hat{\beta}}}
is a \eqn{k \times 1} vector of estimates
of \eqn{k} unknown regression coefficients.}

\item{X}{Matrix.
The data matrix
\eqn{\mathbf{X}}
is an \eqn{n \times k} matrix
of \eqn{n} observations
of \eqn{k} regressors,
which includes a regressor
whose value is 1 for each observation.}

\item{y}{Vector or \code{n} by \code{1} matrix.
The vector
\eqn{\mathbf{y}}
is an \eqn{n \times 1} vector
of observations on the regressand variable.}
}
\value{
Returns the coefficient of determination \eqn{R^2}.
}
\description{
Calculates the coefficient of determination using
\deqn{
    R^2
    =
    1 - \frac{\textrm{Residual sum of squares}}{\textrm{Total sum of squares}}
  }
}
\details{
If \code{rss = NULL},
\code{rss} is computed
using \code{\link[=ss_r]{ss_r()}}
with \code{X} and \code{y} as required arguments
and \code{beta_hat} as an optional argument.
If \code{tss = NULL},
\code{tss} is computed
using \code{\link[=ss_t]{ss_t()}}
with \code{y} as a required argument.
If \code{rss} and \code{tss} are provided,
\code{beta_hat}, \code{X}, and \code{y}
are not needed.
}
\references{
\href{https://en.wikipedia.org/wiki/Residual_sum_of_squares}{Wikipedia: Residual Sum of Squares}

\href{https://en.wikipedia.org/wiki/Explained_sum_of_squares}{Wikipedia: Explained Sum of Squares}

\href{https://en.wikipedia.org/wiki/Total_sum_of_squares}{Wikipedia: Total Sum of Squares}

\href{https://en.wikipedia.org/wiki/Coefficient_of_determination}{Wikipedia: Coefficient of Determination}
}
\seealso{
Other assessment of model quality functions: 
\code{\link{mse}()},
\code{\link{r2_ess}()},
\code{\link{r2}()},
\code{\link{rbar2_r2}()},
\code{\link{rbar2}()},
\code{\link{rmse}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{assessment of model quality functions}
