% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss.R
\name{r_bar_sqr}
\alias{r_bar_sqr}
\title{Adjusted R-square}
\usage{
r_bar_sqr(r2 = NULL, n, k, ...)
}
\arguments{
\item{r2}{Coefficient of determination \eqn{R^2}.}

\item{n}{Integer.
Sample size.}

\item{k}{Integer.
Number of regressors
including a regressor
whose value is 1 for each observation.}

\item{...}{Arguments to pass to \code{\link[=r_sqr]{r_sqr()}} if \code{r2 = NULL}.}
}
\value{
Returns the adjusted coefficient of determination \eqn{\bar{R}^{2}}.
}
\description{
Calculates the adjusted coefficient of determination
\deqn{
    \bar{R}^{2}
    =
    1
    -
    \left(
      \frac{RSS / \left( n - k \right)}{TSS / \left(n - 1 \right)}
    \right)
    =
    1
    -
    \left(
      1 - R^2
    \right)
    \frac{n - 1}{n - k}.
  }
}
\references{
\href{https://en.wikipedia.org/wiki/Residual_sum_of_squares}{Wikipedia: Residual Sum of Squares}

\href{https://en.wikipedia.org/wiki/Explained_sum_of_squares}{Wikipedia: Explained Sum of Squares}

\href{https://en.wikipedia.org/wiki/Total_sum_of_squares}{Wikipedia: Total Sum of Squares}

\href{https://en.wikipedia.org/wiki/Coefficient_of_determination}{Wikipedia: Coefficient of Determination}
}
\seealso{
Other assessment of model quality functions: 
\code{\link{mse}()},
\code{\link{r_sqr}()},
\code{\link{rmse}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{assessment of model quality functions}
