% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss.R
\name{r_sqr}
\alias{r_sqr}
\title{R-square}
\usage{
r_sqr(beta_hat = NULL, X, y, rss = TRUE)
}
\arguments{
\item{beta_hat}{Vector or \code{k} by \code{1} matrix.
The vector
\eqn{\boldsymbol{\hat{\beta}}}
is a \eqn{k \times 1} vector of estimates
of \eqn{k} unknown regression coefficients.}

\item{X}{Matrix.
The data matrix
\eqn{\mathbf{X}}
is an \eqn{n \times k} matrix
of \eqn{n} observations
of \eqn{k} regressors,
which includes a regressor
whose value is 1 for each observation.}

\item{y}{Vector or \code{n} by \code{1} matrix.
The vector
\eqn{\mathbf{y}}
is an \eqn{n \times 1} vector
of observations on the regressand variable.}

\item{rss}{Logical.
If \code{TRUE}, the function uses the residual sum of squares in the calculation.
If \code{FALSE}, the function uses the estimated sum of squares in the calculation.}
}
\value{
Returns the coefficient of determination \eqn{R^2}.
}
\description{
Calculates the coefficient of determination
\deqn{
    R^2
    =
    1 - \frac{\textrm{Residual sum of squares}}{\textrm{Total sum of squares}}
  }
or
\deqn{R^2 = \frac{\textrm{Explained sum of squares}}{\textrm{Total sum of squares}}.
}
}
\references{
\href{https://en.wikipedia.org/wiki/Residual_sum_of_squares}{Wikipedia: Residual Sum of Squares}

\href{https://en.wikipedia.org/wiki/Explained_sum_of_squares}{Wikipedia: Explained Sum of Squares}

\href{https://en.wikipedia.org/wiki/Total_sum_of_squares}{Wikipedia: Total Sum of Squares}

\href{https://en.wikipedia.org/wiki/Coefficient_of_determination}{Wikipedia: Coefficient of Determination}
}
\seealso{
Other assessment of model quality functions: 
\code{\link{mse}()},
\code{\link{r_bar_sqr}()},
\code{\link{rmse}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{assessment of model quality functions}
