% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2.R
\name{rbar2_r2}
\alias{rbar2_r2}
\title{Adjusted R-square (from \eqn{R^2})}
\usage{
rbar2_r2(r2 = NULL, n, k, beta_hat = NULL, X, y, fromrss = TRUE)
}
\arguments{
\item{r2}{Numeric.
Coefficient of determinims \eqn{R^2}.}

\item{n}{Integer.
Sample size.}

\item{k}{Integer.
Number of regressors
including a regressor
whose value is 1 for each observation.}

\item{beta_hat}{Vector or \code{k} by \code{1} matrix.
The vector
\eqn{\boldsymbol{\hat{\beta}}}
is a \eqn{k \times 1} vector of estimates
of \eqn{k} unknown regression coefficients.}

\item{X}{Matrix.
The data matrix
\eqn{\mathbf{X}}
is an \eqn{n \times k} matrix
of \eqn{n} observations
of \eqn{k} regressors,
which includes a regressor
whose value is 1 for each observation.}

\item{y}{Vector or \code{n} by \code{1} matrix.
The vector
\eqn{\mathbf{y}}
is an \eqn{n \times 1} vector
of observations on the regressand variable.}

\item{fromrss}{Logical.
If \code{TRUE},
calculates the coefficient of determinism
from \code{RSS}.
If \code{FALSE},
calculates the coefficient of determinism
from \code{ESS}.}
}
\value{
Returns the adjusted coefficient of determination \eqn{\bar{R}^{2}}.
}
\description{
Calculates the adjusted coefficient of determination
\deqn{
    \bar{R}^{2}
    =
    1
    -
    \left(
      \frac{RSS / \left( n - k \right)}{TSS / \left(n - 1 \right)}
    \right)
    =
    1
    -
    \left(
      1 - R^2
    \right)
    \frac{n - 1}{n - k}.
  }
}
\details{
If \code{r2 = NULL},
\code{r2} is computed
using \code{\link[=r2]{r2()}}
with \code{X} and \code{y} as required arguments
and \code{beta_hat} as an optional argument.
If \code{r2} is provided,
\code{beta_hat}, \code{X}, and \code{y}
are not needed.
}
\references{
\href{https://en.wikipedia.org/wiki/Residual_sum_of_squares}{Wikipedia: Residual Sum of Squares}

\href{https://en.wikipedia.org/wiki/Explained_sum_of_squares}{Wikipedia: Explained Sum of Squares}

\href{https://en.wikipedia.org/wiki/Total_sum_of_squares}{Wikipedia: Total Sum of Squares}

\href{https://en.wikipedia.org/wiki/Coefficient_of_determination}{Wikipedia: Coefficient of Determination}
}
\seealso{
Other assessment of model quality functions: 
\code{\link{mse_rss}()},
\code{\link{mse}()},
\code{\link{r2_ess}()},
\code{\link{r2_rss}()},
\code{\link{r2}()},
\code{\link{rbar2}()},
\code{\link{rmse_rss}()},
\code{\link{rmse}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{assessment of model quality functions}
