% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residual.plot.R
\name{residual.plot}
\alias{residual.plot}
\title{Residual Plots}
\usage{
residual.plot(X, y)
}
\arguments{
\item{X}{\code{n} by \code{k} numeric matrix.
The data matrix \eqn{\mathbf{X}}
(also known as design matrix, model matrix or regressor matrix)
is an \eqn{n \times k} matrix of \eqn{n} observations of \eqn{k} regressors,
which includes a regressor whose value is 1 for each observation on the first column.}

\item{y}{Numeric vector of length \code{n} or \code{n} by \code{1} matrix.
The vector \eqn{\mathbf{y}} is an \eqn{n \times 1} vector of observations
on the regressand variable.}
}
\description{
Residual Plots
}
\examples{
# Simple regression------------------------------------------------
X <- jeksterslabRdatarepo::wages.matrix[["X"]]
X <- X[, c(1, ncol(X))]
y <- jeksterslabRdatarepo::wages.matrix[["y"]]
residual.plot(X = X, y = y)

# Multiple regression----------------------------------------------
X <- jeksterslabRdatarepo::wages.matrix[["X"]]
# age is removed
X <- X[, -ncol(X)]
residual.plot(X = X, y = y)
}
\seealso{
Other plotting functions: 
\code{\link{scatter.plot}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{plotting functions}
\keyword{plot}
