% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mse.R
\name{rmse_rss}
\alias{rmse_rss}
\title{Root Mean Square Error (from \eqn{RSS})}
\usage{
rmse_rss(rss = NULL, n, beta_hat = NULL, X, y)
}
\arguments{
\item{rss}{Numeric.
Residual sum of squares.}

\item{n}{Integer.
Sample size.}

\item{beta_hat}{Vector or \code{k} by \code{1} matrix.
The vector
\eqn{\boldsymbol{\hat{\beta}}}
is a \eqn{k \times 1} vector of estimates
of \eqn{k} unknown regression coefficients.}

\item{X}{Matrix.
The data matrix
\eqn{\mathbf{X}}
is an \eqn{n \times k} matrix
of \eqn{n} observations
of \eqn{k} regressors,
which includes a regressor
whose value is 1 for each observation.}

\item{y}{Vector or \code{n} by \code{1} matrix.
The vector
\eqn{\mathbf{y}}
is an \eqn{n \times 1} vector
of observations on the regressand variable.}
}
\value{
Returns the root mean square error.
}
\description{
Calculates the root mean square error (RMSE) using
\deqn{
    RMSE
    =
    \sqrt{
      \frac{1}{n}
      \sum_{i = 1}^{n}
      \left(
        Y_i - \hat{Y}_{i}
      \right)^{2}
    }
    =
    \sqrt{\frac{RSS}{n}}.
  }
}
\details{
If \code{rss = NULL},
the \code{rss} vector is computed
using \code{\link[=ss_r]{ss_r()}}
with \code{X} and \code{y} as required arguments
and \code{beta_hat} as an optional argument.
If \code{rss} is provided,
\code{beta_hat}, \code{X}, and \code{y}
are not needed.
}
\references{
\href{https://en.wikipedia.org/wiki/Root-mean-square_deviation}{Wikipedia: Root Mean Square Deviation}
}
\seealso{
Other assessment of model quality functions: 
\code{\link{mse_rss}()},
\code{\link{mse}()},
\code{\link{r2_ess}()},
\code{\link{r2_rss}()},
\code{\link{r2}()},
\code{\link{rbar2_r2}()},
\code{\link{rbar2}()},
\code{\link{rmse}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{assessment of model quality functions}
