% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss.R
\name{rss}
\alias{rss}
\title{Residual Sum of Square}
\usage{
rss(betahat = NULL, X, y)
}
\arguments{
\item{betahat}{Vector or \code{k} by \code{1} matrix.
The vector
\eqn{\boldsymbol{\hat{\beta}}}
is a \eqn{k \times 1} vector of estimates
of \eqn{k} unknown regression coefficients.}

\item{X}{Matrix.
The data matrix
\eqn{\mathbf{X}}
is an \eqn{n \times k} matrix
of \eqn{n} observations
of \eqn{k} regressors,
which includes a regressor
whose value is 1 for each observation.}

\item{y}{Vector or \code{n} by \code{1} matrix.
The vector
\eqn{\mathbf{y}}
is an \eqn{n \times 1} vector
of observations on the regressand variable.}
}
\value{
Returns residual sum of squares.
}
\description{
Calculates the residual sum of squares (RSS) using
\deqn{
    RSS
    =
    \sum_{i = 1}^{n}
    \left(
      \mathbf{y}
      -
      \mathbf{X}
      \boldsymbol{\hat{\beta}}
    \right)^{2}
    =
    \sum_{i = 1}^{n} e_{i}^{2}
    =
    \mathbf{e}^{\prime}
    \mathbf{e}.
  }
Equivalent formulas
\deqn{
    RSS
    =
    \left(
      \mathbf{y}
      -
      \mathbf{X}
      \boldsymbol{\hat{\beta}}
    \right)^{\prime}
    \left(
      \mathbf{y}
      -
      \mathbf{X}
      \boldsymbol{\hat{\beta}}
    \right)
    =
    \mathbf{y}^{\prime}
    \mathbf{y}
    -
    2
    \boldsymbol{\hat{\beta}}
    \mathbf{X}^{\prime}
    \mathbf{y}
    +
    \boldsymbol{\hat{\beta}}^{\prime}
    \mathbf{X}^{\prime}
    \mathbf{X}
    \boldsymbol{\hat{\beta}}.
  }
Note that
\deqn{
    TSS
    =
    ESS
    +
    RSS.
  }
}
\details{
If \code{betahat = NULL},
the \code{betahat} vector is computed
using \code{\link[=betahat_inv]{betahat_inv()}}.
}
\references{
\href{https://en.wikipedia.org/wiki/Residual_sum_of_squares}{Wikipedia: Residual Sum of Squares}

\href{https://en.wikipedia.org/wiki/Explained_sum_of_squares}{Wikipedia: Explained Sum of Squares}

\href{https://en.wikipedia.org/wiki/Total_sum_of_squares}{Wikipedia: Total Sum of Squares}

\href{https://en.wikipedia.org/wiki/Coefficient_of_determination}{Wikipedia: Coefficient of Determination}
}
\seealso{
Other sum of squares functions: 
\code{\link{.rss}()},
\code{\link{ess}()},
\code{\link{tss}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{sum of squares functions}
