% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcov.R
\name{s_sqr}
\alias{s_sqr}
\title{Residual Variance}
\usage{
s_sqr(beta_hat = NULL, X, y, s_sqr_est = "both")
}
\arguments{
\item{beta_hat}{Vector or \code{k} by \code{1} matrix.
The vector
\eqn{\boldsymbol{\hat{\beta}}}
is a \eqn{k \times 1} vector of estimates
of \eqn{k} unknown regression coefficients.}

\item{X}{Matrix.
The data matrix
\eqn{\mathbf{X}}
is an \eqn{n \times k} matrix
of \eqn{n} observations
of \eqn{k} regressors,
which includes a regressor
whose value is 1 for each observation.}

\item{y}{Vector or \code{n} by \code{1} matrix.
The vector
\eqn{\mathbf{y}}
is an \eqn{n \times 1} vector
of observations on the regressand variable.}

\item{s_sqr_est}{String.
Residual variance estimator.
If \code{"both"},
returns both OLS and ML estimates as a vector.
If \code{"ols"},
returns OLS estimate.
If \code{"ml"},
returns ML estimate.}
}
\value{
Returns the estimated residual variance.
}
\description{
Calculates estimates of the residual variance
\deqn{
    \mathbf{E}
      \left(
        \sigma^2
      \right)
  =
  s^2
  },
\deqn{
    s_{\textrm{OLS}}^{2}
    =
    \frac{
      \mathbf{e^{\prime} e }
    }
    {
      n - k
    }
  },
\deqn{
    s_{\textrm{ML}}^{2}
    =
    \frac{
      \mathbf{e^{\prime} e }
    }
    {
      n
    }
  }.
}
\author{
Ivan Jacob Agaloos Pesigan
}
