% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sehat.R
\name{sehatbetahatprimebiased}
\alias{sehatbetahatprimebiased}
\title{Standard Errors of Standardized Estimates of Regression Coefficients}
\usage{
sehatbetahatprimebiased(X, y)
}
\arguments{
\item{X}{\code{n} by \code{k} numeric matrix.
The data matrix \eqn{\mathbf{X}}
(also known as design matrix, model matrix or regressor matrix)
is an \eqn{n \times k} matrix of \eqn{n} observations of \eqn{k} regressors,
which includes a regressor whose value is 1 for each observation on the first column.}

\item{y}{Numeric vector of length \code{n} or \code{n} by \code{1} matrix.
The vector \eqn{\mathbf{y}} is an \eqn{n \times 1} vector of observations
on the regressand variable.}
}
\description{
\deqn{
    \mathrm{\widehat{se}}_{\hat{\beta}^{\prime}} =
    \mathrm{\widehat{se}}_{\hat{\beta}} \frac{\hat{\beta}^{\prime}}{\hat{\beta}}
  }
According to Yuan and Chan (2011), this standard error is biased.
}
\examples{
X <- jeksterslabRdatarepo::wages.matrix[["X"]]
# age is removed
X <- X[, -ncol(X)]
y <- jeksterslabRdatarepo::wages.matrix[["y"]]
sehatbetahatprimebiased(
  X = X,
  y = y
)
}
\seealso{
Other standard errors of estimates of regression coefficients functions: 
\code{\link{.sehatbetahatbiased}()},
\code{\link{.sehatbetahatprimebiased}()},
\code{\link{.sehatbetahat}()},
\code{\link{sehatbetahatbiased}()},
\code{\link{sehatbetahat}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{standard errors of estimates of regression coefficients functions}
\keyword{inference}
