% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sehat.R
\name{sehatslopeshatprimedelta}
\alias{sehatslopeshatprimedelta}
\title{Standard Errors of Standardized Estimates of Regression Coefficients (Yuan and Chan (2011))}
\usage{
sehatslopeshatprimedelta(X, y, adjust = FALSE)
}
\arguments{
\item{X}{\code{n} by \code{k} numeric matrix.
The data matrix \eqn{\mathbf{X}}
(also known as design matrix, model matrix or regressor matrix)
is an \eqn{n \times k} matrix of \eqn{n} observations of \eqn{k} regressors,
which includes a regressor whose value is 1 for each observation on the first column.}

\item{y}{Numeric vector of length \code{n} or \code{n} by \code{1} matrix.
The vector \eqn{\mathbf{y}} is an \eqn{n \times 1} vector of observations
on the regressand variable.}

\item{adjust}{Logical.
Use \eqn{n - 3} adjustment for small samples.}
}
\description{
Standard Errors of Standardized Estimates of Regression Coefficients (Yuan and Chan (2011))
}
\details{
The \eqn{p}th estimated standard error is calculated using
\deqn{
    \mathbf{\widehat{se}}_{\boldsymbol{\hat{\beta}}_{\text{p}}^{\prime}} =
    \sqrt{
      \frac{\hat{\sigma}_{X_{p}}^{2} \hat{c}_{p} \hat{\sigma}_{\hat{\varepsilon}}^{2}}{n \hat{\sigma}_{y}^{2}} +
    \frac{\hat{\beta}_{p}
      \left[
        \hat{\sigma}_{X_{p}}^{2}
        \left( \boldsymbol{\hat{\beta}}^{T} \boldsymbol{\hat{\Sigma}}_{X} \boldsymbol{\hat{\beta}} \right) -
        \hat{\sigma}_{X_{p}}^{2} \hat{\sigma}_{\hat{\varepsilon}}^{2} -
        \hat{\sigma}_{y, X_{p}}^{2}
      \right]}{n \hat{\sigma}_{y}^{4}}
    }
  }
where
\itemize{
\item \eqn{p = \left\{2, 3, \cdots, k \right\}}
\item \eqn{\hat{\sigma}_{\hat{\varepsilon}}^{2}} is the estimated residual variance
\item \eqn{\boldsymbol{\hat{\beta}}_{2, 3, \cdots, k} = \left\{ \hat{\beta}_{2}, \hat{\beta}_{3}, \cdots, \hat{\beta}_{k}\right\}^{T}}
is the \eqn{p \times 1} column vector of estimated regression slopes
\item \eqn{\hat{\sigma}_{y}^{2}} is the variance of the regressand variable \eqn{y}
\item \eqn{\boldsymbol{\hat{\Sigma}}_{\mathbf{X}}} is the \eqn{p \times p} estimated covariance matrix of the regressor variables \eqn{X_2, X_3, \cdots, X_k}
\item \eqn{\hat{\sigma}_{X_p}^{2}} is the variance of the corresponding \eqn{p}th regressor variable.
\item \eqn{\hat{\sigma}_{y, X_{p}}^{2}} is the covariance of the regressand variable \eqn{y} and the regressor variables \eqn{X_2, X_3, \cdots, X_k}
\item \eqn{c_p} is the diagonal element that corresponds to the regressor variable in \eqn{\boldsymbol{\Sigma}_{\mathbf{X}}^{-1}}
\item \eqn{n} is the sample size
}
}
\examples{
# Simple regression------------------------------------------------
X <- jeksterslabRdatarepo::wages.matrix[["X"]]
X <- X[, c(1, ncol(X))]
y <- jeksterslabRdatarepo::wages.matrix[["y"]]
sehatslopeshatprimedelta(X = X, y = y)

# Multiple regression----------------------------------------------
X <- jeksterslabRdatarepo::wages.matrix[["X"]]
# age is removed
X <- X[, -ncol(X)]
sehatslopeshatprimedelta(X = X, y = y)
}
\references{
Yuan, K., Chan, W. (2011).
Biases and Standard Errors of Standardized Regression Coefficients.
\emph{Psychometrika} 76, 670-690.
\href{https://doi.org/10.1007/s11336-011-9224-6}{doi:10.1007/s11336-011-9224-6}.
}
\seealso{
Other standard errors of estimates of regression coefficients functions: 
\code{\link{.sehatbetahatbiased}()},
\code{\link{.sehatbetahat}()},
\code{\link{.sehatslopeshatprimedelta}()},
\code{\link{.sehatslopeshatprimetb}()},
\code{\link{sehatbetahatbiased}()},
\code{\link{sehatbetahat}()},
\code{\link{sehatslopeshatprimetb}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{standard errors of estimates of regression coefficients functions}
\keyword{se}
