% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sehat.R
\name{sehatslopeshatprimetb}
\alias{sehatslopeshatprimetb}
\title{Standard Errors of Standardized Estimates of Regression Coefficients (Textbook)}
\usage{
sehatslopeshatprimetb(X, y)
}
\arguments{
\item{X}{\code{n} by \code{k} numeric matrix.
The data matrix \eqn{\mathbf{X}}
(also known as design matrix, model matrix or regressor matrix)
is an \eqn{n \times k} matrix of \eqn{n} observations of \eqn{k} regressors,
which includes a regressor whose value is 1 for each observation on the first column.}

\item{y}{Numeric vector of length \code{n} or \code{n} by \code{1} matrix.
The vector \eqn{\mathbf{y}} is an \eqn{n \times 1} vector of observations
on the regressand variable.}
}
\description{
Standard Errors of Standardized Estimates of Regression Coefficients (Textbook)
}
\examples{
# Simple regression------------------------------------------------
X <- jeksterslabRdatarepo::wages.matrix[["X"]]
X <- X[, c(1, ncol(X))]
y <- jeksterslabRdatarepo::wages.matrix[["y"]]
sehatslopeshatprimetb(X = X, y = y)

# Multiple regression----------------------------------------------
X <- jeksterslabRdatarepo::wages.matrix[["X"]]
# age is removed
X <- X[, -ncol(X)]
sehatslopeshatprimetb(X = X, y = y)
}
\references{
Yuan, K., Chan, W. (2011).
Biases and Standard Errors of Standardized Regression Coefficients.
\emph{Psychometrika} 76, 670-690.
\href{https://doi.org/10.1007/s11336-011-9224-6}{doi:10.1007/s11336-011-9224-6}.
}
\seealso{
Other standard errors of estimates of regression coefficients functions: 
\code{\link{.sehatbetahatbiased}()},
\code{\link{.sehatbetahat}()},
\code{\link{.sehatslopeshatprimedelta}()},
\code{\link{.sehatslopeshatprimetb}()},
\code{\link{sehatbetahatbiased}()},
\code{\link{sehatbetahat}()},
\code{\link{sehatslopeshatprimedelta}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{standard errors of estimates of regression coefficients functions}
\keyword{se}
