% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sehat.R
\name{sehatslopesprimedelta}
\alias{sehatslopesprimedelta}
\title{Standard Errors of Standardized Estimates of Regression Coefficients (Yuan and Chan (2011))}
\usage{
sehatslopesprimedelta(X, y, adjust = FALSE)
}
\arguments{
\item{X}{\code{n} by \code{k} numeric matrix.
The data matrix \eqn{\mathbf{X}}
(also known as design matrix, model matrix or regressor matrix)
is an \eqn{n \times k} matrix of \eqn{n} observations of \eqn{k} regressors,
which includes a regressor whose value is 1 for each observation on the first column.}

\item{y}{Numeric vector of length \code{n} or \code{n} by \code{1} matrix.
The vector \eqn{\mathbf{y}} is an \eqn{n \times 1} vector of observations
on the regressand variable.}

\item{adjust}{Logical.
Use \eqn{n - 3} adjustment for small samples.}
}
\description{
Standard Errors of Standardized Estimates of Regression Coefficients (Yuan and Chan (2011))
}
\examples{
# Simple regression------------------------------------------------
X <- jeksterslabRdatarepo::wages.matrix[["X"]]
X <- X[, c(1, ncol(X))]
y <- jeksterslabRdatarepo::wages.matrix[["y"]]
sehatslopesprimedelta(X = X, y = y)

# Multiple regression----------------------------------------------
X <- jeksterslabRdatarepo::wages.matrix[["X"]]
# age is removed
X <- X[, -ncol(X)]
sehatslopesprimedelta(X = X, y = y)
}
\seealso{
Other standard errors of estimates of regression coefficients functions: 
\code{\link{.sehatbetahatbiased}()},
\code{\link{.sehatbetahat}()},
\code{\link{.sehatslopesprimedelta}()},
\code{\link{.sehatslopesprimetb}()},
\code{\link{sehatbetahatbiased}()},
\code{\link{sehatbetahat}()},
\code{\link{sehatslopesprimetb}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{standard errors of estimates of regression coefficients functions}
\keyword{inference}
