% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigma2.R
\name{sigma2}
\alias{sigma2}
\title{Error Variance \eqn{\sigma^2}}
\usage{
sigma2(slopes, sigmayX, SigmaX)
}
\arguments{
\item{slopes}{Vector or \code{p} by \code{1} matrix.
Regression slopes \eqn{\left( \boldsymbol{\beta}_{2 \cdots k} \right)} .}

\item{sigmayX}{Vector or \code{p} by \code{1} matrix of covariances between the regressand and the regressors.
\eqn{\left( \boldsymbol{\sigma}_{\mathbf{yX}} \right)}.}

\item{SigmaX}{\code{p} by \code{p} numeric matrix.
Variances and covariances of \eqn{{X}_{2}, {X}_{3}, \cdots, {X}_{k}}
\eqn{\left( \boldsymbol{\Sigma}_{\mathbf{X}} \right)}.}
}
\description{
Error variance \eqn{\sigma^2} from the slopes
\eqn{\boldsymbol{\beta}_{2 \cdots k}}
and variance-covariance matrix \eqn{\boldsymbol{\Sigma}}.
}
\seealso{
Other model-implied functions: 
\code{\link{Sigmatheta}()},
\code{\link{intercept}()},
\code{\link{mutheta}()},
\code{\link{slopes}()},
\code{\link{stdslopes}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{model-implied functions}
\keyword{model-implied,}
\keyword{parameters}
