% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcov.R
\name{sigma2_hat_rss}
\alias{sigma2_hat_rss}
\title{Residual Variance (from \eqn{RSS})}
\usage{
sigma2_hat_rss(
  rss = NULL,
  n,
  k,
  type = "unbiased",
  beta_hat = NULL,
  X = NULL,
  y = NULL
)
}
\arguments{
\item{rss}{Numeric.
Residual sum of squares.}

\item{n}{Integer.
Sample size.}

\item{k}{Integer.
Number of regressors
including a regressor
whose value is 1 for each observation.}

\item{type}{String.
Residual variance estimator.
If \code{type = "unbiased"},
returns unbiased estimate.
If \code{type = "biased"},
returns biased estimate.
If \code{type = "both"},
returns a vector of
unbiased and biased estimates.}

\item{beta_hat}{Vector or \code{k} by \code{1} matrix.
The vector
\eqn{\boldsymbol{\hat{\beta}}}
is a \eqn{k \times 1} vector of estimates
of \eqn{k} unknown regression coefficients.}

\item{X}{Matrix.
The data matrix
\eqn{\mathbf{X}}
is an \eqn{n \times k} matrix
of \eqn{n} observations
of \eqn{k} regressors,
which includes a regressor
whose value is 1 for each observation.}

\item{y}{Vector or \code{n} by \code{1} matrix.
The vector
\eqn{\mathbf{y}}
is an \eqn{n \times 1} vector
of observations on the regressand variable.}
}
\value{
Returns the estimated residual variance.
}
\description{
Calculates estimates of the error variance
\deqn{
    \mathbf{E}
      \left(
        \sigma^2
      \right)
  =
  \hat{\sigma}^2
  }
\deqn{
    \hat{\sigma}^2_{\textrm{unbiased}}
    =
    \frac{
    \mathbf{e}^{\prime}
    \mathbf{e}
    }
    {
      n - k
    }
    =
    \frac{
    RSS
    }
    {
      n - k
    }
  }
or
\deqn{
    \hat{\sigma}^2_{\textrm{biased}}
    =
    \frac{
    \mathbf{e}^{\prime}
    \mathbf{e}
    }
    {
      n
    }
    =
    \frac{
    RSS
    }
    {
      n
    }
  }
where
\eqn{\mathbf{e}} is the vector of residuals,
\eqn{RSS} is the residual sum of squares,
\eqn{n} is the sample size,
and \eqn{k} is the number of regressors
including a regressor whose value is 1 for each observation.
}
\details{
If \code{rss = NULL},
\code{rss} is computed
using \code{\link[=ss_r]{ss_r()}}
with \code{X} and \code{y} as a required arguments
and \code{beta_hat} as an optional argument.
If \code{rss} is provided,
\code{beta_hat}, \code{X}, and \code{y}
are not needed.
}
\references{
\href{https://en.wikipedia.org/wiki/Linear_regression}{Wikipedia: Linear Regression}

\href{https://en.wikipedia.org/wiki/Ordinary_least_squares}{Wikipedia: Ordinary Least Squares}
}
\author{
Ivan Jacob Agaloos Pesigan
}
