% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigma2epsilon.R
\name{sigma2epsilon}
\alias{sigma2epsilon}
\title{Error Variance \eqn{\sigma_{\varepsilon}^{2}}}
\usage{
sigma2epsilon(slopes, sigma2y, sigmayX, SigmaX)
}
\arguments{
\item{slopes}{Numeric vector of length \code{p} or \code{p} by \code{1} matrix.
\eqn{p \times 1} column vector of regression slopes \eqn{\left( \boldsymbol{\beta}_{2, 3, \cdots, k} = \left\{ \beta_2, \beta_3, \cdots, \beta_k \right\} \right)} .}

\item{sigma2y}{Numeric.
Variance of the regressand variable \eqn{\left( \sigma_{y}^{2} \right)}.}

\item{sigmayX}{Numeric vector of length \code{p} or \code{p} by \code{1} matrix.
\eqn{p \times 1} vector of covariances between the regressand \eqn{y} variable
and regressor variables \eqn{X_2, X_3, \cdots, X_k}
\eqn{\left( \boldsymbol{\sigma}_{\mathbf{y}, \mathbf{X}}
    = \left\{ \sigma_{y, X_2}, \sigma_{y, X_3}, \cdots, \sigma_{y, X_k} \right\}^{T} \right)}.}

\item{SigmaX}{\code{p} by \code{p} numeric matrix.
\eqn{p \times p} matrix of variances and covariances between regressor variables \eqn{{X}_{2}, {X}_{3}, \cdots, {X}_{k}}
\eqn{\left( \boldsymbol{\Sigma}_{\mathbf{X}} \right)}.}
}
\description{
Error variance \eqn{\sigma_{\varepsilon}^{2}} from the slopes
\eqn{\boldsymbol{\beta}_{2, \cdots, k}}
and variance-covariance matrix \eqn{\boldsymbol{\Sigma}}.
}
\examples{
slopes <- c(0.2076475, 0.4510391)
sigma2y <- 1.288163
sigmayX <- c(0.4661224, 0.5771429)
SigmaX <- matrix(
  data = c(1.2934694, 0.4379592, 0.4379592, 1.0779592),
  ncol = 2
)
sigma2epsilon(slopes = slopes, sigma2y = sigma2y, sigmayX = sigmayX, SigmaX = SigmaX)
}
\seealso{
Other parameter functions: 
\code{\link{.intercept}()},
\code{\link{.slopesprime}()},
\code{\link{.slopes}()},
\code{\link{intercept}()},
\code{\link{slopesprime}()},
\code{\link{slopes}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{parameter functions}
\keyword{parameter}
