% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcov.R
\name{sigma2hat}
\alias{sigma2hat}
\title{Residual Variance (\eqn{\hat{\sigma}^2})}
\usage{
sigma2hat(X, y, type = "unbiased")
}
\arguments{
\item{X}{Matrix.
The data matrix
\eqn{\mathbf{X}}
(also known as
design matrix,
model matrix or
regressor matrix)
is an \eqn{n \times k} matrix
of \eqn{n} observations
of \eqn{k} regressors,
which includes a regressor
whose value is 1 for each observation.}

\item{y}{Vector or \code{n} by \code{1} matrix.
The vector
\eqn{\mathbf{y}}
is an \eqn{n \times 1} vector
of observations on the regressand variable.}

\item{type}{String.
Residual variance estimator.
If \code{type = "unbiased"},
returns unbiased estimate.
If \code{type = "biased"},
returns biased estimate.
If \code{type = "both"},
returns a vector of
unbiased and biased estimates.}
}
\value{
Returns the estimated
residual variance (\eqn{\hat{\sigma}^2}).
}
\description{
Calculates estimates of the error variance
\deqn{
    \mathbf{E}
      \left(
        \sigma^2
      \right)
  =
  \hat{\sigma}^2
  }
\deqn{
    \hat{\sigma}^2_{\textrm{unbiased}}
    =
    \frac{1}{n - k}
    \sum_{i = 1}^{n}
    \left(
      \mathbf{y}
      -
      \mathbf{X} \boldsymbol{\hat{\beta}}
    \right)^2 \\
    =
    \frac{
    \mathbf{e}^{\prime}
    \mathbf{e}
    }
    {
      n - k
    } \\
    =
    \frac{
    RSS
    }
    {
      n - k
    }
  }
or
\deqn{
    \hat{\sigma}^2_{\textrm{biased}}
    =
    \frac{1}{n}
    \sum_{i = 1}^{n}
    \left(
      \mathbf{y}
      -
      \mathbf{X} \boldsymbol{\hat{\beta}}
    \right)^2 \\
    =
    \frac{
    \mathbf{e}^{\prime}
    \mathbf{e}
    }
    {
      n
    } \\
    =
    \frac{
    RSS
    }
    {
      n
    }
  }
where
\eqn{\mathbf{e}} is the vector of residuals,
\eqn{RSS} is the residual sum of squares,
\eqn{n} is the sample size,
and \eqn{k} is the number of regressors
including a regressor whose value is 1 for each observation.
}
\references{
\href{https://en.wikipedia.org/wiki/Linear_regression}{Wikipedia: Linear Regression}

\href{https://en.wikipedia.org/wiki/Ordinary_least_squares}{Wikipedia: Ordinary Least Squares}
}
\seealso{
Other residual variance functions: 
\code{\link{.sigma2hat}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{residual variance functions}
