% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ram.R
\name{Sigmatheta}
\alias{Sigmatheta}
\title{Model-Implied Variance-Covariance Matrix
\eqn{\boldsymbol{\Sigma} \left( \boldsymbol{\theta} \right)}}
\usage{
Sigmatheta(slopes, sigma2, SigmaX)
}
\arguments{
\item{slopes}{Vector or \code{p} by \code{1} matrix.
Regression slopes \eqn{\left( \boldsymbol{\beta}_{2 \cdots k} \right)} .}

\item{sigma2}{Numeric.
Variance of the error term \eqn{\varepsilon} \eqn{\left( \sigma^2 \right)}.}

\item{SigmaX}{\code{p} by \code{p} numeric matrix.
Variances and covariances of \eqn{{X}_{2}, {X}_{3}, \dots, {X}_{k}}
\eqn{\left( \boldsymbol{\Sigma}_{\mathbf{X}} \right)}.}
}
\value{
Returns the model-implied variance-covariance matrix
\eqn{\boldsymbol{\Sigma} \left( \boldsymbol{\theta} \right)}.
}
\description{
Model-implied variance-covariance matrix
\eqn{\boldsymbol{\Sigma} \left( \boldsymbol{\theta} \right)}
from parameters of a \eqn{k}-variable linear regression model.
}
\details{
The following are the parameters of a linear regression model
for the covariance structure
\itemize{
\item \eqn{\boldsymbol{\beta}_{\mathrm{slopes}}} - \eqn{p \times 1} regression slopes,
\item \eqn{\sigma^2} - variance of the error term
\eqn{\varepsilon}, and
\item \eqn{\boldsymbol{\Sigma}_{\mathbf{X}}} - \eqn{p \times p}
variances and covariances of \eqn{{X}_{2}, {X}_{3}, \dots, {X}_{k}}.
}

The model-implied variance-covariance matrix is given by
\deqn{
    \boldsymbol{\Sigma} \left( \boldsymbol{\theta} \right)
    =
    \left( \mathrm{I} - \mathrm{A} \right)^{-1}
    \mathrm{S}
    \left[ \left( \mathrm{I} - \mathrm{A} \right)^{-1} \right]^{\prime}
  }
where
\deqn{
    \mathbf{A}
    =
    \begin{bmatrix}
      0 & \beta_2 & \beta_3 & \cdots & \beta_k \\
      0 & 0       & 0       & \cdots & 0       \\
      0 & 0       & 0       & \cdots & 0       \\
      0 & 0       & 0       & \cdots & 0       \\
      0 & 0       & 0       & \cdots & 0
    \end{bmatrix}
  }

\deqn{
    \mathbf{S}
    =
    \begin{bmatrix}
      \sigma^2 & 0                         & 0                         & \cdots & 0                         \\
      0        & \mathrm{Var}_{X_2}        & \mathrm{Cov}_{X_{2}X_{3}} & \cdots & \mathrm{Cov}_{X_{2}X_{k}} \\
      0        & \mathrm{Cov}_{X_{3}X_{2}} & \mathrm{Var}_{X_3}        & \cdots & \mathrm{Cov}_{X_{3}X_{k}} \\
      0        & \vdots                    & \vdots                    & \ddots & \vdots                    \\
      0        & \mathrm{Cov}_{X_{k}X_{2}} & \mathrm{Cov}_{X_{k}X_{3}} & \cdots & \mathrm{Var}_{X_k}
    \end{bmatrix}
  }

\deqn{
    \mathbf{I}
    =
    \begin{bmatrix}
      1 & 0   & 0   & \cdots & 0   \\
      0 & 1   & 0   & \cdots & 0   \\
      0 & 0   & 1   & \cdots & 0   \\
      0 & 0   & 0   & 1      & 0   \\
      0 & 0   & 0   & \cdots & 1
    \end{bmatrix}
  }
}
\seealso{
Other model-implied functions: 
\code{\link{intercept}()},
\code{\link{mutheta}()},
\code{\link{slopes}()},
\code{\link{stdslopes}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{model-implied functions}
\keyword{model-implied}
