% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta.R
\name{slopes}
\alias{slopes}
\title{Regression Slopes \eqn{\boldsymbol{\beta}_{2 \cdots k}}}
\usage{
slopes(SigmaX, sigmayX)
}
\arguments{
\item{SigmaX}{\code{p} by \code{p} numeric matrix.
Variances and covariances of \eqn{{X}_{2}, {X}_{3}, \dots, {X}_{k}}
\eqn{\left( \boldsymbol{\Sigma}_{\mathbf{X}} \right)}.}

\item{sigmayX}{Vector or \code{p} by \code{1} matrix of covariances between the regressand and the regressors.
\eqn{\left( \boldsymbol{\sigma}_{\mathbf{yX}} \right)}.}
}
\value{
Returns the slopes \eqn{\left( \boldsymbol{\beta}_{2 \cdots k} \right)}
of a linear regression model derived from the variance-covariance matrix.
}
\description{
Derives the slopes \eqn{\boldsymbol{\beta}_{2 \cdots k}} of a linear regression model
(\eqn{\boldsymbol{\beta}} minus the intercept)
as a function of covariances.
}
\details{
The linear regression slopes is given by
\deqn{
    \boldsymbol{\beta}_{2 \cdots k}
    =
    \mathrm{V}_{\mathbf{X}}^{\prime}
    \mathrm{v}_{\mathbf{yX}}
  }
where \eqn{\boldsymbol{\Sigma}_{\mathbf{X}}}
is the \eqn{p \times p} covariance matrix of the regressor variables and
\eqn{\boldsymbol{\sigma}_{\mathbf{yX}}}
is the \eqn{p \times 1} column vector of the covariances between the regressand and the regressors.
}
\seealso{
Other model-implied functions: 
\code{\link{Sigmatheta}()},
\code{\link{intercept}()},
\code{\link{mutheta}()},
\code{\link{stdslopes}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{model-implied functions}
\keyword{beta}
\keyword{model-implied,}
