% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta.R
\name{slopesprime}
\alias{slopesprime}
\title{Regression Standardized Slopes \eqn{\boldsymbol{\beta}_{2, \cdots, k}^{\prime}}}
\usage{
slopesprime(X, y)
}
\arguments{
\item{X}{\code{n} by \code{k} numeric matrix.
The data matrix \eqn{\mathbf{X}}
(also known as design matrix, model matrix or regressor matrix)
is an \eqn{n \times k} matrix of \eqn{n} observations of \eqn{k} regressors,
which includes a regressor whose value is 1 for each observation on the first column.}

\item{y}{Numeric vector of length \code{n} or \code{n} by \code{1} matrix.
The vector \eqn{\mathbf{y}} is an \eqn{n \times 1} vector of observations
on the regressand variable.}
}
\value{
Returns the standardized slopes
\eqn{\boldsymbol{\beta}_{2, \cdots, k}^{\prime}}
of a linear regression model derived from the correlation matrix.
}
\description{
Derives the standardized slopes
\eqn{\boldsymbol{\beta}_{2, \cdots, k}^{\prime}} of a linear regression model
as a function of correlations.
}
\details{
The linear regression standardized slopes is given by
\deqn{
    \boldsymbol{\beta}_{2, \cdots, k}^{\prime} =
    \mathrm{R}_{\mathbf{X}}^{T} \mathrm{r}_{\mathbf{yX}}
  }

where
\itemize{
\item \eqn{\mathbf{R}_{\mathbf{X}}}
is the \eqn{p \times p} correlation matrix of the regressor variables and
\item \eqn{\mathbf{r}_{\mathbf{yX}}}
is the \eqn{p \times 1} column vector
of the correlations between the regressand and the regressors.
}
}
\examples{
X <- jeksterslabRdatarepo::wages.matrix[["X"]]
# age is removed
X <- X[, -ncol(X)]
y <- jeksterslabRdatarepo::wages.matrix[["y"]]
slopesprime(X = X, y = y)
}
\seealso{
Other parameter functions: 
\code{\link{.intercept}()},
\code{\link{.slopesprime}()},
\code{\link{.slopes}()},
\code{\link{intercept}()},
\code{\link{sigma2epsilon}()},
\code{\link{slopes}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{parameter functions}
\keyword{parameter}
