% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta.R
\name{stdslopes}
\alias{stdslopes}
\title{Regression Standardized Slopes \eqn{\boldsymbol{\beta}_{2 \cdots k}^{*}}}
\usage{
stdslopes(RX, ryX)
}
\arguments{
\item{RX}{\code{p} by \code{p} numeric matrix.
Correlations between the regressors.}

\item{ryX}{Vector or \code{p} by \code{1} matrix of correlations between the regressand and the regressors.}
}
\value{
Returns the standardized slopes \eqn{\left( \boldsymbol{\beta}_{2 \cdots k}^{*} \right)}
of a linear regression model derived from the correlation matrix.
}
\description{
Derives the standardized slopes \eqn{\boldsymbol{\beta}_{2 \cdots k}^{*}} of a linear regression model
as a function of correlations.
}
\details{
The linear regression standardized slopes is given by
\deqn{
    \boldsymbol{\beta}_{2 \cdots k}^{*}
    =
    \mathrm{R}_{\mathbf{X}}^{\prime}
    \mathrm{r}_{\mathbf{yX}}
  }
where \eqn{\mathbf{R}_{\mathbf{X}}}
is the \eqn{p \times p} correlation matrix of the regressor variables and
\eqn{\mathbf{r}_{\mathbf{yX}}}
is the \eqn{p \times 1} column vector of the correlations between the regressand and the regressors.
}
\seealso{
Other model-implied functions: 
\code{\link{Sigmatheta}()},
\code{\link{intercept}()},
\code{\link{mutheta}()},
\code{\link{slopes}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{model-implied functions}
\keyword{beta}
\keyword{model-implied,}
