% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tepsilonhat.R
\name{tepsilonhat}
\alias{tepsilonhat}
\title{Studentized Residuals}
\usage{
tepsilonhat(X, y)
}
\arguments{
\item{X}{\code{n} by \code{k} numeric matrix.
The data matrix \eqn{\mathbf{X}}
(also known as design matrix, model matrix or regressor matrix)
is an \eqn{n \times k} matrix of \eqn{n} observations of \eqn{k} regressors,
which includes a regressor whose value is 1 for each observation on the first column.}

\item{y}{Numeric vector of length \code{n} or \code{n} by \code{1} matrix.
The vector \eqn{\mathbf{y}} is an \eqn{n \times 1} vector of observations
on the regressand variable.}
}
\value{
Returns studentized residuals.
}
\description{
Calculates studentized residuals using
\deqn{
    t_i = \frac{\hat{\varepsilon}_{i}}{\hat{\sigma}_{\varepsilon}^{2} \sqrt{1 - h_{ii}}}
  }
}
\examples{
# Simple regression------------------------------------------------
X <- jeksterslabRdatarepo::wages.matrix[["X"]]
X <- X[, c(1, ncol(X))]
y <- jeksterslabRdatarepo::wages.matrix[["y"]]
tepsilonhat <- tepsilonhat(X = X, y = y)
hist(tepsilonhat)

# Multiple regression----------------------------------------------
X <- jeksterslabRdatarepo::wages.matrix[["X"]]
# age is removed
X <- X[, -ncol(X)]
tepsilonhat <- tepsilonhat(X = X, y = y)
hist(tepsilonhat)
}
\seealso{
Other residuals functions: 
\code{\link{.My}()},
\code{\link{.tepsilonhat}()},
\code{\link{.yminusyhat}()},
\code{\link{My}()},
\code{\link{epsilonhat}()},
\code{\link{yminusyhat}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{residuals functions}
\keyword{residuals}
