% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss.R
\name{tss}
\alias{tss}
\title{Total Sum of Squares.}
\usage{
tss(y)
}
\arguments{
\item{y}{Vector or \code{n} by \code{1} matrix.
The vector
\eqn{\mathbf{y}}
is an \eqn{n \times 1} vector
of observations on the regressand variable.}
}
\value{
Returns the total sum of squares.
}
\description{
Calculates the total sum of squares (TSS)
\deqn{
    TSS
    =
    \mathbf{y}^{\prime}
    \mathbf{y}
    -
    n
    \mathbf{\bar{Y}}^{2}.
  }
Note that
\deqn{
    TSS
    =
    ESS
    +
    RSS.
  }
}
\references{
\href{https://en.wikipedia.org/wiki/Residual_sum_of_squares}{Wikipedia: Residual Sum of Squares}

\href{https://en.wikipedia.org/wiki/Explained_sum_of_squares}{Wikipedia: Explained Sum of Squares}

\href{https://en.wikipedia.org/wiki/Total_sum_of_squares}{Wikipedia: Total Sum of Squares}

\href{https://en.wikipedia.org/wiki/Coefficient_of_determination}{Wikipedia: Coefficient of Determination}
}
\seealso{
Other sum of squares functions: 
\code{\link{.rss}()},
\code{\link{ess}()},
\code{\link{rss}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{sum of squares functions}
