% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcov.R
\name{vcov_betahat}
\alias{vcov_betahat}
\title{Variance-Covariance Matrix of Estimates of Regression Coefficients}
\usage{
vcov_betahat(X, y, type = "unbiased")
}
\arguments{
\item{X}{Matrix.
The data matrix
\eqn{\mathbf{X}}
(also known as
design matrix,
model matrix or
regressor matrix)
is an \eqn{n \times k} matrix
of \eqn{n} observations
of \eqn{k} regressors,
which includes a regressor
whose value is 1 for each observation.}

\item{y}{Vector or \code{n} by \code{1} matrix.
The vector
\eqn{\mathbf{y}}
is an \eqn{n \times 1} vector
of observations on the regressand variable.}

\item{type}{String.
Residual variance estimator.
If \code{type = "unbiased"},
returns unbiased estimate.
If \code{type = "biased"},
returns biased estimate.
If \code{type = "both"},
returns a vector of
unbiased and biased estimates.}
}
\value{
Returns the variance-covariance matrix
of estimates of regression coefficients.
}
\description{
Calculates the variance-covariance matrix
of estimates of regression coefficients using
\deqn{
    \hat{\sigma}^2
    \left(
      \mathbf{X}^{\prime}
      \mathbf{X}
    \right)^{-1}
  }
where \eqn{\hat{\sigma}^2}
is the estimate of the error variance \eqn{{\sigma}^2}
and \eqn{\mathbf{X}} is
the data matrix,
that is,
an \eqn{n \times k} matrix
of \eqn{n} observations
of \eqn{k} regressors,
which includes a regressor
whose value is 1 for each observation.
}
\references{
\href{https://en.wikipedia.org/wiki/Linear_regression}{Wikipedia: Linear Regression}

\href{https://en.wikipedia.org/wiki/Ordinary_least_squares}{Wikipedia: Ordinary Least Squares}
}
\seealso{
Other variance-covariance functions: 
\code{\link{.vcov_betahat}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{variance-covariance functions}
