% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/y_hat.R
\name{y_hat_Py}
\alias{y_hat_Py}
\title{y-hat (\eqn{\mathbf{\hat{y}} = \mathbf{P} \mathbf{y}})}
\usage{
y_hat_Py(y, P = NULL, X = NULL)
}
\arguments{
\item{y}{Vector or \code{n} by \code{1} matrix.
The vector
\eqn{\mathbf{y}}
is an \eqn{n \times 1} vector
of observations on the regressand variable.}

\item{P}{Matrix.
The projection matrix
(\eqn{\mathbf{P}}).}

\item{X}{Matrix.
The data matrix
\eqn{\mathbf{X}}
is an \eqn{n \times k} matrix
of \eqn{n} observations
of \eqn{k} regressors,
which includes a regressor
whose value is 1 for each observation.}
}
\value{
Returns y-hat (\eqn{\mathbf{\hat{y}}}).
}
\description{
Calculates y-hat using
\deqn{
    \mathbf{\hat{y}}
    =
    \mathbf{P}
    \mathbf{y}.
  }
}
\details{
If \code{P = NULL},
the \code{P} matrix is computed
using \code{\link[=proj_P]{proj_P()}}
with \code{X} as its argument.
If \code{P} is provided,
\code{X} is not needed.
}
\seealso{
Other y-hat functions: 
\code{\link{y_hat_Xbeta_hat}()},
\code{\link{y_hat_test}()},
\code{\link{y_hat}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{y-hat functions}
