% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/y_hat.R
\name{y_hat_test}
\alias{y_hat_test}
\title{y-hat for Testing Data  Set (\eqn{\mathbf{\hat{y}} = \mathbf{X} \boldsymbol{\hat{\beta}}})}
\usage{
y_hat_test(beta_hat, X, y = NULL)
}
\arguments{
\item{beta_hat}{Vector or \code{k} by \code{1} matrix.
\eqn{k \times 1} vector of estimates
of \eqn{k} unknown regression coefficients
(\eqn{\boldsymbol{\hat{\beta}}}).
Estimated from \strong{training data set}.}

\item{X}{Matrix.
The data matrix (\eqn{\mathbf{X}}),
that is an \eqn{n \times k} matrix of \eqn{n} observations
of \eqn{k} regressors,
which includes a regressor whose value is 1 for each observation
from the \strong{testing data set}.}

\item{y}{Vector or \code{n} by \code{1} matrix.
\eqn{n \times 1} vector of observations on the regressand variable
(\eqn{\mathbf{y}})
from the \strong{testing data set}.
This is an optional argument
for model performance assessment.}
}
\value{
Returns a list with the following elements
\describe{
\item{y_hat}{Predicted values of \eqn{\mathbf{y}} using the \code{beta_hat} (\eqn{\mathbf{\hat{\beta}}}) from traning data set and \code{X} (\eqn{\mathbf{X}}) from testing data set.}
\item{RSS}{Residual sum of squares. \code{NA} if \code{y = NULL}.}
\item{MSE}{Mean-square error. \code{NA} if \code{y = NULL}.}
\item{RMSE}{Root-mean-square error. \code{NA} if \code{y = NULL}.}
\item{R_squared}{Coefficient of determinism. \code{NA} if \code{y = NULL}.}
\item{adjusted_R_squared}{Adjusted R-squared. \code{NA} if \code{y = NULL}.}
}
}
\description{
Calculates y-hat (\eqn{\mathbf{\hat{y}}}), that is,
the predicted value of \eqn{\mathbf{y}}
given \eqn{\mathbf{X}} using
\deqn{
    \mathbf{\hat{y}}
    =
    \mathbf{X}
    \boldsymbol{\hat{\beta}}
  }
for the \strong{testing data set}.
}
\references{
\href{https://en.wikipedia.org/wiki/Linear_regression}{Wikipedia: Linear Regression}

\href{https://en.wikipedia.org/wiki/Ordinary_least_squares}{Wikipedia: Ordinary Least Squares}
}
\seealso{
Other y-hat functions: 
\code{\link{y_hat_Py}()},
\code{\link{y_hat_Xbeta_hat}()},
\code{\link{y_hat}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{y-hat functions}
