% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/y_hat.R
\name{y_hat_Xbeta_hat}
\alias{y_hat_Xbeta_hat}
\title{y-hat (\eqn{\mathbf{\hat{y}} = \mathbf{X} \boldsymbol{\hat{\beta}}})}
\usage{
y_hat_Xbeta_hat(X, beta_hat = NULL, y = NULL)
}
\arguments{
\item{X}{Matrix.
The data matrix
\eqn{\mathbf{X}}
is an \eqn{n \times k} matrix
of \eqn{n} observations
of \eqn{k} regressors,
which includes a regressor
whose value is 1 for each observation.}

\item{beta_hat}{Vector or \code{k} by \code{1} matrix.
The vector
\eqn{\boldsymbol{\hat{\beta}}}
is a \eqn{k \times 1} vector of estimates
of \eqn{k} unknown regression coefficients.}

\item{y}{Vector or \code{n} by \code{1} matrix.
The vector
\eqn{\mathbf{y}}
is an \eqn{n \times 1} vector
of observations on the regressand variable.}
}
\value{
Returns y-hat (\eqn{\mathbf{\hat{y}}}).
}
\description{
Calculates y-hat (\eqn{\mathbf{\hat{y}}}), that is,
the predicted value of \eqn{\mathbf{y}}
given \eqn{\mathbf{X}} using
\deqn{
    \mathbf{\hat{y}}
    =
    \mathbf{X}
    \boldsymbol{\hat{\beta}}.
  }
}
\details{
If \code{beta_hat = NULL},
the \code{beta_hat} vector is computed
using \code{\link[=beta_hat_inv]{beta_hat_inv()}}
with \code{X} and \code{y}
as arguments.
If \code{beta_hat} is provided,
\code{y} is not needed.
}
\references{
\href{https://en.wikipedia.org/wiki/Linear_regression}{Wikipedia: Linear Regression}

\href{https://en.wikipedia.org/wiki/Ordinary_least_squares}{Wikipedia: Ordinary Least Squares}
}
\seealso{
Other y-hat functions: 
\code{\link{y_hat_Py}()},
\code{\link{y_hat_test}()},
\code{\link{y_hat}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{y-hat functions}
