% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e.R
\name{y_minus_yhat}
\alias{y_minus_yhat}
\title{Residuals (\eqn{\mathbf{e} = \mathbf{y} - \mathbf{\hat{y}}})}
\usage{
y_minus_yhat(y, yhat = NULL, X = NULL, betahat = NULL)
}
\arguments{
\item{y}{Vector or \code{n} by \code{1} matrix.
The vector
\eqn{\mathbf{y}}
is an \eqn{n \times 1} vector
of observations on the regressand variable.}

\item{yhat}{Vector or \code{n} by \code{1} matrix.
\eqn{n \times 1} vector
of predicted values of \eqn{\mathbf{y}}
(\eqn{\mathbf{\hat{y}}}).}

\item{X}{Matrix.
The data matrix
\eqn{\mathbf{X}}
is an \eqn{n \times k} matrix
of \eqn{n} observations
of \eqn{k} regressors,
which includes a regressor
whose value is 1 for each observation.}

\item{betahat}{Vector or \code{k} by \code{1} matrix.
The vector
\eqn{\boldsymbol{\hat{\beta}}}
is a \eqn{k \times 1} vector of estimates
of \eqn{k} unknown regression coefficients.}
}
\value{
Returns residuals (\eqn{\mathbf{e}}),
that is,
the difference between
the observed and predicted value of \eqn{\mathbf{y}}
(\eqn{\mathbf{e} = \mathbf{y}} - \eqn{\mathbf{\hat{y}}}).
}
\description{
Calculates residuals using
\deqn{
    \mathbf{e}
    =
    \mathbf{y}
    -
    \mathbf{\hat{y}}
    =
    \mathbf{y}
    -
    \mathbf{X}
    \boldsymbol{\hat{\beta}}.
  }
}
\details{
If \code{yhat = NULL},
the \code{yhat} vector is computed
using \code{\link[=Xbetahat]{Xbetahat()}}
with \code{X} as a required argument
and \code{betahat} as an optional argument.
If \code{yhat} is provided,
\code{X} and \code{betahat} are not needed.
}
\references{
\href{https://en.wikipedia.org/wiki/Errors_and_residuals}{Wikipedia: Errors and Residuals}
}
\seealso{
Other residuals functions: 
\code{\link{My}()},
\code{\link{e}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{residuals functions}
