% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci.R
\name{ci}
\alias{ci}
\title{Confidence Intervals of Estimates of Regression Coefficients}
\usage{
ci(betahat, se, n, k)
}
\arguments{
\item{betahat}{Vector of length \code{k} or \code{k} by \code{1} matrix.
The vector \eqn{\boldsymbol{\hat{\beta}}} is a \eqn{k \times 1} vector of estimates
of \eqn{k} unknown regression coefficients.}

\item{se}{Numeric vector.
Standard errors of regression coefficients.}

\item{n}{Integer.
Sample size.}

\item{k}{Integer.
Number of regressors including a regressor
whose value is 1 for each observation on the first column.}
}
\description{
Confidence Intervals of Estimates of Regression Coefficients
}
\seealso{
Other hypothesis testing functions: 
\code{\link{.anovatable}()},
\code{\link{anovatable}()},
\code{\link{nhst}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{hypothesis testing functions}
\keyword{inference}
