% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptives.R
\name{descriptives}
\alias{descriptives}
\title{Descriptive Statistics}
\usage{
descriptives(X, y, muhat = TRUE)
}
\arguments{
\item{X}{Matrix.
The data matrix
\eqn{\mathbf{X}}
(also known as
design matrix,
model matrix or
regressor matrix)
is an \eqn{n \times k} matrix
of \eqn{n} observations
of \eqn{k} regressors,
which includes a regressor
whose value is 1 for each observation.}

\item{y}{Vector or \code{n} by \code{1} matrix.
The vector
\eqn{\mathbf{y}}
is an \eqn{n \times 1} vector
of observations on the regressand variable.}

\item{muhat}{Logical.
If \code{TRUE}, compute means.}
}
\value{
Returns a list with the following items
\describe{
\item{Sigmahat}{
Variance-covariance matrix
\eqn{
      \left(
        \boldsymbol{
          \hat{
            \Sigma
          }
        }
      \right)
    } .
}
\item{R}{
Correlation matrix
\eqn{
      \left(
        \mathrm{
          R
        }
      \right)
    } .
}
\item{VX}{Variance-covariance matrix
\eqn{
      \left(
        \mathbf{
          V
        }_{
          \mathbf{
            X
          }
        }
      \right)
    } .
}
\item{vyX}{Vector of covariances between
\eqn{
      \mathbf{
        y
      }
    }
and
\eqn{
      \mathbf{
        X
      }
    }
\eqn{
      \left(
        \mathbf{
          v
        }_{
          \mathbf{
            yX
          }
        }
      \right)
    } .
}
\item{RX}{Correlation matrix
\eqn{
      \left(
        \mathbf{
          R
        }_{
          \mathbf{
            X
          }
        }
      \right)
    } .
}
\item{ryX}{Vector of correlations between
\eqn{
      \mathbf{
        y
      }
    }
and
\eqn{
      \mathbf{
        X
      }
    }
\eqn{
      \left(
        \mathbf{
          r
        }_{
          \mathbf{
            yX
          }
        }
      \right)
    } .
}
\item{muhatX}{Means of regressor variables
\eqn{
      \boldsymbol{
        \hat{
          \mu
        }
      }_{
        \mathbf{
          X
        }
      }
    } .
}
\item{muhaty}{Mean of regressand variable
\eqn{
      \hat{
        \mu
      }_{
        \mathbf{
          y
        }
      }
    } .
}
}
Note that \code{muhatX} and \code{muhaty} are \code{NA}
when \code{muhat = FALSE}.
}
\description{
Calculate means, covariances and correlations.
}
\author{
Ivan Jacob Agaloos Pesigan
}
