% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptives.R
\name{descriptives}
\alias{descriptives}
\title{Descriptive Statistics}
\usage{
descriptives(
  X,
  y,
  varnamesX = NULL,
  varnamey = NULL,
  plot = TRUE,
  msd = TRUE,
  cor = TRUE
)
}
\arguments{
\item{X}{\code{n} by \code{k} numeric matrix.
The data matrix \eqn{\mathbf{X}}
(also known as design matrix, model matrix or regressor matrix)
is an \eqn{n \times k} matrix of \eqn{n} observations of \eqn{k} regressors,
which includes a regressor whose value is 1 for each observation on the first column.}

\item{y}{Numeric vector of length \code{n} or \code{n} by \code{1} matrix.
The vector \eqn{\mathbf{y}} is an \eqn{n \times 1} vector of observations
on the regressand variable.}

\item{varnamesX}{Character vector of length \code{k}.
Variable names for matrix \code{X}.}

\item{varnamey}{Character string.
Variable name for vector \code{y}.}

\item{plot}{Logical.
Display scatter plot matrix.}

\item{msd}{Logical.
Print means and standard deviations.}

\item{cor}{Logical.
Print correlations.}
}
\value{
Returns descriptive statistics useful in k-variable linear regression model.
}
\description{
Descriptive Statistics
}
\examples{
X <- jeksterslabRdatarepo::wages.matrix[["X"]]
y <- jeksterslabRdatarepo::wages.matrix[["y"]]
descriptives(
  X = X,
  y = y
)
}
\author{
Ivan Jacob Agaloos Pesigan
}
\keyword{descriptives}
