% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptives.R
\name{descriptives}
\alias{descriptives}
\title{Descriptive Statistics}
\usage{
descriptives(
  X,
  y,
  varnamesX = NULL,
  varnamey = NULL,
  plot = TRUE,
  moments = TRUE,
  cor = TRUE,
  mardia = TRUE
)
}
\arguments{
\item{X}{\code{n} by \code{k} numeric matrix.
The data matrix \eqn{\mathbf{X}}
(also known as design matrix, model matrix or regressor matrix)
is an \eqn{n \times k} matrix of \eqn{n} observations of \eqn{k} regressors,
which includes a regressor whose value is 1 for each observation on the first column.}

\item{y}{Numeric vector of length \code{n} or \code{n} by \code{1} matrix.
The vector \eqn{\mathbf{y}} is an \eqn{n \times 1} vector of observations
on the regressand variable.}

\item{varnamesX}{Optional. Character vector of length \code{k}.
Variable names for matrix \code{X}.}

\item{varnamey}{Optional. Character string.
Variable name for vector \code{y}.}

\item{plot}{Logical.
Display scatter plot matrix.}

\item{moments}{Logical.
Print central moments (means, standard deviations, skewness, and kurtosis).}

\item{cor}{Logical.
Print correlations.}

\item{mardia}{Logical.
Estimate Mardia's multivariate skewness and kurtosis.}
}
\value{
Returns descriptive statistics useful in k-variable linear regression model.
}
\description{
Descriptive Statistics
}
\examples{
# Simple regression------------------------------------------------
X <- jeksterslabRdatarepo::wages.matrix[["X"]]
X <- X[, c(1, ncol(X))]
y <- jeksterslabRdatarepo::wages.matrix[["y"]]
out <- descriptives(X = X, y = y)
str(out)

# Multiple regression----------------------------------------------
X <- jeksterslabRdatarepo::wages.matrix[["X"]]
# age is removed
X <- X[, -ncol(X)]
out <- descriptives(X = X, y = y)
str(out)
}
\author{
Ivan Jacob Agaloos Pesigan
}
\keyword{descriptives}
