% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSE.R
\name{.MSE}
\alias{.MSE}
\title{Mean Squared Error (from \eqn{\mathrm{RSS}})}
\usage{
.MSE(RSS = NULL, n, X, y)
}
\arguments{
\item{RSS}{Numeric.
Residual sum of squares.}

\item{n}{Integer.
Sample size.}

\item{X}{\code{n} by \code{k} numeric matrix.
The data matrix \eqn{\mathbf{X}}
(also known as design matrix, model matrix or regressor matrix)
is an \eqn{n \times k} matrix of \eqn{n} observations of \eqn{k} regressors,
which includes a regressor whose value is 1 for each observation on the first column.}

\item{y}{Numeric vector of length \code{n} or \code{n} by \code{1} matrix.
The vector \eqn{\mathbf{y}} is an \eqn{n \times 1} vector of observations
on the regressand variable.}
}
\value{
Returns the mean squared error.
}
\description{
Calculates the mean squared error \eqn{\left( \mathrm{MSE} \right)} using
\deqn{
    \mathrm{MSE}  = \frac{1}{n} \sum_{i = 1}^{n}
    \left( \mathbf{y} - \mathbf{X} \boldsymbol{\hat{\beta}} \right)^{2} \\
    = \frac{1}{n} \sum_{i = 1}^{n} \left( \mathbf{y} - \mathbf{\hat{y}} \right)^{2} \\
    = \frac{\mathrm{RSS}}{n} .
  }
}
\details{
If \code{RSS = NULL}, the \code{RSS} vector is computed using \code{\link[=RSS]{RSS()}}
with \code{X} and \code{y} as required arguments and \code{betahat} as an optional argument.
If \code{RSS} is provided, \code{X}, and \code{y} are not needed.
}
\references{
\href{https://en.wikipedia.org/wiki/Mean_squared_error}{Wikipedia: Mean Squared Error}
}
\seealso{
Other assessment of model quality functions: 
\code{\link{.R2fromESS}()},
\code{\link{.R2fromRSS}()},
\code{\link{.RMSE}()},
\code{\link{.Rbar2}()},
\code{\link{.model}()},
\code{\link{MSE}()},
\code{\link{R2}()},
\code{\link{RMSE}()},
\code{\link{Rbar2}()},
\code{\link{model}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{assessment of model quality functions}
\keyword{model-assessment}
