% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSE.R
\name{.RMSE}
\alias{.RMSE}
\title{Root Mean Squared Error (from \eqn{RSS})}
\usage{
.RMSE(RSS = NULL, n, X, y, betahat = NULL)
}
\arguments{
\item{RSS}{Numeric.
Residual sum of squares.}

\item{n}{Integer.
Sample size.}

\item{X}{Matrix.
The data matrix
\eqn{\mathbf{X}}
(also known as
design matrix,
model matrix or
regressor matrix)
is an \eqn{n \times k} matrix
of \eqn{n} observations
of \eqn{k} regressors,
which includes a regressor
whose value is 1 for each observation.}

\item{y}{Vector or \code{n} by \code{1} matrix.
The vector
\eqn{\mathbf{y}}
is an \eqn{n \times 1} vector
of observations on the regressand variable.}

\item{betahat}{Vector or \code{k} by \code{1} matrix.
The vector
\eqn{\boldsymbol{\hat{\beta}}}
is a \eqn{k \times 1} vector of estimates
of \eqn{k} unknown regression coefficients.}
}
\value{
Returns the root mean squared error.
}
\description{
Calculates the root mean squared error (RMSE) using
\deqn{
    RMSE
    =
    \sqrt{
      \frac{1}{n}
      \sum_{i = 1}^{n}
      \left(
        \mathbf{y} - \mathbf{X} \boldsymbol{\hat{\beta}}
      \right)^{2}
    } \\
    =
    \sqrt{
      \frac{1}{n}
      \sum_{i = 1}^{n}
      \left(
        \mathbf{y} - \mathbf{\hat{y}}
      \right)^{2}
    } \\
    =
    \sqrt{\frac{RSS}{n}}.
  }
}
\details{
If \code{RSS = NULL},
the \code{RSS} vector is computed
using \code{\link[=RSS]{RSS()}}
with \code{X} and \code{y} as required arguments
and \code{betahat} as an optional argument.
If \code{RSS} is provided,
\code{betahat}, \code{X}, and \code{y}
are not needed.
}
\references{
\href{https://en.wikipedia.org/wiki/Root-mean-square_deviation}{Wikipedia: Root Mean Square Deviation}
}
\seealso{
Other assessment of model quality functions: 
\code{\link{.MSE}()},
\code{\link{.R2_ESS}()},
\code{\link{.R2_RSS}()},
\code{\link{.Rbar2}()},
\code{\link{MSE}()},
\code{\link{R2}()},
\code{\link{RMSE}()},
\code{\link{Rbar2}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{assessment of model quality functions}
