% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta.R
\name{.slopes}
\alias{.slopes}
\title{Regression Slopes \eqn{\boldsymbol{\beta}_{2, \cdots, k}}}
\usage{
.slopes(SigmaX = NULL, sigmayX = NULL, X, y)
}
\arguments{
\item{SigmaX}{\code{p} by \code{p} numeric matrix.
\eqn{p \times p} variances and covariances of \eqn{{X}_{2}, {X}_{3}, \cdots, {X}_{k}}
\eqn{\left( \boldsymbol{\Sigma}_{\mathbf{X}} \right)}.}

\item{sigmayX}{Numeric vector of length \code{p} or \code{p} by \code{1} matrix.
\eqn{p \times 1} covariances between the regressand and the regressors
\eqn{\left( \boldsymbol{\sigma}_{\mathbf{yX}} \right)}.}

\item{X}{\code{n} by \code{k} numeric matrix.
The data matrix \eqn{\mathbf{X}}
(also known as design matrix, model matrix or regressor matrix)
is an \eqn{n \times k} matrix of \eqn{n} observations of \eqn{k} regressors,
which includes a regressor whose value is 1 for each observation on the first column.}

\item{y}{Numeric vector of length \code{n} or \code{n} by \code{1} matrix.
The vector \eqn{\mathbf{y}} is an \eqn{n \times 1} vector of observations
on the regressand variable.}
}
\value{
Returns the slopes \eqn{\boldsymbol{\beta}_{2, \cdots, k}}
of a linear regression model derived from the variance-covariance matrix.
}
\description{
Derives the slopes \eqn{\boldsymbol{\beta}_{2, \cdots, k}}
of a linear regression model (\eqn{\boldsymbol{\beta}} minus the intercept)
as a function of covariances.
}
\details{
The linear regression slopes is given by
\deqn{
    \boldsymbol{\beta}_{2, \cdots, k} =
    \mathrm{V}_{\mathbf{X}}^{T} \mathrm{v}_{\mathbf{yX}}
  }

where
\itemize{
\item \eqn{\boldsymbol{\Sigma}_{\mathbf{X}}}
is the \eqn{p \times p} covariance matrix of the regressor variables and
\item \eqn{\boldsymbol{\sigma}_{\mathbf{yX}}}
is the \eqn{p \times 1} column vector
of the covariances between the regressand and the regressors.
}
}
\seealso{
Other parameter functions: 
\code{\link{.intercept}()},
\code{\link{.slopesprime}()},
\code{\link{intercept}()},
\code{\link{sigma2epsilon}()},
\code{\link{slopesprime}()},
\code{\link{slopes}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{parameter functions}
\keyword{parameter}
