% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SS.R
\name{ESS}
\alias{ESS}
\title{Explained Sum of Squares}
\usage{
ESS(X, y)
}
\arguments{
\item{X}{\code{n} by \code{k} numeric matrix.
The data matrix \eqn{\mathbf{X}}
(also known as design matrix, model matrix or regressor matrix)
is an \eqn{n \times k} matrix of \eqn{n} observations of \eqn{k} regressors,
which includes a regressor whose value is 1 for each observation on the first column.}

\item{y}{Numeric vector of length \code{n} or \code{n} by \code{1} matrix.
The vector \eqn{\mathbf{y}} is an \eqn{n \times 1} vector of observations
on the regressand variable.}
}
\value{
Returns the explained sum of squares \eqn{\left( ESS \right)}.
}
\description{
Calculates the explained sum of squares \eqn{\left( ESS \right)}
using
\deqn{
    ESS = \sum_{i = 1}^{n} \left( \hat{Y}_{i} - \bar{Y} \right)^2 \\
    = \sum_{i = 1}^{n} \left( \hat{\beta}_{1} + \hat{\beta}_{2} X_{2i} +
    \hat{\beta}_{3} X_{3i} + \dots + \hat{\beta}_{k} X_{ki} - \bar{Y} \right)^2
  }
In matrix form
\deqn{
    ESS = \sum_{i = 1}^{n} \left( \mathbf{\hat{y}} - \mathbf{\bar{Y}} \right)^2 \\
    = \sum_{i = 1}^{n} \left( \mathbf{X} \boldsymbol{\hat{\beta}} - \mathbf{\bar{Y}} \right)^2
  }
where
\eqn{\mathbf{\hat{y}}} \eqn{\left( \mathbf{X} \boldsymbol{\hat{\beta}} \right)}
is an \eqn{n \times 1} matrix of predicted values of \eqn{\mathbf{y}},
and \eqn{\mathbf{\bar{Y}}} is the mean of \eqn{\mathbf{y}}.
Equivalent computational matrix formula
\deqn{
    ESS = \boldsymbol{\hat{\beta}}^{\prime} \mathbf{X}^{\prime} \mathbf{X}
    \boldsymbol{\hat{\beta}} - n \mathbf{\bar{Y}}^{2}.
  }
Note that
\deqn{
    TSS = ESS + RSS .
  }
}
\references{
\href{https://en.wikipedia.org/wiki/Residual_sum_of_squares}{Wikipedia: Residual Sum of Squares}

\href{https://en.wikipedia.org/wiki/Explained_sum_of_squares}{Wikipedia: Explained Sum of Squares}

\href{https://en.wikipedia.org/wiki/Total_sum_of_squares}{Wikipedia: Total Sum of Squares}

\href{https://en.wikipedia.org/wiki/Coefficient_of_determination}{Wikipedia: Coefficient of Determination}
}
\seealso{
Other sum of squares functions: 
\code{\link{.ESS}()},
\code{\link{.RSS}()},
\code{\link{RSS}()},
\code{\link{TSS}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{sum of squares functions}
\keyword{SS}
