% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{linreg}
\alias{linreg}
\title{Linear Regression}
\usage{
linreg(
  X,
  y,
  varnamesX = NULL,
  varnamey = NULL,
  qr = TRUE,
  unbiased = TRUE,
  plot = TRUE,
  print = TRUE
)
}
\arguments{
\item{X}{\code{n} by \code{k} numeric matrix.
The data matrix \eqn{\mathbf{X}}
(also known as design matrix, model matrix or regressor matrix)
is an \eqn{n \times k} matrix of \eqn{n} observations of \eqn{k} regressors,
which includes a regressor whose value is 1 for each observation on the first column.}

\item{y}{Numeric vector of length \code{n} or \code{n} by \code{1} matrix.
The vector \eqn{\mathbf{y}} is an \eqn{n \times 1} vector of observations
on the regressand variable.}

\item{varnamesX}{Character vector of length \code{k}.
Variable names for matrix \code{X}.}

\item{varnamey}{Character string.
Variable name for vector \code{y}.}

\item{qr}{Logical.
If \code{TRUE}, use QR decomposition when normal equations fail.
If \code{FALSE}, use singular value decompositon when normal equations fail.}

\item{unbiased}{Logical.
Use unbiased standard errors for regression coefficients
hypothesis test.}

\item{plot}{Logical.
Display plots.}

\item{print}{Logical.
Display summary output.}
}
\description{
Linear Regression
}
\author{
Ivan Jacob Agaloos Pesigan
}
\keyword{linreg}
