% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{linreg}
\alias{linreg}
\title{Linear Regression}
\usage{
linreg(
  X,
  y,
  varnamesX = NULL,
  varnamey = NULL,
  qr = TRUE,
  sehatbetahattype = "unbiased",
  sehatslopesprimetype = "textbook",
  adjust = FALSE,
  plot = TRUE,
  print = TRUE
)
}
\arguments{
\item{X}{\code{n} by \code{k} numeric matrix.
The data matrix \eqn{\mathbf{X}}
(also known as design matrix, model matrix or regressor matrix)
is an \eqn{n \times k} matrix of \eqn{n} observations of \eqn{k} regressors,
which includes a regressor whose value is 1 for each observation on the first column.}

\item{y}{Numeric vector of length \code{n} or \code{n} by \code{1} matrix.
The vector \eqn{\mathbf{y}} is an \eqn{n \times 1} vector of observations
on the regressand variable.}

\item{varnamesX}{Optional. Character vector of length \code{k}.
Variable names for matrix \code{X}.}

\item{varnamey}{Optional. Character string.
Variable name for vector \code{y}.}

\item{qr}{Logical.
If \code{TRUE}, use QR decomposition when normal equations fail.
If \code{FALSE}, use singular value decompositon when normal equations fail.}

\item{sehatbetahattype}{Character string.
Standard errors for regression coefficients hypothesis test.
Options are \code{sehatbetahattype = "unbiased"} and \code{sehatbetahattype = "biased"}.}

\item{sehatslopesprimetype}{Character string.
Standard errors for standardized regression slopes hypothesis test.
Options are \code{sehatslopesprimetype = "textbook"} and \code{sehatslopesprimetype = "delta"}.}

\item{adjust}{Logical.
If \code{sehatslopesprimetype = "delta"} and \code{adjust = TRUE},
uses \code{n - 3} to adjust \code{sehatslopesprime} for bias.
This adjustment is recommended for small sample sizes.}

\item{plot}{Logical.
Display plots.}

\item{print}{Logical.
Display summary output.}
}
\description{
Linear Regression
}
\examples{
# Simple regression------------------------------------------------
X <- jeksterslabRdatarepo::wages.matrix[["X"]]
X <- X[, c(1, ncol(X))]
y <- jeksterslabRdatarepo::wages.matrix[["y"]]
linreg(X = X, y = y)

# Multiple regression----------------------------------------------
X <- jeksterslabRdatarepo::wages.matrix[["X"]]
# age is removed
X <- X[, -ncol(X)]
linreg(X = X, y = y)
}
\author{
Ivan Jacob Agaloos Pesigan
}
\keyword{linreg}
