% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{linreg}
\alias{linreg}
\title{Linear Regression}
\usage{
linreg(X, y, FUN = betahat_inv, output = c("coef", "model", "anova"))
}
\arguments{
\item{X}{Matrix.
The data matrix
\eqn{\mathbf{X}}
is an \eqn{n \times k} matrix
of \eqn{n} observations
of \eqn{k} regressors,
which includes a regressor
whose value is 1 for each observation.}

\item{y}{Vector or \code{n} by \code{1} matrix.
The vector
\eqn{\mathbf{y}}
is an \eqn{n \times 1} vector
of observations on the regressand variable.}

\item{FUN}{Function.
Beta-hat function to use.
\code{\link[=betahat_inv]{betahat_inv()}} is used by default.}

\item{output}{Character vector.
Results to print.}
}
\description{
Linear Regression
}
\author{
Ivan Jacob Agaloos Pesigan
}
