% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proj.R
\name{M}
\alias{M}
\title{M Matrix}
\usage{
M(X)
}
\arguments{
\item{X}{\code{n} by \code{k} numeric matrix.
The data matrix \eqn{\mathbf{X}}
(also known as design matrix, model matrix or regressor matrix)
is an \eqn{n \times k} matrix of \eqn{n} observations of \eqn{k} regressors,
which includes a regressor whose value is 1 for each observation on the first column.}
}
\value{
Returns the residual maker matrix \eqn{\left( \mathbf{M} \right)} .
}
\description{
Calculates the residual maker matrix
\eqn{\left( \mathbf{M} \right)} using
\deqn{
    \mathbf{M} = \mathbf{I} - \mathbf{P}
  }
where
\deqn{
    \mathbf{P} = \mathbf{X} \left( \mathbf{X}^{T} \mathbf{X} \right)^{-1} \mathbf{X}^{T} .
  }
The residual maker matrix \eqn{\left( \mathbf{M} \right)}
transforms the \eqn{\mathbf{y}} vector
to the vector of residuals \eqn{\left( \mathbf{e} = \mathbf{My} \right)} .
}
\examples{
# Simple regression------------------------------------------------
X <- jeksterslabRdatarepo::wages.matrix[["X"]]
X <- X[, c(1, ncol(X))]
M <- M(X = X)
str(M, list.len = 6)

# Multiple regression----------------------------------------------
X <- jeksterslabRdatarepo::wages.matrix[["X"]]
# age is removed
X <- X[, -ncol(X)]
M <- M(X = X)
str(M, list.len = 6)
}
\references{
\href{https://en.wikipedia.org/wiki/Projection_matrix}{Wikipedia: Projection Matrix}
}
\seealso{
Other projection matrix functions: 
\code{\link{.M}()},
\code{\link{.h}()},
\code{\link{P}()},
\code{\link{h}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{projection matrix functions}
\keyword{projection}
