% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss.R
\name{mse}
\alias{mse}
\title{Mean Square Error}
\usage{
mse(rss = NULL, n, ...)
}
\arguments{
\item{rss}{Numeric.
Residual sum of squares.}

\item{n}{Integer.
Sample size.}

\item{...}{Arguments to pass to \code{\link[=ss_r]{ss_r()}}
if \code{rss = NULL}.}
}
\description{
Calculates the mean square error (MSE) using
\deqn{
    MSE
    =
    \frac{1}{n}
    \sum_{i = 1}^{n}
    \left(
      Y_i - \hat{Y}_{i}
    \right)^{2}
    =
    \frac{RSS}{n}.
  }
}
\details{
If \code{rss = NULL},
the \code{rss} is computed
using \code{\link[=ss_r]{ss_r()}}.
In this case the arguments
\code{X},
\code{y}, and
the optional argument \code{beta_hat}
are supplied using the \code{...} argument.
}
\references{
\href{https://en.wikipedia.org/wiki/Mean_squared_error}{Wikipedia: Mean Squared Error}
}
\seealso{
Other assessment of model quality functions: 
\code{\link{r_bar_sqr}()},
\code{\link{r_sqr}()},
\code{\link{rmse}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{assessment of model quality functions}
