% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mse.R
\name{mse}
\alias{mse}
\title{Mean Squared Error}
\usage{
mse(betahat = NULL, X, y)
}
\arguments{
\item{betahat}{Vector or \code{k} by \code{1} matrix.
The vector
\eqn{\boldsymbol{\hat{\beta}}}
is a \eqn{k \times 1} vector of estimates
of \eqn{k} unknown regression coefficients.}

\item{X}{Matrix.
The data matrix
\eqn{\mathbf{X}}
is an \eqn{n \times k} matrix
of \eqn{n} observations
of \eqn{k} regressors,
which includes a regressor
whose value is 1 for each observation.}

\item{y}{Vector or \code{n} by \code{1} matrix.
The vector
\eqn{\mathbf{y}}
is an \eqn{n \times 1} vector
of observations on the regressand variable.}
}
\value{
Returns the mean square error.
}
\description{
Calculates the mean square error (MSE) using
\deqn{
    MSE
    =
    \frac{1}{n}
    \sum_{i = 1}^{n}
    \left(
      Y_i - \hat{Y}_{i}
    \right)^{2}
    =
    \frac{RSS}{n}.
  }
}
\references{
\href{https://en.wikipedia.org/wiki/Mean_squared_error}{Wikipedia: Mean Squared Error}
}
\seealso{
Other assessment of model quality functions: 
\code{\link{.mse}()},
\code{\link{.r2_ess}()},
\code{\link{.r2_rss}()},
\code{\link{.rbar2}()},
\code{\link{.rmse}()},
\code{\link{r2}()},
\code{\link{rbar2}()},
\code{\link{rmse}()}
}
\concept{assessment of model quality functions}
