% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ram.R
\name{mutheta}
\alias{mutheta}
\title{Model-Implied Mean Vector
\eqn{\boldsymbol{\mu} \left( \boldsymbol{\theta} \right)}}
\usage{
mutheta(beta, muX)
}
\arguments{
\item{beta}{Numeric vector of length \code{k} or \code{k} by \code{1} matrix.
\eqn{k \times 1} regression coefficients \eqn{\left( \boldsymbol{\beta} \right)}.}

\item{muX}{Numeric vector of length \code{p} or \code{p} by \code{1} matrix. \eqn{p \times 1} vector
of means of the regressors \eqn{{X}_{2}, {X}_{3}, \cdots, {X}_{k}}
\eqn{\left( \boldsymbol{\mu}_{\mathbf{X}} \right)} .}
}
\value{
Returns the model-implied mean vector
\eqn{\boldsymbol{\mu} \left( \boldsymbol{\theta} \right)}.
Note that the first item corresponds to the expected value of \code{y}.
The rest of the items correspond to \code{muX}.
}
\description{
Model-implied mean vector
\eqn{\boldsymbol{\mu} \left( \boldsymbol{\theta} \right)}
from parameters of a \eqn{k}-variable linear regression model.
}
\details{
The following are the parameters of a linear regression model
for the mean structure
\itemize{
\item \eqn{\boldsymbol{\beta}} - \eqn{k \times k} regression coefficients and
\item \eqn{\boldsymbol{\mu}_{\mathbf{X}}} - \eqn{p \times 1}
vector of means of the regressors
\eqn{{X}_{2}, {X}_{3}, \cdots, {X}_{k}} .
}
}
\examples{
beta <- c(-12.7128845, 0.2076475, 0.4510391)
muX <- c(70.18, 3.06)
mutheta(beta = beta, muX = muX)
}
\seealso{
Other model-implied functions: 
\code{\link{Sigmatheta}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{model-implied functions}
\keyword{model-implied}
