% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_implied.R
\name{mutheta}
\alias{mutheta}
\title{Model-Implied Mean Vector
(\eqn{\boldsymbol{\mu} \left( \boldsymbol{\theta} \right)})}
\usage{
mutheta(beta, muX)
}
\arguments{
\item{beta}{Vector or \code{k} by \code{1} matrix.
The vector
\eqn{\boldsymbol{\beta}}
is a \eqn{k \times 1} vector
of \eqn{k} regression coefficients.}

\item{muX}{Vector or \code{p} by \code{1} matrix.
\eqn{\left( p \right) \times 1} vector
of means of
the regressors
\eqn{
      {X}_{2},
      {X}_{3},
      \dots,
      {X}_{k}
    }
excluding the mean for the vector of constants
(\eqn{\boldsymbol{\mu}_{\mathbf{X}}}).}
}
\description{
Model-implied mean vector
(\eqn{\boldsymbol{\mu} \left( \boldsymbol{\theta} \right)})
from parameters of a \eqn{k}-variable linear regression model.
The parameters of a linear regression model
for the mean structure
are
the regression coefficients
(\eqn{\boldsymbol{\beta}}),
and
the means of
the regressors
\eqn{
    {X}_{2},
    {X}_{3},
    \dots,
    {X}_{k}
  }
excluding the mean for the vector of constants
(\eqn{\boldsymbol{\mu}_{\mathbf{X}}}).
}
\author{
Ivan Jacob Agaloos Pesigan
}
