% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ram.R
\name{mutheta}
\alias{mutheta}
\title{Model-Implied Mean Vector
\eqn{\boldsymbol{\mu} \left( \boldsymbol{\theta} \right)}}
\usage{
mutheta(beta, muX)
}
\arguments{
\item{beta}{Vector or \code{k} by \code{1} matrix.
\eqn{k \times 1} regression coefficients \eqn{\left( \boldsymbol{\beta} \right)}.}

\item{muX}{Vector or \code{p} by \code{1} matrix. \eqn{p \times 1} vector
of means of the regressors \eqn{{X}_{2}, {X}_{3}, \cdots, {X}_{k}}
\eqn{\left( \boldsymbol{\mu}_{\mathbf{X}} \right)} .}
}
\description{
Model-implied mean vector
\eqn{\boldsymbol{\mu} \left( \boldsymbol{\theta} \right)}
from parameters of a \eqn{k}-variable linear regression model.
}
\details{
The following are the parameters of a linear regression model
for the mean structure
\itemize{
\item \eqn{\boldsymbol{\beta}} - \eqn{k \times k} regression coefficients and
\item \eqn{\boldsymbol{\mu}_{\mathbf{X}}} - \eqn{p \times 1} vector of means of the regressors
\eqn{{X}_{2}, {X}_{3}, \cdots, {X}_{k}} .
}
}
\seealso{
Other model-implied functions: 
\code{\link{Sigmatheta}()},
\code{\link{intercept}()},
\code{\link{sigma2}()},
\code{\link{slopes}()},
\code{\link{stdslopes}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{model-implied functions}
\keyword{model-implied}
