% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e.R
\name{My}
\alias{My}
\title{Residuals (\eqn{\mathbf{e} = \mathbf{My}})}
\usage{
My(y, M = NULL, X = NULL, P = NULL)
}
\arguments{
\item{y}{Vector or \code{n} by \code{1} matrix.
The vector
\eqn{\mathbf{y}}
is an \eqn{n \times 1} vector
of observations on the regressand variable.}

\item{M}{Matrix.
The residual maker matrix (\eqn{\mathbf{M}}).}

\item{X}{Matrix.
The data matrix
\eqn{\mathbf{X}}
is an \eqn{n \times k} matrix
of \eqn{n} observations
of \eqn{k} regressors,
which includes a regressor
whose value is 1 for each observation.}

\item{P}{Matrix.
The projection matrix
(\eqn{\mathbf{P}}).}
}
\value{
Returns residuals (\eqn{\mathbf{e}}),
that is,
the difference between
the observed and predicted value of \eqn{\mathbf{y}}
(\eqn{\mathbf{e} = \mathbf{y}} - \eqn{\mathbf{\hat{y}}}).
}
\description{
Calculates residuals using
\deqn{
    \mathbf{e}
    =
    \mathbf{My}.
  }
}
\details{
If \code{M = NULL},
the \code{M} matrix is computed
using \code{\link[=M]{M()}}
with \code{X} as a required argument
and \code{P} as an optional argument.
If \code{M} is provided,
\code{X} and \code{P} are not needed.
}
\references{
\href{https://en.wikipedia.org/wiki/Errors_and_residuals}{Wikipedia: Errors and Residuals}
}
\seealso{
Other residuals functions: 
\code{\link{e}()},
\code{\link{y_minus_yhat}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{residuals functions}
