% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epsilonhat.R
\name{My}
\alias{My}
\title{Residuals
\eqn{\left( \boldsymbol{\hat{\varepsilon}} = \mathbf{My} \right)}}
\usage{
My(X, y)
}
\arguments{
\item{X}{\code{n} by \code{k} numeric matrix.
The data matrix \eqn{\mathbf{X}}
(also known as design matrix, model matrix or regressor matrix)
is an \eqn{n \times k} matrix of \eqn{n} observations of \eqn{k} regressors,
which includes a regressor whose value is 1 for each observation on the first column.}

\item{y}{Numeric vector of length \code{n} or \code{n} by \code{1} matrix.
The vector \eqn{\mathbf{y}} is an \eqn{n \times 1} vector of observations
on the regressand variable.}
}
\value{
Returns an \eqn{n \times 1} matrix of residuals
\eqn{\left( \boldsymbol{\hat{\varepsilon}} \right)},
that is, the difference between the observed \eqn{\left( \mathbf{y} \right)}
and predicted \eqn{\left( \mathbf{\hat{y}} \right)} values of the regressand variable
\eqn{\left( \boldsymbol{\hat{\varepsilon}} = \mathbf{y} - \mathbf{\hat{y}} \right)}.
}
\description{
Calculates residuals using
\deqn{
    \boldsymbol{\hat{\varepsilon}} =  \mathbf{My} .
  }
where
\deqn{
    \mathbf{M}
    = \mathbf{I} - \mathbf{P} \\
    = \mathbf{I} - \mathbf{X}
      \left( \mathbf{X}^{T} \mathbf{X} \right)^{-1} \mathbf{X}^{T} .
  }
}
\examples{
X <- jeksterslabRdatarepo::wages.matrix[["X"]]
# age is removed
X <- X[, -ncol(X)]
y <- jeksterslabRdatarepo::wages.matrix[["y"]]
My(X = X, y = y)
}
\references{
\href{https://en.wikipedia.org/wiki/Errors_and_residuals}{Wikipedia: Errors and Residuals}
}
\seealso{
Other residuals functions: 
\code{\link{.My}()},
\code{\link{.tepsilonhat}()},
\code{\link{.yminusyhat}()},
\code{\link{epsilonhat}()},
\code{\link{tepsilonhat}()},
\code{\link{yminusyhat}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{residuals functions}
\keyword{residuals}
