% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2.R
\name{r2}
\alias{r2}
\title{R-square}
\usage{
r2(beta_hat = NULL, X, y, fromrss = TRUE)
}
\arguments{
\item{beta_hat}{Vector or \code{k} by \code{1} matrix.
The vector
\eqn{\boldsymbol{\hat{\beta}}}
is a \eqn{k \times 1} vector of estimates
of \eqn{k} unknown regression coefficients.}

\item{X}{Matrix.
The data matrix
\eqn{\mathbf{X}}
is an \eqn{n \times k} matrix
of \eqn{n} observations
of \eqn{k} regressors,
which includes a regressor
whose value is 1 for each observation.}

\item{y}{Vector or \code{n} by \code{1} matrix.
The vector
\eqn{\mathbf{y}}
is an \eqn{n \times 1} vector
of observations on the regressand variable.}

\item{fromrss}{Logical.
If \code{TRUE},
calculates the coefficient of determinism
from \code{RSS}.
If \code{FALSE},
calculates the coefficient of determinism
from \code{ESS}.}
}
\description{
Calculates the coefficient of determination using
\deqn{
    R^2
    =
    1 - \frac{\textrm{Residual sum of squares}}{\textrm{Total sum of squares}}
  }
or
\deqn{
    R^2
    =
    \frac{\textrm{Explained sum of squares}}{\textrm{Total sum of squares}}.
  }
}
\references{
\href{https://en.wikipedia.org/wiki/Residual_sum_of_squares}{Wikipedia: Residual Sum of Squares}

\href{https://en.wikipedia.org/wiki/Explained_sum_of_squares}{Wikipedia: Explained Sum of Squares}

\href{https://en.wikipedia.org/wiki/Total_sum_of_squares}{Wikipedia: Total Sum of Squares}

\href{https://en.wikipedia.org/wiki/Coefficient_of_determination}{Wikipedia: Coefficient of Determination}
}
\seealso{
Other assessment of model quality functions: 
\code{\link{mse}()},
\code{\link{r2_ess}()},
\code{\link{r2_rss}()},
\code{\link{rbar2_r2}()},
\code{\link{rbar2}()},
\code{\link{rmse}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{assessment of model quality functions}
