% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R2.R
\name{R2}
\alias{R2}
\title{Coefficient of Determination \eqn{R^2}}
\usage{
R2(X, y, fromRSS = TRUE)
}
\arguments{
\item{X}{\code{n} by \code{k} numeric matrix.
The data matrix \eqn{\mathbf{X}}
(also known as design matrix, model matrix or regressor matrix)
is an \eqn{n \times k} matrix of \eqn{n} observations of \eqn{k} regressors,
which includes a regressor whose value is 1 for each observation on the first column.}

\item{y}{Numeric vector of length \code{n} or \code{n} by \code{1} matrix.
The vector \eqn{\mathbf{y}} is an \eqn{n \times 1} vector of observations
on the regressand variable.}

\item{fromRSS}{Logical.
If \code{TRUE}, calculates the coefficient of determination from \code{RSS}.
If \code{FALSE}, calculates the coefficient of determination from \code{ESS}.}
}
\description{
Calculates the coefficient of determination using
\deqn{
    R^2 = 1 - \frac{\textrm{Residual sum of squares}}
    {\textrm{Total sum of squares}}
  }
or
\deqn{
    R^2 = \frac{\textrm{Explained sum of squares}}
    {\textrm{Total sum of squares}} .
}
}
\examples{
X <- jeksterslabRdatarepo::wages.matrix[["X"]]
# age is removed
X <- X[, -ncol(X)]
y <- jeksterslabRdatarepo::wages.matrix[["y"]]
R2(
  X = X,
  y = y
)
}
\references{
\href{https://en.wikipedia.org/wiki/Residual_sum_of_squares}{Wikipedia: Residual Sum of Squares}

\href{https://en.wikipedia.org/wiki/Explained_sum_of_squares}{Wikipedia: Explained Sum of Squares}

\href{https://en.wikipedia.org/wiki/Total_sum_of_squares}{Wikipedia: Total Sum of Squares}

\href{https://en.wikipedia.org/wiki/Coefficient_of_determination}{Wikipedia: Coefficient of Determination}
}
\seealso{
Other assessment of model quality functions: 
\code{\link{.MSE}()},
\code{\link{.R2fromESS}()},
\code{\link{.R2fromRSS}()},
\code{\link{.RMSE}()},
\code{\link{.Rbar2}()},
\code{\link{.model}()},
\code{\link{MSE}()},
\code{\link{RMSE}()},
\code{\link{Rbar2}()},
\code{\link{model}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{assessment of model quality functions}
\keyword{model-assessment}
