% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2.R
\name{rbar2}
\alias{rbar2}
\title{Adjusted R-square}
\usage{
rbar2(beta_hat = NULL, X, y)
}
\arguments{
\item{beta_hat}{Vector or \code{k} by \code{1} matrix.
The vector
\eqn{\boldsymbol{\hat{\beta}}}
is a \eqn{k \times 1} vector of estimates
of \eqn{k} unknown regression coefficients.}

\item{X}{Matrix.
The data matrix
\eqn{\mathbf{X}}
is an \eqn{n \times k} matrix
of \eqn{n} observations
of \eqn{k} regressors,
which includes a regressor
whose value is 1 for each observation.}

\item{y}{Vector or \code{n} by \code{1} matrix.
The vector
\eqn{\mathbf{y}}
is an \eqn{n \times 1} vector
of observations on the regressand variable.}
}
\value{
Returns the adjusted coefficient of determination \eqn{\bar{R}^{2}}.
}
\description{
Calculates the adjusted coefficient of determination
\deqn{
    \bar{R}^{2}
    =
    1
    -
    \left(
      \frac{RSS / \left( n - k \right)}{TSS / \left(n - 1 \right)}
    \right)
    =
    1
    -
    \left(
      1 - R^2
    \right)
    \frac{n - 1}{n - k}.
  }
}
\references{
\href{https://en.wikipedia.org/wiki/Residual_sum_of_squares}{Wikipedia: Residual Sum of Squares}

\href{https://en.wikipedia.org/wiki/Explained_sum_of_squares}{Wikipedia: Explained Sum of Squares}

\href{https://en.wikipedia.org/wiki/Total_sum_of_squares}{Wikipedia: Total Sum of Squares}

\href{https://en.wikipedia.org/wiki/Coefficient_of_determination}{Wikipedia: Coefficient of Determination}
}
\seealso{
Other assessment of model quality functions: 
\code{\link{mse}()},
\code{\link{r2_ess}()},
\code{\link{r2_rss}()},
\code{\link{r2}()},
\code{\link{rbar2_r2}()},
\code{\link{rmse}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{assessment of model quality functions}
