% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residual.plot.R
\name{residual.plot}
\alias{residual.plot}
\title{Residual Plots}
\usage{
residual.plot(yhat, tepsilonhat, epsilonhat)
}
\arguments{
\item{yhat}{Vector of length \code{n} or \code{n} by \code{1} matrix.
\eqn{n \times 1} vector of predicted values of \eqn{\mathbf{y}}
\eqn{\left( \mathbf{\hat{y}} \right)}.}

\item{tepsilonhat}{Studentized residuals.}

\item{epsilonhat}{Numeric vector of length \code{n} or \code{n} by 1 matrix.
\eqn{n \time 1} vector of residuals.}
}
\description{
Residual Plots
}
\seealso{
Other plotting functions: 
\code{\link{scatter.plot}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{plotting functions}
\keyword{plot}
