% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSE.R
\name{RMSE}
\alias{RMSE}
\title{Root Mean Squared Error}
\usage{
RMSE(X, y)
}
\arguments{
\item{X}{\code{n} by \code{k} numeric matrix.
The data matrix \eqn{\mathbf{X}}
(also known as design matrix, model matrix or regressor matrix)
is an \eqn{n \times k} matrix of \eqn{n} observations of \eqn{k} regressors,
which includes a regressor whose value is 1 for each observation on the first column.}

\item{y}{Numeric vector of length \code{n} or \code{n} by \code{1} matrix.
The vector \eqn{\mathbf{y}} is an \eqn{n \times 1} vector of observations
on the regressand variable.}
}
\value{
Returns the root mean squared error.
}
\description{
Calculates the root mean squared error
\eqn{\left( \mathrm{RMSE} \right)} using
\deqn{
    \mathrm{RMSE}
    = \sqrt{\frac{1}{n} \sum_{i = 1}^{n}
      \left( \mathbf{y} - \mathbf{X} \boldsymbol{\hat{\beta}} \right)^{2}} \\
    = \sqrt{\frac{1}{n} \sum_{i = 1}^{n}
      \left( \mathbf{y} - \mathbf{\hat{y}} \right)^{2}} \\
    = \sqrt{\frac{\mathrm{RSS}}{n}} .
  }
}
\examples{
# Simple regression------------------------------------------------
X <- jeksterslabRdatarepo::wages.matrix[["X"]]
X <- X[, c(1, ncol(X))]
y <- jeksterslabRdatarepo::wages.matrix[["y"]]
RMSE(X = X, y = y)

# Multiple regression----------------------------------------------
X <- jeksterslabRdatarepo::wages.matrix[["X"]]
# age is removed
X <- X[, -ncol(X)]
RMSE(X = X, y = y)
}
\references{
\href{https://en.wikipedia.org/wiki/Root-mean-square_deviation}{Wikipedia: Root-mean-square deviation}
}
\seealso{
Other assessment of model quality functions: 
\code{\link{.MSE}()},
\code{\link{.R2fromESS}()},
\code{\link{.R2fromRSS}()},
\code{\link{.RMSE}()},
\code{\link{.Rbar2}()},
\code{\link{.model}()},
\code{\link{MSE}()},
\code{\link{R2}()},
\code{\link{Rbar2}()},
\code{\link{model}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{assessment of model quality functions}
\keyword{model-assessment}
