% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigma2epsilon.R
\name{sigma2epsilon}
\alias{sigma2epsilon}
\title{Error Variance \eqn{\sigma_{\varepsilon}^{2}}}
\usage{
sigma2epsilon(slopes, sigma2y, sigmayX, SigmaX)
}
\arguments{
\item{slopes}{Numeric vector of length \code{p} or \code{p} by \code{1} matrix.
\eqn{p \times 1} regression slopes \eqn{\left( \boldsymbol{\beta}_{2, \cdots, k} \right)} .}

\item{sigma2y}{Numeric.
Variance of the regressand variable \eqn{\left( \sigma_{y}^{2} \right)}.}

\item{sigmayX}{Numeric vector of length \code{p} or \code{p} by \code{1} matrix.
\eqn{p \times 1} covariances between the regressand and the regressors
\eqn{\left( \boldsymbol{\sigma}_{\mathbf{yX}} \right)}.}

\item{SigmaX}{\code{p} by \code{p} numeric matrix.
\eqn{p \times p} variances and covariances of \eqn{{X}_{2}, {X}_{3}, \cdots, {X}_{k}}
\eqn{\left( \boldsymbol{\Sigma}_{\mathbf{X}} \right)}.}
}
\description{
Error variance \eqn{\sigma_{\varepsilon}^{2}} from the slopes
\eqn{\boldsymbol{\beta}_{2, \cdots, k}}
and variance-covariance matrix \eqn{\boldsymbol{\Sigma}}.
}
\seealso{
Other parameter functions: 
\code{\link{intercept}()},
\code{\link{slopesprime}()},
\code{\link{slopes}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{parameter functions}
\keyword{parameter}
