% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_implied.R
\name{Sigmatheta}
\alias{Sigmatheta}
\title{Model-Implied Variance-Covariance Matrix
(\eqn{\boldsymbol{\Sigma} \left( \boldsymbol{\theta} \right)})}
\usage{
Sigmatheta(beta_slopes, sigma2, SigmaX)
}
\arguments{
\item{beta_slopes}{Vector or \code{p} by \code{1} matrix.
\eqn{\left( p \right) \times 1}
vector of regression slopes.}

\item{sigma2}{Numeric.
Variance of the error term
\eqn{\boldsymbol{\varepsilon}}
(\eqn{\sigma^2}).}

\item{SigmaX}{Matrix.
Variances and covariances of
\eqn{
    {X}_{2},
    {X}_{3},
    \dots,
    {X}_{k}
  }
(\eqn{\boldsymbol{\Sigma}_{\mathbf{X}}}).}
}
\value{
Returns the model-implied
variance-covariance matrix
(\eqn{\boldsymbol{\Sigma} \left( \boldsymbol{\theta} \right)}).
}
\description{
Model-implied variance-covariance matrix
(\eqn{\boldsymbol{\Sigma} \left( \boldsymbol{\theta} \right)})
from parameters of a \eqn{k}-variable linear regression model.
The parameters of a linear regression model
for the covariance structure
are
the slopes
(\eqn{\boldsymbol{\beta}_{\mathrm{slopes}}}),
the variance of the error term \eqn{\boldsymbol{\varepsilon}}
(\eqn{\sigma^2}),
and
the variances and covariances of
\eqn{
    {X}_{2},
    {X}_{3},
    \dots,
    {X}_{k}
  }
(\eqn{\boldsymbol{\Sigma}_{\mathbf{X}}}).
}
\author{
Ivan Jacob Agaloos Pesigan
}
