% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ram.R
\name{Sigmatheta}
\alias{Sigmatheta}
\title{Model-Implied Variance-Covariance Matrix
\eqn{\boldsymbol{\Sigma} \left( \boldsymbol{\theta} \right)}}
\usage{
Sigmatheta(slopes, sigma2epsilon, SigmaX)
}
\arguments{
\item{slopes}{Numeric vector of length \code{p} or \code{p} by \code{1} matrix.
\eqn{p \times 1} regression slopes \eqn{\left( \boldsymbol{\beta}_{2, \cdots, k} \right)} .}

\item{sigma2epsilon}{Numeric.
Variance of the error term \eqn{\varepsilon} \eqn{\left( \sigma_{\varepsilon}^{2} \right)}.}

\item{SigmaX}{\code{p} by \code{p} numeric matrix.
\eqn{p \times p} variances and covariances of \eqn{{X}_{2}, {X}_{3}, \cdots, {X}_{k}}
\eqn{\left( \boldsymbol{\Sigma}_{\mathbf{X}} \right)}.}
}
\value{
Returns the model-implied variance-covariance matrix
\eqn{\boldsymbol{\Sigma} \left( \boldsymbol{\theta} \right)}.
Note that the first item corresponds to \code{y}.
The rest of the items correspond to how \code{SigmaX} is arranged.
}
\description{
Model-implied variance-covariance matrix
\eqn{\boldsymbol{\Sigma} \left( \boldsymbol{\theta} \right)}
from parameters of a \eqn{k}-variable linear regression model.
}
\details{
The following are the parameters of a linear regression model
for the covariance structure
\itemize{
\item \eqn{\boldsymbol{\beta}_{2, \cdots, k}} - \eqn{p \times 1} regression slopes,
\item \eqn{\sigma_{\varepsilon}^{2}} - variance of the error term
\eqn{\varepsilon}, and
\item \eqn{\boldsymbol{\Sigma}_{\mathbf{X}}} - \eqn{p \times p}
variances and covariances of \eqn{{X}_{2}, {X}_{3}, \cdots, {X}_{k}}.
}
}
\seealso{
Other model-implied functions: 
\code{\link{mutheta}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{model-implied functions}
\keyword{model-implied}
