% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ram.R
\name{Sigmatheta}
\alias{Sigmatheta}
\title{Model-Implied Variance-Covariance Matrix
\eqn{\boldsymbol{\Sigma} \left( \boldsymbol{\theta} \right)}}
\usage{
Sigmatheta(slopes, sigma2, SigmaX)
}
\arguments{
\item{slopes}{Vector or \code{p} by \code{1} matrix.
Regression slopes \eqn{\left( \boldsymbol{\beta}_{2 \cdots k} \right)} .}

\item{sigma2}{Numeric.
Variance of the error term \eqn{\varepsilon} \eqn{\left( \sigma^2 \right)}.}

\item{SigmaX}{\code{p} by \code{p} numeric matrix.
Variances and covariances of \eqn{{X}_{2}, {X}_{3}, \cdots, {X}_{k}}
\eqn{\left( \boldsymbol{\Sigma}_{\mathbf{X}} \right)}.}
}
\value{
Returns the model-implied variance-covariance matrix
\eqn{\boldsymbol{\Sigma} \left( \boldsymbol{\theta} \right)}.
}
\description{
Model-implied variance-covariance matrix
\eqn{\boldsymbol{\Sigma} \left( \boldsymbol{\theta} \right)}
from parameters of a \eqn{k}-variable linear regression model.
}
\details{
The following are the parameters of a linear regression model
for the covariance structure
\itemize{
\item \eqn{\boldsymbol{\beta}_{2 \cdots k}} - \eqn{p \times 1} regression slopes,
\item \eqn{\sigma^2} - variance of the error term
\eqn{\varepsilon}, and
\item \eqn{\boldsymbol{\Sigma}_{\mathbf{X}}} - \eqn{p \times p}
variances and covariances of \eqn{{X}_{2}, {X}_{3}, \cdots, {X}_{k}}.
}
}
\seealso{
Other model-implied functions: 
\code{\link{intercept}()},
\code{\link{mutheta}()},
\code{\link{sigma2epsilon}()},
\code{\link{slopesprime}()},
\code{\link{slopes}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{model-implied functions}
\keyword{model-implied}
