% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta.R
\name{slopesprime}
\alias{slopesprime}
\title{Regression Standardized Slopes \eqn{\boldsymbol{\beta}_{2, \cdots, k}^{\prime}}}
\usage{
slopesprime(RX, ryX)
}
\arguments{
\item{RX}{\code{p} by \code{p} numeric matrix.
\eqn{p \times p} correlations between the regressors.}

\item{ryX}{Numeric vector of length \code{p} or \code{p} by \code{1} matrix.
\eqn{p \times 1} correlations between the regressand and the regressors.}
}
\value{
Returns the standardized slopes
\eqn{\boldsymbol{\beta}_{2, \cdots, k}^{\prime}}
of a linear regression model derived from the correlation matrix.
}
\description{
Derives the standardized slopes
\eqn{\boldsymbol{\beta}_{2, \cdots, k}^{\prime}} of a linear regression model
as a function of correlations.
}
\details{
The linear regression standardized slopes is given by
\deqn{
    \boldsymbol{\beta}_{2, \cdots, k}^{\prime} =
    \mathrm{R}_{\mathbf{X}}^{T} \mathrm{r}_{\mathbf{yX}}
  }

where
\itemize{
\item \eqn{\mathbf{R}_{\mathbf{X}}}
is the \eqn{p \times p} correlation matrix of the regressor variables and
\item \eqn{\mathbf{r}_{\mathbf{yX}}}
is the \eqn{p \times 1} column vector
of the correlations between the regressand and the regressors.
}
}
\seealso{
Other parameter functions: 
\code{\link{intercept}()},
\code{\link{sigma2epsilon}()},
\code{\link{slopes}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{parameter functions}
\keyword{parameter}
