% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SS.R
\name{TSS}
\alias{TSS}
\title{Total Sum of Squares.}
\usage{
TSS(y)
}
\arguments{
\item{y}{Numeric vector of length \code{n} or \code{n} by \code{1} matrix.
The vector \eqn{\mathbf{y}} is an \eqn{n \times 1} vector of observations
on the regressand variable.}
}
\value{
Returns the total sum of squares \eqn{\left( TSS \right)}.
}
\description{
Calculates the total sum of squares \eqn{\left( TSS \right)} using
\deqn{
    TSS = \sum_{i = 1}^{n} \left( Y_i - \bar{Y} \right)^2 \\
    = \sum_{i = 1}^{n} Y_{i}^{2} - n \bar{Y}^2
  }
In matrix form
\deqn{
    TSS = \sum_{i = 1}^{n} \left( \mathbf{y} - \mathbf{\bar{y}} \right)^2
  }
Equivalent computational matrix formula
\deqn{
    TSS = \mathbf{y}^{\prime} \mathbf{y} - n \mathbf{\bar{Y}}^{2}.
  }
Note that
\deqn{
    TSS = ESS + RSS .
  }
}
\references{
\href{https://en.wikipedia.org/wiki/Residual_sum_of_squares}{Wikipedia: Residual Sum of Squares}

\href{https://en.wikipedia.org/wiki/Explained_sum_of_squares}{Wikipedia: Explained Sum of Squares}

\href{https://en.wikipedia.org/wiki/Total_sum_of_squares}{Wikipedia: Total Sum of Squares}

\href{https://en.wikipedia.org/wiki/Coefficient_of_determination}{Wikipedia: Coefficient of Determination}
}
\seealso{
Other sum of squares functions: 
\code{\link{.ESS}()},
\code{\link{.RSS}()},
\code{\link{ESS}()},
\code{\link{RSS}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{sum of squares functions}
\keyword{SS}
